==> scripts/1_summary_stats.Rmd <==
---
title: "Summary statistics"
author: "Christopher L. Crawford"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: html_document
editor_options: 
  chunk_output_type: console
---

The markdown document calculates the primary summary statistics for the Crawford et al. abandonment trajectories manuscript (see: https://github.com/chriscra/abandonment_trajectories).
It is loosely organized as follows:

1. Prepping the derived abandonment length, persistence, area, and turnover data.

2. Calculating **summary statistics** for the manuscript, particularly the mean duration of abandonment at each site.
  a. Saving plots of the mean abandonment duration for our five-year threshold, and across a range of thresholds (1, 3, 5, 7, 10), 
  b. Plotting the mean observed abandonment duration along with the underlying distribution of all abandonment periods. This resulted in the main text Figure 4 (and a corresponding figure in the SI for max lengths, for potential abandonment without recultivation, and recultivation periods themselves).
  
3. Calculating the **area abandoned**, resulting in *area_summary_df*. This also involved:
  a. Calculating the area of those pixels classified as cropland at least once during the time series (*the total cropland extent*), 
  b. Plotting in a three plot panel figure, 1) Area abandoned as of 2017, 2) Area abandoned at least once between 1987-2017, and 3) Area abandoned as of 2017 as a proportion of total cropland extent.
  c. Calculating the mean proportion of abandoned land that is recultivated at each site (allowing for multiple instances of abandonment for pixel), and how this changes based on the chosen definition of abandonment (1, 3, 5, 7, or 10 years). [Note that the script "cluster/6_lc_of_abn.R" calculates the proportion of abandonment (as of 2017) that is classified as woody vegetation vs. grassy (herbaceous) vegetation.]

4. Comparing our method for identifying abandonment using the full annual time series to a commonly used two-year method, using just two distinct time points.
  a. This involved creating maps of abandonment as of 2017 using these two methods, 
  b. Calculating the Jaccard similarity between the two (a measure of overlap),
  c. Calculating the proportion of the area identified as "abandonment" using the two-year method is actually "young" abandonment, i.e., has been abandoned for less than five years (our chosen definition of abandonment)
  d. Producing a summary data.frame including the percent difference and Jaccard similarity values, which is then used to produce a table for the SI.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
mean_length_df$abn_threshold %>% unique()
```

```{r initialize}
source("/Users/christophercrawford/work/projects/abandonment_trajectories/scripts/0_start.R")
```

```{r set-site-params, eval = FALSE}
# Make data.frame containing site names, labels, original land cover codes, and noting whether to 1) merge layers into a stack, 2) update land cover codes, or 3) trim extra years

# Primary sites
# 1. Shaanxi - He
# 2. Belarus (Smolensk) - He
# 3. Chongqing - Yanhua (not great -fields are small, mountainous, and lots of clouds) - look at the data to see. 
# 4. Goias - Ben
# 5. Mato Grasso - Amintas
# 6. Nebraska - Seth
# 7. Wisconsin - David
# 
# ok
# 8. Volgograd - Natalia
# 9. Orenburg - Natalia
# 10. Bosnia & Herzegovina (medium)
# 11. Iraq (medium good, low accuracy over years, but good single accuracy)
# 
# exclude 
# 12. Nepal - Johanna
# 13. Sardinia - Kasia
# 14. Uganda - Niwaeli

# CAREFUL!!!!! Land Cover codes are different for each file:
# Land use class codes: (for He's site, Shaanxi and Belarus)
#       1. Non-vegetated area (e.g. water, urban, barren land)
#       2. Woody vegetation
#       3. Cropland 
#       4. Herbaceous land (e.g. grassland)
# 
# Also be careful of years: Nebraska runs from 1986-2018, Wisconsin 1987:2018. All other sites run 1987-2017.

# Shaanxi (China):      1 others;     2 forest,     3 cropland;     4 grassland
# Belarus / Russia:     1 others;     2 forest,     3 cropland;     4 grassland
# Goias (Brazil):       1 others;     2 forest,     3 cropland;     4 grassland
# Mato Grosso (Brazil): 1 others;     2 forest;     3 cropland;     4 grassland
# Bosnia & Herzegovina: 1 others;     2 forest;     3 cropland;     4 grassland
# Chongqing (China):    1 others;     2 forest;     3 cropland;     4 grassland

# Iraq:                 1 others;     2 cropland;   3 forest;       4 grassland
# Nebraska (US):        1 cropland;   2 forest;     3 others;       4 grassland
# Orenburg (Russia):    1 others;     2 cropland;   3 grassland;    4 forest
# Volgograd (Russia):   1 others;     2 cropland;   3 grassland;    4 forest
# Wisconsin (US):       1 cropland;   2 grassland;  3 forest;       4 others

# list of all sites
site_list <- c("belarus", "bosnia_herzegovina", "chongqing", 
               "goias", "iraq", "mato_grosso", 
               "nebraska", "orenburg", "shaanxi", 
               "volgograd", "wisconsin")


site_label_list <- c("_b", "_bh", "_c", 
                     "_g", "_i", "_mg", 
                     "_n", "_o", "_s", 
                     "_v", "_w")

site_description <- c(
  "Vitebsk, Belarus / Smolensk, Russia",
  "Bosnia & Herzegovina", 
  "Chongqing, China",
  "Goiás, Brazil",
  "Iraq", 
  "Mato Grosso, Brazil",
  "Nebraska/Wyoming, USA", 
  "Orenburg, Russia / Uralsk, Kazakhstan",
  "Shaanxi/Shanxi, China", 
  "Volgograd, Russia", 
  "Wisconsin, USA"
  )

site_merge_layers <- site_list[c(3, 4, 5, 7, 8, 9, 10)]
c("chongqing", "goias", "iraq", "nebraska", "orenburg", "shaanxi", "volgograd")

site_update_lc <- site_list[c(5, 7, 8, 10, 11)]
c("iraq", "nebraska", "orenburg", "volgograd", "wisconsin")


# make a data.frame outlining site names, labels, whether to update land cover, whether to merge, whether to trim years, original land cover codes etc. 
site_df <- data.frame(site = site_list,
                      label = site_label_list) %>% 
  mutate(merge_layers = ifelse(site %in% 
                                 c("chongqing", "goias", "iraq", 
                                   "nebraska", "orenburg", "shaanxi", "volgograd"), 
                               "Yes", "No"),
         update_lc = ifelse(site %in% 
                              c("iraq", "nebraska", "orenburg", 
                                "volgograd", "wisconsin"), 
                            "Yes", "No"),
         trim_years = ifelse(site %in% 
                               c("nebraska", "wisconsin"), 
                             "Yes", "No"),
         other = 1, 
         woody_veg = 2,
         cropland = 3,
         grassland = 4,
         description = site_description
         )

# Update land cover classes for the following sites:
# Iraq:             1 others;     2 cropland;   3 forest;       4 grassland
# Nebraska:         1 cropland;   2 forest;     3 others;       4 grassland
# Russia Orenburg:  1 others;     2 cropland;   3 grassland;    4 forest
# Russia Volgograd: 1 others;     2 cropland;   3 grassland;    4 forest
# Wisconsin:        1 cropland;   2 grassland;  3 forest;       4 others

site_df[site_df$site == "iraq", 
        c("other", "cropland", "woody_veg", "grassland")] <- 1:4

site_df[site_df$site == "nebraska", 
        c("cropland", "woody_veg", "other", "grassland")] <- 1:4

site_df[site_df$site == "orenburg", 
        c("other", "cropland", "grassland", "woody_veg")] <- 1:4

site_df[site_df$site == "volgograd", 
        c("other", "cropland", "grassland", "woody_veg")] <- 1:4

site_df[site_df$site == "wisconsin", 
        c("cropland", "grassland", "woody_veg", "other")] <- 1:4

site_df

# save site_df
write_csv(site_df, file = paste0(p_dat_derived, "site_df.csv"))
```

# Prep datasets
```{r run-label}
run_label # just check that this is the right one
# see "_util_main.R"
```


```{r persistence-dat}

# load csvs into a list
persistence_dat <- lapply(1:11, function(i) {
  read.csv(file = paste0(p_dat_derived, run_label, "/", 
                         site_df$site[i], "_result_persistence", run_label, ".csv"))
})


names(persistence_dat) <- site_df$site
persistence_dat <- persistence_dat %>% bind_rows(.id = "site") %>% as_tibble()

persistence_dat <- left_join(persistence_dat,
            persistence_dat %>% filter(age == 5) %>% select(site, year_abn, initial_area_abn = area_ha), 
            by = c("site", "year_abn")) %>%
  mutate(year_abn_bins = cut(year_abn, breaks = 5, include.lowest = TRUE)) %>%
  mutate(year_abn_bins = as_factor(year_abn_bins), # convert both to factors
         time = age - 5,
         time_1 = time + 1,
         cohort = factor(year_abn)) %>% 
  mutate(site = as_factor(site)) %>% filter(!is.na(proportion)) # remove all na proportions


# save persistence_dat
write_csv(persistence_dat, file = paste0(p_derived2, "persistence_dat", run_label, ".csv"))






# ---------------------------------------------------------------------------- #
# ---------------------------------------------------------------------------- #
# variables of interest
# ---------------------------------------------------------------------------- #
# ---------------------------------------------------------------------------- #
persistence_dat
persistence_dat$proportion # response variable: the proportion of originally abandoned land that remains in a given year. 
persistence_dat$year # the current year, allowing us to measure area abandoned in the current year, relative to the amount initially abandoned
persistence_dat$year_abn # cohort of land abandoned in a particular year, which is then followed year after year to track persistence.
persistence_dat$cohort # a factor equivalent of "year_abn" to be used for cohort 
persistence_dat$area_ha # area abandoned (in hectares) for a given cohort in that specific year.
persistence_dat$age # the number of years the land in a cohort has been abandoned for.

persistence_dat$time # age - 5; adjusted time abandoned beyond the initial abandonment threshold. So, 0 corresponds to an abandonment age of 5, 1 -> 6, 2 -> 7 etc.

persistence_dat$time_1 # age - 4 (i.e. time + 1). This is so that time starts from 1, to avoid issues with taking the log of 0. 

persistence_dat$bins # bins of abandoned land of similar ages (5-10 years)
persistence_dat$year_abn_bins # bins of land abandoned during the same five to six year window
persistence_dat$initial_area_abn # the area (in ha) initially abandoned in a given year (cohort)

```

```{r area-dat}

# load csvs into a list
area_dat <- lapply(1:11, function(i) {
  read.csv(file = paste0(p_dat_derived, run_label, "/", 
                         site_df$site[i], "_result_area", run_label, ".csv"))
})

names(area_dat) <- site_df$site
area_dat <- area_dat %>% bind_rows(.id = "site") %>% as_tibble()


write_csv(area_dat, file = paste0(p_derived2, "area_dat", run_label, ".csv"))
# area_dat <- read_csv(file = paste0(p_derived2, "area_dat", run_label, ".csv"))

# ---------------------------------------------------------------------- #
# plot facet grid
# ---------------------------------------------------------------------- #
ggplot(data = area_dat %>% filter(lc != "Abandoned (>=1)")
    ) +
    theme_classic() +
    # theme(axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0)) +
    labs(y = expression("Area  (10"^{6}*" ha)") , 
         x = "Year", 
         title = "Area by Land Cover, with Abandonment",
         #subtitle = subtitle,
         color = NULL #"Land Cover"
         ) + 
    geom_line(mapping = aes(x = year, y = area_ha / (10^6), # convert to millions of ha
                            group = lc, color = lc),
              size = 1.5) + 
    theme(legend.position = "bottom",
          legend.title = element_text(size = 10)) +
    guides(color = guide_legend(nrow = 2, ncol = 3)) +
    
    scale_color_manual(values = plot_cols) + 
  facet_wrap(vars(site))

```



```{r turnover-dat}
# load csvs into a list
turnover_dat <- lapply(1:11, function(i) {
  read.csv(file = paste0(p_dat_derived, run_label, "/", 
                         site_df$site[i], "_result_turnover", run_label, ".csv"))
})

names(turnover_dat) <- site_df$site
turnover_dat <- turnover_dat %>% bind_rows(.id = "site") %>% as_tibble()

write_csv(turnover_dat, file = paste0(p_derived2, "turnover_dat", run_label, ".csv"))
# turnover_dat <- read_csv(file = paste0(p_derived2, "turnover_dat", run_label, ".csv"))

ggplot() +
    theme_classic() + 
    labs(y = expression("Annual area turnover (10"^{3}*" ha)"), 
         x = "Year", 
         #title = "Change in Area of Abandoned Land",
         #subtitle = subtitle,
         fill = NULL, #"Annual Change",
         color = NULL) + 
    scale_color_manual(values = c("black"),
                       labels = "Net Change") + 
    scale_fill_manual(values = brewer_pal(palette = "Greens")(5)[3:2],
                      labels = c("Area Gained", "Area Lost")) + 
    theme(legend.position = "bottom",
      #legend.position = c(0.2, 0.7),
      # legend.direction = "horizontal",
      legend.title = element_text(size = 10)
    ) +
    geom_col(data = filter(turnover_dat, change != "net"),
             mapping = aes(x = year, y = area_ha / (10^3), 
                           group = change, fill = change)) + 
    geom_line(data = filter(turnover_dat, change == "net"),
              mapping = aes(x = year, y = area_ha / (10^3), color = "Net Change in Area"),
              size = 1.5) + 
  facet_wrap(vars(site))
```

```{r potential-area-at}

# load csvs into a list
potential_area_dat <- lapply(1:11, function(i) {
  read.csv(file = paste0(p_dat_derived, run_label, "/", 
                         site_df$site[i], "_result_potential_area", run_label, ".csv"))
})

names(potential_area_dat) <- site_df$site
potential_area_dat <- potential_area_dat %>% bind_rows(.id = "site") %>% as_tibble()

write_csv(potential_area_dat, file = paste0(p_derived2, "potential_area_dat", run_label, ".csv"))
# potential_area_dat <- read_csv(file = paste0(p_derived2, "potential_area_dat", run_label, ".csv"))

# ---------------------------------------------------------------------- #
# plot facet grid
# ---------------------------------------------------------------------- #
ggplot(data = potential_area_dat %>% filter(lc != "Abandoned (>=1)")
    ) +
    theme_classic() +
    # theme(axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0)) +
    labs(y = expression("Area  (10"^{6}*" ha)") , 
         x = "Year", 
         title = "Area by Land Cover, with Abandonment",
         #subtitle = subtitle,
         color = NULL #"Land Cover"
         ) + 
    geom_line(mapping = aes(x = year, y = area_ha / (10^6), # convert to millions of ha
                            group = lc, color = lc),
              size = 1.5) + 
    theme(legend.position = "bottom",
          legend.title = element_text(size = 10)) +
    guides(color = guide_legend(nrow = 2, ncol = 3)) +
    
    scale_color_manual(values = plot_cols) + 
  facet_wrap(vars(site))
```


```{r potential-persistence-dat}

# load csvs into a list
potential_persistence_dat <- lapply(1:11, function(i) {
  read.csv(file = paste0(p_dat_derived, run_label, "/", 
                         site_df$site[i], "_result_potential_persistence", run_label, ".csv"))
})


names(potential_persistence_dat) <- site_df$site
potential_persistence_dat <- potential_persistence_dat %>% bind_rows(.id = "site") %>% as_tibble()

potential_persistence_dat <- 
  left_join(potential_persistence_dat,
            potential_persistence_dat %>% filter(age == 5) %>% select(site, year_abn, initial_area_abn = area_ha), 
            by = c("site", "year_abn")) %>%
  mutate(year_abn_bins = cut(year_abn, breaks = 5, include.lowest = TRUE)) %>%
  mutate(year_abn_bins = as_factor(year_abn_bins), # convert both to factors
         time = age - 5,
         time_1 = time + 1,
         cohort = factor(year_abn)) %>% 
  mutate(site = as_factor(site)) %>% filter(!is.na(proportion)) # remove all na proportions


# save potential_persistence_dat
write_csv(potential_persistence_dat, file = paste0(p_derived2, "potential_persistence_dat", run_label, ".csv"))

```

```{r load-data}

# load combined persistence, area, and turnover datasets:

persistence_dat <- read_csv(file = paste0(p_derived2, "persistence_dat", run_label, ".csv")) %>%
  mutate(site = as.factor(site),
         bins = as.factor(bins),
         year_abn_bins = as.factor(year_abn_bins),
         cohort = as.factor(cohort))

area_dat <- read_csv(file = paste0(p_derived2, "area_dat", run_label, ".csv"))
turnover_dat <- read_csv(file = paste0(p_derived2, "turnover_dat", run_label, ".csv"))
potential_area_dat <- read_csv(file = paste0(p_derived2, "potential_area_dat", run_label, ".csv"))


site_df <- read.csv(file = paste0(p_dat_derived, "site_df.csv"))

```


# Mean abandonment duration

This involved using "length_distill_df" (a summary of the number of abandonment periods of specific lengths of time at each site over the course of the entire time series, see "cluster/3_distill_lengths.R") to produce "mean_length_df," a data.frame with the mean, median, and standard deviation, for each site, for both "all" lengths or just the "max" length per pixel, and for a range of abandonment definitions (1, 3, 5, 7, and 10 years).

"mean_length_df" was then used to produce "summary_stats_all_sites" (summary stats across the 11 sites, e.g., means of means, means of standard deviations, etc.). "length_distill_df" was used to produce a similar summary table called "summary_stats_all_sites_pooled."

```{r calc-mean-lengths}
# -------------------------------------------------------- #
# Christopher Crawford, Princeton University, March 9th, 2021

# Script to produce length plots based on distilled length data.tables from "3_distill_lengths.R"
# -------------------------------------------------------- #
# Download via _3_4.slurm:
# cd /Users/christophercrawford/work/projects/abandonment_trajectories/data_derived/_2022_01_31
# scp clc6@della:/scratch/gpfs/clc6/abandonment_trajectories/data_derived/input_rasters/*result*_2022_01_31* ./
# scp clc6@della:/scratch/gpfs/clc6/abandonment_trajectories/data_derived/input_rasters/length_distill_df_2022_01_31.csv ./
# scp clc6@della:/scratch/gpfs/clc6/abandonment_trajectories/data_derived/input_rasters/recult_length_distill_df_2022_01_31.csv ./
# scp clc6@della:/scratch/gpfs/clc6/abandonment_trajectories/data_derived/abn_*_2022_01_31* ../


# set up parameters:
# data.frame of all sites contains information about sites
site_df

# 0. Load distilled length data
length_distill_df <- read_csv(file = paste0(p_derived2,
                                            "length_distill_df", run_label, ".csv"))
# length_distill_df %>% filter(site == "shaanxi") %>% select(length_type) %>% unique()


mean_length_df <- lapply(c(1, 3, 5, 7, 10), function(x) {
  length_distill_df %>% as_tibble() %>% 
    mutate(product = length*freq) %>% 
    filter(length >= x) %>% # to filter by length # this is important for max, since some pixels have max length of 0
    group_by(site, length_type) %>% 
    summarise(
      n_abn_periods = sum(freq), # number of abandonment periods at each site:
      mean_duration = sum(product)/sum(freq),
      median_duration = median(rep(length, freq)),
      sd_duration = sd(rep(length, freq))
      ) %>% 
    mutate(abn_threshold = x) 
}) %>% 
  bind_rows() %>% 
  ungroup() %>%

# add order for plotting:
  left_join(., 
            filter(., abn_threshold == 5, length_type == "all") %>% 
              arrange(mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order)
            )


# save mean_length_df
write_csv(mean_length_df, file = paste0(p_derived2, "mean_length_df", run_label, ".csv"))
cat(fill = TRUE, "Saved mean_length_df to:", paste0(p_derived2, "mean_length_df", run_label, ".csv"))

# mean_length_df <- read_csv(file = paste0(p_derived2, "mean_length_df", run_label, ".csv"))
```





```{r all-sites-summary-stats, include = FALSE}
# ---------------------------------------------------------------- #
# 1. Take mean of the site mean lengths:
# note, this is the value given for the horizontal dashed lines on the main mean abn duration figure
# knitr::include_graphics(path = paste0(p_plots, run_label, "/mean_lengths", run_label, ".png"))


# load files ------------- #
# load all length data (distilled)
length_distill_df <- read_csv(file = paste0(p_derived2, 
                                            "length_distill_df", run_label, ".csv"))

# load mean length data:
mean_length_df <- read_csv(file = paste0(p_derived2, "mean_length_df", run_label, ".csv"))

mean_length_df %>% filter(abn_threshold == 5, length_type == "all") %>% .$sd_duration %>% summary()

summary_stats_all <- mean_length_df %>% 
  group_by(abn_threshold, length_type) %>%
  summarise(mean_of_means = mean(mean_duration, na.rm = TRUE),
            mean_of_medians = mean(median_duration, na.rm = TRUE),
            mean_of_sds = mean(sd_duration, na.rm = TRUE)
            )

# just # our standard 5 year abandonment threshold: 
summary_stats_all_sites <- mean_length_df %>%
  filter(abn_threshold == 5,
         length_type == "all") %>% 
  group_by(abn_threshold, length_type) %>%
  summarise(mean_of_means = mean(mean_duration),
            mean_of_medians = mean(median_duration),
            sd_of_means = sd(mean_duration),
            mean_of_sds = mean(sd_duration),
            mean_n_abn_periods = mean(n_abn_periods)
            )

summary_stats_all_sites$mean_of_means
summary_stats_all_sites$sd_of_means
summary_stats_all_sites$mean_of_sds

sd(filter(mean_length_df, length_type == "all", abn_threshold == 5)$mean_duration)


# ---------------------------------------------------------------- #
# 2. Pool all raw abandonment periods together
summary_stats_all_sites_pooled <- length_distill_df %>%
  as_tibble() %>%
  filter(length >= 5, length_type == "all") %>%
  mutate(product = length*freq) %>%
  group_by() %>%
  summarise(
    total_n_abn_periods = sum(freq),
    mean_duration = sum(product)/sum(freq),
    median_duration = median(rep(length, freq)),
    # mean_test = mean(rep(length, freq)), # same as above
    sd_duration = sd(rep(length, freq))
    )

summary_stats_all_sites_pooled$mean_duration

# alternatively, calculate the weighted mean, using the number of abandonment periods at each site:

mean_length_df %>% 
  filter(length_type == "all", abn_threshold == 5) %>% 
  mutate(prod = mean_duration * n_abn_periods) %>%
  group_by() %>%
  summarise(sum_prod = sum(prod),
            sum_n = sum(n_abn_periods),
            w_mean = sum_prod/sum_n)


# save files:

write_csv(summary_stats_all_sites, 
          file = paste0(p_derived2, "summary_stats_all_sites", run_label, ".csv"))
write_csv(summary_stats_all_sites_pooled, 
          file = paste0(p_derived2, "summary_stats_all_sites_pooled", run_label, ".csv"))


# summary_stats_all_sites <- read_csv(file = paste0(p_derived2, "summary_stats_all_sites", run_label, ".csv"))
# summary_stats_all_sites_pooled <- read_csv(file = paste0(p_derived2, "summary_stats_all_sites_pooled", run_label, ".csv"))
```

## Save mean abandonment duration plots

1. Summary figure: mean length (all periods and max lengths) for each site
2. Summary lengths by abandonment threshold (like 1, but with different thresholds)

```{r plot-mean-abn-duration}
# create directory for plot outputs
if(!dir.exists(paste0(p_output, "plots/", run_label))) {
  dir.create(paste0(p_output, "plots/", run_label))
}

# -------------------------------------------------------- #
# 1. Plot mean length of time abandoned (>= 5) ----

gg_mean_length <-
  ggplot(data = mean_length_df %>%
           filter(abn_threshold == "5"),
         mapping = aes(x = fct_reorder(site, order), #fct_reorder2(site, abn_threshold, desc(mean_duration)),
                       y = mean_duration, 
                       color = length_type)) + 
  theme_classic() +
  theme(axis.text.x = element_text(angle = 330, vjust = 1, hjust = 0),
        plot.caption = element_text(face = "italic")) +
  labs(#title = "Mean length of abandonment",
       #caption = "Counting only recultivation of 2 or more continuous years",
       linetype = NULL,
       color = "Mean of:", 
       y = "Mean abandonment duration (years)", 
       x = "Site") +
  geom_point(size = 2, alpha = 0.85) + 
  geom_hline(data = mean_length_df %>% 
               group_by(abn_threshold, length_type) %>%
               summarise(overall_mean = mean(mean_duration, na.rm = TRUE)) %>% 
               filter(abn_threshold == "5"), 
             mapping = aes(yintercept = overall_mean, 
                           col = length_type, 
                           linetype = "Mean \nAcross \nSites")) + 
  scale_linetype_manual(values = c("Mean \nAcross \nSites" = "dashed")) + 
  scale_color_discrete(labels = c(all = "All periods", 
                                   max = "Max. length \nper pixel")) +
  # scale_shape_manual(values = c(16, 21), # filled and closed circles
  #                    labels = c(all = "All periods", 
  #                               max = "Max. length \nper pixel")) + 
  scale_x_discrete(breaks = site_df$site,
                   labels = site_df$description) +
  scale_y_continuous(breaks = c(12:20)) + 
  guides(color = guide_legend(order = 1))
    


# ----------------- save ---------------- #
ggsave(plot = gg_mean_length,
       filename = paste0(p_plots, run_label, "/mean_lengths_long", run_label, ".pdf"), 
    width = 5.5, height = 5, units = "in")

# png(filename = paste0(p_output, "plots/", run_label, "/mean_lengths", run_label, ".png"), 
#     # width = 6, height = 5, 
#     width = 5, height = 4, 
#     units = "in", res = 400)
# print(gg_mean_length)
# dev.off()
# 
# library(colorblindr)
# cvd_grid(gg_mean_length)
```


```{r mean-duration-across-thresholds}
# 2. Plot mean length across multiple abandonment thresholds ------

gg_mean_length_by_threshold_old <-
  ggplot(data = mean_length_df,
         mapping = aes(x = fct_reorder(site, mean_duration), y = mean_duration, fill = fct_reorder(site, mean_duration))) + 
  theme_classic() +
  theme(
    axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0, size = 7),
        # plot.caption = element_text(face = "italic"),
        legend.position = "bottom") +
  labs(
    # title = "Mean length of abandonment",
       #caption = "Counting only recultivation of 2 or more continuous years",
       linetype = NULL,
       fill = "Site", 
       y = "Mean length of time abandoned (years)",
       x = "Site") +
  geom_col(width = 0.75) + #coord_flip() + 
  facet_grid(
    length_type ~ abn_threshold,
    # abn_threshold ~ length_type,
             labeller = labeller(length_type = c(all = "all lengths", #old = "new",
                                                 max = "max length per pixel"),
                                 abn_threshold = c("1" = "abn threshold = 1",
                                                   "3" = "abn threshold = 3",
                                                   "5" = "abn threshold = 5",
                                                   "7" = "abn threshold = 7",
                                                   "10" = "abn threshold = 10"))) +
  geom_hline(data = mean_length_df %>% 
               group_by(abn_threshold, length_type) %>%
               summarise(overall_mean = mean(mean_duration, na.rm = TRUE)), 
             mapping = aes(yintercept = overall_mean, linetype = "Mean across sites"),
             show.legend = TRUE,
             color = "black", size = 0.75) + 
  scale_linetype_manual(values = c("Mean across sites" = "dashed")) + 
  guides(
    # fill = guide_legend(override.aes = list(linetype = 0))
    fill = "none"
         )

gg_mean_length_by_threshold_side <-
  ggplot(data = mean_length_df,
         mapping = aes(x = fct_reorder(site, desc(mean_duration)), 
                       y = mean_duration, fill = as_factor(abn_threshold))) + 
  theme_classic() +
  theme(legend.position = "bottom",
        axis.text.x = element_text(angle = 345, vjust = 1, hjust = 0)) +
  labs(linetype = NULL,
       fill = "Abandonment threshold (years)", 
       y = "Mean length of time abandoned (years)",
       x = "Site") +
  geom_col(width = 0.75, position = position_dodge()) + #coord_flip() +
  # geom_point(size = 2, position = position_dodge()) + #coord_flip() + 
  facet_grid(vars(length_type),
             labeller = labeller(length_type = c(all = "all lengths", #old = "new",
                                                 max = "max length per pixel"))) +
  geom_hline(data = mean_length_df %>% 
               group_by(abn_threshold, length_type) %>%
               summarise(overall_mean = mean(mean_duration, na.rm = TRUE)), 
             mapping = aes(yintercept = overall_mean, 
                           linetype = "Mean across sites",
                           color = as_factor(abn_threshold)),
             show.legend = TRUE, #color = "black",
             size = 0.75
             ) + 
  scale_linetype_manual(values = c("Mean across sites" = "dashed")) + 
  guides(
    fill = guide_legend(override.aes = list(linetype = 0)),
    # fill = "none",
    color = "none"
         ) 

# ----------------- save ---------------- #

ggsave(plot = gg_mean_length_by_threshold_old,
       filename = paste0(p_plots, run_label, "/mean_lengths_by_threshold_old", run_label, ".pdf"), 
    width = 8.5, height = 5, units = "in")

ggsave(plot = gg_mean_length_by_threshold_side + 
         scale_fill_manual(values = viridis(n = 5)) + 
         scale_color_manual(values = viridis(n = 5)),
       filename = paste0(p_plots, run_label, "/mean_lengths_by_threshold_side", run_label, ".pdf"), 
    width = 8.5, height = 5, units = "in")








gg_mean_length_by_threshold_flipped <-
  ggplot(data = mean_length_df,
         mapping = aes(y = fct_reorder(site, mean_duration), 
                       x = mean_duration, fill = as_factor(abn_threshold))) + 
  theme_classic() +
  labs(linetype = NULL,
       fill = "Abandonment\nthreshold\n(years)", 
       x = "Mean length of time abandoned (years)",
       y = "Site") +
  geom_col(width = 0.75,
           position = position_dodge()) +
  scale_y_discrete(labels = fancy_labels) +
  facet_grid(.~length_type,
             labeller = labeller(length_type = c(all = "All periods", #old = "new",
                                                 max = "Max. length per pixel"))) +
  geom_vline(data = mean_length_df %>% 
               group_by(abn_threshold, length_type) %>%
               summarise(overall_mean = mean(mean_duration, na.rm = TRUE)), 
             mapping = aes(xintercept = overall_mean, 
                           linetype = "Mean",
                           color = as_factor(abn_threshold)),
             show.legend = TRUE, #color = "black",
             size = 0.75
             ) + 
  scale_linetype_manual(values = c("Mean" = "dashed"), labels = "Mean\nacross\nsites") +
  guides(
    fill = guide_legend(override.aes = list(linetype = 0)),
    # fill = "none",
    color = "none"
    )  + 
  scale_fill_manual(values = viridis(n = 5)) + 
  scale_color_manual(values = viridis(n = 5)) + 
  coord_cartesian(xlim = c(5, 21))


ggsave(plot = gg_mean_length_by_threshold_flipped,
       filename = paste0(p_plots, run_label, "/mean_lengths_by_threshold", run_label, ".pdf"), 
    width = 8.5, height = 6, units = "in")
```

```{r recult-duration-by-threshold}

gg_mean_recult_length_by_threshold_flipped <-
  ggplot(data = mean_recult_length_df,
         mapping = aes(y = fct_reorder(site, mean_duration), 
                       x = mean_duration, fill = as_factor(recult_threshold))) + 
  theme_classic() +
  labs(linetype = NULL,
       fill = "Recultivation\nthreshold\n(years)", 
       x = "Mean length of time recultivated (years)",
       y = "Site") +
  geom_col(width = 0.75,
           position = position_dodge()) +
  scale_y_discrete(labels = fancy_labels) +
  # facet_grid(.~length_type, 
  #            labeller = labeller(length_type = c(all = "All periods", #old = "new",
  #                                                max = "Max. length per pixel"))) +
  geom_vline(data = mean_recult_length_df %>% 
               group_by(recult_threshold, length_type) %>%
               summarise(overall_mean = mean(mean_duration, na.rm = TRUE)), 
             mapping = aes(xintercept = overall_mean, 
                           linetype = "Mean",
                           color = as_factor(recult_threshold)),
             show.legend = TRUE, #color = "black",
             size = 0.75
             ) + 
  scale_linetype_manual(values = c("Mean" = "dashed"), labels = "Mean\nacross\nsites") +
  guides(
    fill = guide_legend(override.aes = list(linetype = 0)),
    # fill = "none",
    color = "none"
    )  + 
  scale_fill_manual(values = viridis(n = 5)) + 
  scale_color_manual(values = viridis(n = 5)) #+ 
  # coord_cartesian(xlim = c(5, 21))


ggsave(plot = gg_mean_recult_length_by_threshold_flipped,
       filename = paste0(p_plots, run_label, "/mean_recult_lengths_by_threshold", run_label, ".pdf"), 
    width = 6, height = 5, units = "in")


```



# Distributions of abandonment lengths (durations)

```{r histograms-of-lengths}
# Plot all site histograms together:

gg_raw_hist <-
  ggplot(data = length_distill_df %>% 
           filter(#site == site_df$site[6], 
             length_type == "all",
             length >= 5) %>%
           left_join(., length_distribution %>% select(site, order) %>% unique(), by = "site")) + 
  theme_classic() +
  xlab("Time abandoned (years)") + ylab(expression("Count  (10"^{6}*" pixels)")) +
  geom_col(mapping = aes(x = length, y = freq/(10^6))) +
  facet_grid(rows = vars(fct_reorder(site, order)), scales = "free") +
  scale_x_continuous(breaks = seq(from = 5, to = 30, by = 5))

ggsave(plot = gg_raw_hist, 
       filename = paste0(p_plots, run_label, "/mean_lengths_default_hist", run_label, ".pdf"), 
       width = 5, height = 8, units = "in")


# 3. Save Individual Histograms ----
# For all periods and max lengths for each site
# -------------------------------------------------------------------- #
# ----------------- plot histograms individually  -------------------- #
# -------------------------------------------------------------------- #

for (i in 1:11) {
  gg_hist <-
    ggplot(data = filter(length_distill_df, site == site_df$site[i], length > 0)) + 
    theme_classic() +
    labs(linetype = "", x = "Time abandoned (years)", 
         y = expression("Count  (10"^{6}*" pixels)"),
         color = "Mean") +
    geom_col(mapping = aes(x = length, y = freq/(10^6)), fill = "gray70") +
    facet_grid(rows = vars(length_type), scales = "free",
               labeller = labeller(length_type = c(all = "all abn. periods", #old = "new",
                                                   max = "max. length per pixel"))) +
    # means
    geom_vline(data = filter(mean_length_df, site == site_df$site[i], 
                             !abn_threshold %in% c(3)), 
               mapping = aes(xintercept = mean_duration, 
                             color = as_factor(abn_threshold)#, linetype = "Mean (>=1)"
                             ),
               show.legend = TRUE, size = 1, linetype = "dashed") + 

    scale_color_manual(values = viridis(n = 4), #c("5" = "#C51B7D", "1" = "#F1B6DA"),
                       labels = c("5" = "Mean (\u2265 5)", "1" = "Mean (\u2265 1)",
                                  "7" = "Mean (\u2265 7)", "10" = "Mean (\u2265 10)")) +

    theme(#legend.position = c(0.85, 0.95)
      )

  # save
  png(filename = paste0(p_output, "plots/", run_label, "/length_hist_all", 
                        # "/length_hist", 
                        run_label, site_df$label[i], ".png"), 
      width = 7, height = 5, units = "in", res = 400)
  print(gg_hist)
  dev.off()
}

```


```{r length_distribution}
# Create distribution of observations for violin plots, density plots, histograms, etc.
# This is essentially a recreation of the length distribution, using length_distill, but condensed so as to make plotting manageable.

# producing a tibble called "length_distribution"

length_distribution <- length_distill_df %>% as_tibble() %>% 
    filter(length >= 5,
           # site == "shaanxi",
           length_type == "all") %>% # to filter by length # this is important for max, since some pixels have max length of 0 
  mutate(fr = round(freq/1000, 
                    digits = 0)) %>% 
  group_by(site) %>%
  summarise(site = unique(site), 
            length = rep(length, fr),
            mean = mean(length))


# adding order based on site-level mean abandonment duration (all periods)
length_distribution <- length_distribution %>%
  left_join(., mean_length_df %>% 
              filter(abn_threshold == "5", length_type == "all") %>% 
              arrange(length_type, mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order),
            by = "site")



# -------------------------------------------------------------- #
# the same as length_distribution, but with just the maximum periods, not all periods
# -------------------------------------------------------------- #
length_distribution_max <- length_distill_df %>% as_tibble() %>% 
    filter(length >= 5, length_type == "max") %>% # to filter by length # this is important for max, since some pixels have max length of 0 
  mutate(fr = round(freq/1000, digits = 0)) %>% 
  group_by(site) %>%
  summarise(site = unique(site), 
            length = rep(length, fr),
            mean = mean(length))

# adding order based on site-level mean abandonment duration (max length per pixel)

length_distribution_max <- length_distribution_max %>%
  left_join(., 
            mean_length_df %>% 
              filter(abn_threshold == "5", length_type == "max") %>% 
              arrange(length_type, mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order),
            by = "site")



# -------------------------------------------------------------- #
# recultivation length_distribution
# -------------------------------------------------------------- #
# Note: no recultivation threshold being used.

recult_length_distribution <- recult_length_distill_df %>% as_tibble() %>% 
    filter(#length >= 5,  # no recultivation threshold being used
           length_type == "recult") %>% 
  mutate(fr = round(freq/1000, digits = 0)) %>% 
  group_by(site) %>%
  summarise(site = unique(site), 
            length = rep(length, fr),
            mean = mean(length))

# adding order based on site-level mean abandonment duration (max length per pixel)

recult_length_distribution <- recult_length_distribution %>%
  left_join(., 
            mean_recult_length_df %>% 
              filter(
                recult_threshold == "1",
                length_type == "recult") %>% 
              arrange(length_type, mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order),
            by = "site")


# Save files:
write_csv(length_distribution,
          file = paste0(p_derived2, "length_distribution", run_label, ".csv"))

write_csv(length_distribution_max,
          file = paste0(p_derived2, "length_distribution_max", run_label, ".csv"))

write_csv(recult_length_distribution,
          file = paste0(p_derived2, "recult_length_distribution", run_label, ".csv"))

# load files
# length_distribution <- read_csv(file = paste0(p_derived2, "length_distribution", run_label, ".csv"))
# length_distribution_max <- read_csv(file = paste0(p_derived2, "length_distribution_max", run_label, ".csv"))
# recult_length_distribution <- read_csv(file = paste0(p_derived2, "recult_length_distribution", run_label, ".csv"))
```




## Potential abandonment

```{r potential-mean-lengths}
mean_length_df
# 0. Load distilled length data
potential_length_distill_df <- read_csv(
  file = paste0(p_derived2,
                "potential_length_distill_df", run_label, ".csv"))



potential_mean_length_df <- lapply(c(1, 3, 5, 7, 10), function(x) {
  potential_length_distill_df %>% as_tibble() %>% 
    mutate(product = length*freq) %>% 
    filter(length >= x) %>% # to filter by length # this is important for max, since some pixels have max length of 0
    group_by(site, length_type) %>% 
    summarise(
      n_abn_periods = sum(freq), # number of abandonment periods at each site:
      mean_duration = sum(product)/sum(freq),
      median_duration = median(rep(length, freq)),
      sd_duration = sd(rep(length, freq))
      ) %>% 
    mutate(abn_threshold = x) 
}) %>% 
  bind_rows() %>% 
  ungroup()

# add order for plotting:
potential_mean_length_df <- 
  potential_mean_length_df %>% 
  left_join(., 
            filter(., abn_threshold == 5, 
                   length_type == "potential") %>% 
              arrange(mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order)
            )


# save potential_mean_length_df
write_csv(potential_mean_length_df, 
          file = paste0(p_derived2, "potential_mean_length_df", run_label, ".csv"))
cat(fill = TRUE, "Saved potential_mean_length_df to:", paste0(p_derived2, "potential_mean_length_df", run_label, ".csv"))

# potential_mean_length_df <- read_csv(file = paste0(p_derived2, "potential_mean_length_df", run_label, ".csv"))
```

```{r potential-summary-stats}
potential_mean_length_df
endpoint_half_lives$endpoint %>% unique()

# load files ------------- #
# load all length data (distilled)
potential_length_distill_df <- read_csv(
  file = paste0(p_derived2, 
                "potential_length_distill_df", run_label, ".csv"))

# load mean length data:
potential_mean_length_df <- read_csv(
  file = paste0(p_derived2, "potential_mean_length_df", run_label, ".csv"))

potential_mean_length_df %>% filter(abn_threshold == 5, length_type == "potential") %>% .$sd_duration %>% summary()

potential_summary_stats_all <- potential_mean_length_df %>% 
  group_by(abn_threshold, length_type) %>%
  summarise(mean_of_means = mean(mean_duration, na.rm = TRUE),
            mean_of_medians = mean(median_duration, na.rm = TRUE),
            mean_of_sds = mean(sd_duration, na.rm = TRUE)
            )

# just # our standard 5 year abandonment threshold: 
potential_summary_stats_all_sites <- potential_mean_length_df %>%
  filter(abn_threshold == 5,
         length_type == "potential") %>% 
  group_by(abn_threshold, length_type) %>%
  summarise(mean_of_means = mean(mean_duration),
            mean_of_medians = mean(median_duration),
            sd_of_means = sd(mean_duration),
            mean_of_sds = mean(sd_duration),
            mean_n_abn_periods = mean(n_abn_periods)
            )

potential_summary_stats_all_sites$mean_of_means
potential_summary_stats_all_sites$sd_of_means
potential_summary_stats_all_sites$mean_of_sds

sd(filter(potential_mean_length_df, 
          length_type == "potential", abn_threshold == 5)$mean_duration)


# ---------------------------------------------------------------- #
# 2. Pool all raw abandonment periods together
potential_summary_stats_all_sites_pooled <- potential_length_distill_df %>%
  as_tibble() %>%
  filter(length >= 5, length_type == "potential") %>%
  mutate(product = length*freq) %>%
  group_by() %>%
  summarise(
    total_n_abn_periods = sum(freq),
    mean_duration = sum(product)/sum(freq),
    median_duration = median(rep(length, freq)),
    # mean_test = mean(rep(length, freq)), # same as above
    sd_duration = sd(rep(length, freq))
    )

potential_summary_stats_all_sites_pooled$mean_duration

# alternatively, calculate the weighted mean, using the number of abandonment periods at each site:

potential_mean_length_df %>% 
  filter(length_type == "potential", abn_threshold == 5) %>% 
  mutate(prod = mean_duration * n_abn_periods) %>%
  group_by() %>%
  summarise(sum_prod = sum(prod),
            sum_n = sum(n_abn_periods),
            w_mean = sum_prod/sum_n)


# save files:

write_csv(potential_summary_stats_all_sites, 
          file = paste0(p_derived2, "potential_summary_stats_all_sites", run_label, ".csv"))
write_csv(potential_summary_stats_all_sites_pooled, 
          file = paste0(p_derived2, "potential_summary_stats_all_sites_pooled", run_label, ".csv"))


# potential_summary_stats_all_sites <- read_csv(file = paste0(p_derived2, "potential_summary_stats_all_sites", run_label, ".csv"))
# potential_summary_stats_all_sites_pooled <- read_csv(file = paste0(p_derived2, "potential_summary_stats_all_sites_pooled", run_label, ".csv"))
```


```{r potential-length-dist}

# -------------------------------------------------------------- #
# potential area abandoned, assuming no recultivation:
# -------------------------------------------------------------- #
potential_length_distribution

potential_length_distribution <- 
  potential_length_distill_df %>% as_tibble() %>% 
    filter(length >= 5,
           # site == "shaanxi",
           length_type == "potential") %>% # to filter by length # this is important for max, since some pixels have max length of 0 
  mutate(fr = round(freq/1000, 
                    digits = 0)) %>% 
  group_by(site) %>%
  summarise(site = unique(site), 
            length = rep(length, fr),
            mean = mean(length))


# adding order based on site-level mean abandonment duration (all periods)
potential_length_distribution <- 
  potential_length_distribution %>%
  left_join(., potential_mean_length_df %>% 
              filter(abn_threshold == "5", 
                     length_type == "potential") %>% 
              arrange(length_type, mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order),
            by = "site")



write_csv(potential_length_distribution,
          file = paste0(p_derived2, "potential_length_distribution", run_label, ".csv"))
# potential_length_distribution <- read_csv(file = paste0(p_derived2, "potential_length_distribution", run_label, ".csv"))

```





## Recultivation 

```{r calc-mean-length-recultivation}
# 0. Load distilled length data
recult_length_distill_df <- read_csv(file = paste0(p_derived2, 
                                            "recult_length_distill_df", run_label, ".csv"))


mean_recult_length_df <- lapply(c(1, 3, 5, 7, 10), function(x) {
  recult_length_distill_df %>% as_tibble() %>% 
    mutate(product = length*freq) %>% 
    filter(length >= x) %>% # to filter by length # this is important for max, since some pixels have max length of 0
    group_by(site, length_type) %>% 
    summarise(
      n_abn_periods = sum(freq), # number of abandonment periods at each site:
      mean_duration = sum(product)/sum(freq),
      median_duration = median(rep(length, freq)),
      sd_duration = sd(rep(length, freq))
      ) %>% 
    mutate(recult_threshold = x) 
}) %>% bind_rows() %>% ungroup()

# add order for plotting:
mean_recult_length_df <-
  mean_recult_length_df %>%
  left_join(., 
            filter(., 
                   recult_threshold == 1,  # using the threshold of 1 (i.e., no threshold, for plot ordering)
                   length_type == "recult") %>% 
              arrange(mean_duration) %>% 
              mutate(order = 1:n()) %>% select(site, order)
            )

# save mean_length_df
write_csv(mean_recult_length_df, file = paste0(p_derived2, "mean_recult_length_df", run_label, ".csv"))
cat(fill = TRUE, "Saved mean_recult_length_df to:", paste0(p_derived2, "mean_recult_length_df", run_label, ".csv"))


# mean_recult_length_df <- read_csv(file = paste0(p_derived2, "mean_recult_length_df", run_label, ".csv"))
```

```{r recult-summary-stats}
# ---------------------------------------------------------------- #
# 1. Take mean of the site mean lengths:
# note, this is the value given for the horizontal dashed lines on the main mean abn duration figure
# knitr::include_graphics(path = paste0(p_plots, run_label, "/mean_lengths", run_label, ".png"))


# load files ------------- #
# load all recultivation length data (distilled)
recult_length_distill_df <- read_csv(file = paste0(p_derived2, 
                                            "recult_length_distill_df", run_label, ".csv"))

# load mean length data:
mean_recult_length_df <- read_csv(file = paste0(p_derived2, "mean_recult_length_df", run_label, ".csv"))

mean_recult_length_df %>% 
  filter(recult_threshold == 1, 
         length_type == "recult") %>% 
  .$sd_duration %>% summary()

recult_summary_stats_all <- 
  mean_recult_length_df %>% 
  group_by(recult_threshold, length_type) %>%
  summarise(mean_of_means = mean(mean_duration, na.rm = TRUE),
            mean_of_medians = mean(median_duration, na.rm = TRUE),
            mean_of_sds = mean(sd_duration, na.rm = TRUE),
            mean_n_abn_periods = mean(n_abn_periods)
            )

# just # our 1 year recultivation threshold: 
recult_summary_stats_all_sites <-
  mean_recult_length_df %>%
  filter(recult_threshold == 1, length_type == "recult") %>%
  group_by(recult_threshold, length_type) %>%
  summarise(mean_of_means = mean(mean_duration),
            mean_of_medians = mean(median_duration),
            sd_of_means = sd(mean_duration),
            mean_of_sds = mean(sd_duration),
            mean_n_abn_periods = mean(n_abn_periods),
            n_abn_periods = sum(n_abn_periods)
            )

recult_summary_stats_all_sites$mean_of_means
recult_summary_stats_all_sites$sd_of_means
recult_summary_stats_all_sites$mean_of_sds

sd(filter(mean_recult_length_df, length_type == "recult", recult_threshold == 1)$mean_duration)


# ---------------------------------------------------------------- #
# 2. Pool all raw abandonment periods together
recult_summary_stats_all_sites_pooled <- recult_length_distill_df %>%
  as_tibble() %>%
  filter(length >= 1, length_type == "recult") %>%
  mutate(product = length*freq) %>%
  group_by() %>%
  summarise(
    total_n_abn_periods = sum(freq),
    mean_duration = sum(product)/sum(freq),
    median_duration = median(rep(length, freq)),
    # mean_test = mean(rep(length, freq)), # same as above
    sd_duration = sd(rep(length, freq))
    )

recult_summary_stats_all_sites_pooled$mean_duration

# alternatively, calculate the weighted mean, using the number of abandonment periods at each site:

mean_recult_length_df %>% 
  filter(length_type == "recult", recult_threshold == 1) %>% 
  mutate(prod = mean_duration * n_abn_periods) %>%
  group_by() %>%
  summarise(sum_prod = sum(prod),
            sum_n = sum(n_abn_periods),
            w_mean = sum_prod/sum_n)


# save files:

write_csv(recult_summary_stats_all_sites, 
          file = paste0(p_derived2, "recult_summary_stats_all_sites", run_label, ".csv"))
write_csv(recult_summary_stats_all_sites_pooled, 
          file = paste0(p_derived2, "recult_summary_stats_all_sites_pooled", run_label, ".csv"))


# recult_summary_stats_all_sites <- read_csv(file = paste0(p_derived2, "recult_summary_stats_all_sites", run_label, ".csv"))
# recult_summary_stats_all_sites_pooled <- read_csv(file = paste0(p_derived2, "recult_summary_stats_all_sites_pooled", run_label, ".csv"))
```


## Final distribution figures (*manual*)

The observed abandonment distribution figure was initially Figure 2 in the draft manuscript, which is why these figures are collectively referred to as Fig2 variations.

```{r man-fig2-obs}

gg_fig2_manual <-
  length_distill_df %>% 
  filter(length >= 5, length_type == "all") %>%
  bind_rows(
    tibble(length = 4, freq = 0, site = site_df$site, length_type = "all"),
    tibble(length = 31, freq = 0, site = site_df$site, length_type = "all")) %>%
  left_join(mean_length_df %>% filter(abn_threshold == 5, length_type == "all"),
            by = c("site")
            ) %>%
  mutate(height = freq / n_abn_periods,
         site = fct_reorder(site, order)) %>%
  # .$site %>% levels()
  
  # plot
  ggplot() + 
  geom_col(aes(x = length, y = height, group = site),
           fill = "gray70", color = "gray70") +
  geom_step(aes(x = length, y = height, group = site),
            direction = "mid", size = 0.5) +
  geom_hline(yintercept = 0) +
  geom_point(
    data = mean_length_df %>%
      filter(abn_threshold == "5", length_type == "all") %>%
      mutate(site = fct_reorder(site, order)),
    mapping = aes(y = 0, x = mean_duration, color = length_type),
    inherit.aes = FALSE,
    size = 3, alpha = 0.85) +
  scale_alpha_manual(values=c(0.5))+
  facet_grid(vars(site), scales = "free_x",
             labeller = as_labeller(fancy_labels)
             ) + 
  theme_classic() + 
  scale_y_continuous(
    # breaks = c(0, 0.1, 0.2), labels = c("0", "0.1", "0.2")
                     n.breaks = 3) + 
  labs(y = "Proportion of abandonment periods by duration", 
       x = "Abandonment duration (years)",
       color = NULL, linetype = NULL) +
  geom_vline(
    data = . %>% select(site, mean_duration) %>% unique() %>% 
      summarise(overall_mean = mean(mean_duration, na.rm = TRUE)),
    mapping = aes(xintercept = overall_mean,
                  linetype = "Mean across all sites"),
    color = hue_pal()(3)[1]) +
  
  scale_linetype_manual(values = c("Mean across all sites" = "dashed")) +
  scale_color_discrete(labels = c(potential = "Site-level mean")) +
  
  scale_x_continuous(breaks = c(5, 10, 15, 20, 25, 30), expand = c(0, 0)) +
  guides(color = guide_legend(order = 1, nrow = 1, 
                              override.aes = list(linetype = 0)), 
         linetype = guide_legend(order = 2, nrow = 1)) + 
  coord_cartesian(clip = "off") +
  theme(axis.line.x = element_blank(),
        axis.text.y = element_text(vjust = 0.35, size = 6),
        strip.text.y.right = element_text(size = 8, angle = 0, hjust = 0, vjust = 0),
        strip.background = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        plot.background = element_blank(),
        panel.spacing = unit(1, "mm"),
        panel.grid.major.x = element_line(),
        legend.spacing.x = unit(0.5, "mm"),
        legend.position = c(0.6, -0.12),
        legend.direction = "horizontal",
        legend.box = "horizontal",
        plot.margin = margin(5, 0, 25, 5),
        legend.background = element_blank(),
        legend.box.just = "left",
        legend.text.align = 0,
        legend.title.align = 1)

  
# save plot
ggsave(plot = gg_fig2_manual,
       filename = paste0(p_plots, "/fig2", run_label, ".pdf"), 
    width = 3.5, height = 5, units = "in")

```
```{r man-fig2-max}

gg_fig2_manual_max <-
  length_distill_df %>% 
  filter(length >= 5, length_type == "max") %>%
  bind_rows(
    tibble(length = 4, freq = 0, site = site_df$site, length_type = "max"),
    tibble(length = 31, freq = 0, site = site_df$site, length_type = "max")) %>%
  left_join(mean_length_df %>% filter(abn_threshold == 5, length_type == "max") %>%
              arrange(mean_duration) %>%
              mutate(order = 1:11),
            by = c("site")
            ) %>%
  mutate(height = freq / n_abn_periods,
         site = fct_reorder(site, order)) %>%
  # .$site %>% levels()
  
  # plot
  ggplot() + 
  geom_col(aes(x = length, y = height, group = site),
           fill = "gray70", color = "gray70") +
  geom_step(aes(x = length, y = height, group = site),
            direction = "mid", size = 0.5) +
  geom_hline(yintercept = 0) +
  geom_vline(
    data = mean_length_df %>%
      filter(abn_threshold == "5"#, length_type == "all"
             ) %>%
      mutate(site = fct_reorder(site, order)) %>%
      group_by(length_type) %>% 
      summarise(overall_mean = mean(mean_duration, na.rm = TRUE)),
    mapping = aes(xintercept = overall_mean, color = length_type, 
                  linetype = "Mean\nacross\nall sites")) +
  scale_linetype_manual(values = c("Mean\nacross\nall sites" = "dashed")) +

  geom_point(
    data = mean_length_df %>%
      filter(abn_threshold == "5"#, length_type == "all"
             ) %>%
      mutate(site = fct_reorder(site, order)),
    mapping = aes(y = 0, x = mean_duration, color = length_type),
    inherit.aes = FALSE,
    size = 3, alpha = 0.85) +
  scale_alpha_manual(values=c(0.5))+
  facet_grid(vars(site), scales = "free_x",
             labeller = as_labeller(fancy_labels)
             ) + 
  scale_color_discrete(labels = c(all = "All abandonment periods",
                                   max = "Maximum duration\nper pixel")) +  
  theme_classic() + 
  scale_y_continuous(
    # breaks = c(0, 0.1, 0.2), labels = c("0", "0.1", "0.2")
                     n.breaks = 3) + 
  labs(y = "Proportion of abandonment periods by duration", 
       x = "Abandonment duration (years)",
       color = "Site-level \nmean of:", linetype = NULL) +
  
  scale_x_continuous(breaks = c(5, 10, 15, 20, 25, 30), expand = c(0, 0)) +
  
  guides(color = guide_legend(order = 1, nrow = 2, 
                              override.aes = list(linetype = 0)), 
         linetype = guide_legend(order = 2, label.vjust = 0)) + 
  coord_cartesian(clip = "off") +
  theme(axis.line.x = element_blank(),
        axis.text.y = element_text(vjust = 0.35, size = 6),
        strip.text.y.right = element_text(size = 8, angle = 0, hjust = 0, vjust = 0),
        strip.background = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        plot.background = element_blank(),
        panel.spacing = unit(1, "mm"),
        panel.grid.major.x = element_line(),
        legend.spacing.x = unit(0.5, "mm"),
        legend.position = c(0.65, -0.155),
        legend.direction = "horizontal",
        legend.box = "horizontal",
        plot.margin = margin(5, 0, 40, 5),
        legend.background = element_blank(),
        legend.box.just = "left",
        legend.text.align = 0,
        legend.title.align = 1)

  
# save plot
ggsave(plot = gg_fig2_manual_max,
       filename = paste0(p_plots, "/fig2_w_max", run_label, ".pdf"), 
    width = 3.5, height = 5, units = "in")


```
```{r man-fig2-potential}

# save plot
gg_fig2_manual_potential <-
  potential_length_distill_df %>% 
  filter(length >= 5) %>%
  bind_rows(
    tibble(length = 4, freq = 0, site = site_df$site, length_type = "potential"),
    tibble(length = 31, freq = 0, site = site_df$site, length_type = "potential")) %>%
  left_join(potential_mean_length_df %>% filter(abn_threshold == 5)) %>%
  mutate(height = freq / n_abn_periods,
         site = fct_reorder(site, order)) %>%
  
  # plot
  ggplot() + 
  geom_col(aes(x = length, y = height, group = site),
           fill = "gray70", color = "gray70") +
  geom_step(aes(x = length, y = height, group = site),
            direction = "mid", size = 0.5) +
  geom_hline(yintercept = 0) +
  geom_point(data = . %>% select(site, mean_duration, length_type) %>% unique(),
             mapping = aes(y = 0, x = mean_duration, color = length_type),
             inherit.aes = FALSE,
             size = 3, alpha = 0.85) +
  scale_alpha_manual(values=c(0.5))+
  facet_grid(vars(site), scales = "free_x",
             labeller = as_labeller(fancy_labels)
             ) + 
  theme_classic() + 
  scale_y_continuous(breaks = c(0, 0.1, 0.2), labels = c("0", "0.1", "0.2")) + 
  labs(y = "Proportion of abandonment periods by duration", 
       # x = "Abandonment duration (years)",
       x = "Potential abandonment duration (years),\nassuming no recultivation",
       color = NULL, linetype = NULL) +
  geom_vline(
    data = . %>% select(site, mean_duration) %>% unique() %>% 
      summarise(overall_mean = mean(mean_duration, na.rm = TRUE)),
    mapping = aes(xintercept = overall_mean,
                  linetype = "Mean across all sites"),
    color = hue_pal()(3)[1]) +
  
  scale_linetype_manual(values = c("Mean across all sites" = "dashed")) +
  scale_color_discrete(labels = c(potential = "Site-level mean")) +
  
  scale_x_continuous(breaks = c(5, 10, 15, 20, 25, 30), expand = c(0, 0)) +
  guides(color = guide_legend(order = 1, nrow = 1, 
                              override.aes = list(linetype = 0)), 
         linetype = guide_legend(order = 2, nrow = 1)) + 
  coord_cartesian(clip = "off") +
  theme(axis.line.x = element_blank(),
        axis.text.y = element_text(vjust = 0.35, size = 6),
        strip.text.y.right = element_text(size = 8, angle = 0, hjust = 0, vjust = 0),
        strip.background = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        plot.background = element_blank(),
        panel.spacing = unit(1, "mm"),
        panel.grid.major.x = element_line(),
        legend.spacing.x = unit(0.5, "mm"),
        legend.position = c(0.6, -0.17),
        legend.direction = "horizontal",
        legend.box = "horizontal",
        plot.margin = margin(5, 5, 27, 5),
        legend.background = element_blank(),
        legend.box.just = "left",
        legend.text.align = 0,
        legend.title.align = 1)

  
# save plot
ggsave(plot = gg_fig2_manual_potential,
       filename = paste0(p_plots, "/fig2_potential", run_label, ".pdf"), 
    width = 3.5, height = 5, units = "in")

```

```{r overlaying-obs-pot-distributions}

gg_fig2_manual_overlap_potential <- length_distill_df %>% 
  select(-site_long) %>%
  filter(length >= 5, length_type == "all") %>%
  bind_rows(
    tibble(length = 4, freq = 0, site = site_df$site, length_type = "all"),
    tibble(length = 31, freq = 0, site = site_df$site, length_type = "all")) %>%
  left_join(mean_length_df %>% filter(abn_threshold == 5, length_type == "all") %>%
              arrange(mean_duration) %>% mutate(order = 1:11),
            by = c("site", "length_type")
            ) %>%
  
  bind_rows(
    potential_length_distill_df %>% 
  filter(length >= 5, length_type == "potential") %>%
  bind_rows(
    tibble(length = 4, freq = 0, site = site_df$site, length_type = "potential"),
    tibble(length = 31, freq = 0, site = site_df$site, length_type = "potential")) %>%
  left_join(potential_mean_length_df %>% filter(abn_threshold == 5, length_type == "potential") %>%
              arrange(mean_duration) %>% mutate(order = 1:11),
            by = c("site", "length_type")
            )) %>%
  
  mutate(height = freq / n_abn_periods,
         site = fct_reorder(site, order)) %>%
  # select(length_type) %>% unique
  # .$site %>% levels()
  
  # plot
  ggplot() + 
  geom_col(aes(x = length, y = height, # group = site, 
               fill = length_type, group = length_type,
               # color = length_type
               ), 
           position = "identity",
           # position = "dodge",
           alpha = 0.5,
           show.legend = FALSE
           # color = "gray70"
           ) +
  # geom_step(aes(x = length, y = height, #fill = length_type, 
  #               group = length_type,
  #              color = length_type),
  #           direction = "mid", size = 0.5,
  #           position = "identity",
  #          alpha = 0.5) +
  geom_hline(yintercept = 0) +
  geom_vline(
    data = bind_rows(
      filter(mean_length_df, abn_threshold == "5", length_type == "all"),
      filter(potential_mean_length_df, abn_threshold == "5", length_type == "potential")) %>%
      mutate(site = fct_reorder(site, order)) %>%
      group_by(length_type) %>% 
      summarise(overall_mean = mean(mean_duration, na.rm = TRUE)),
    mapping = aes(xintercept = overall_mean, color = length_type, 
                  linetype = "Mean\nacross\nall sites")) +
  scale_linetype_manual(values = c("Mean\nacross\nall sites" = "dashed")) +

  geom_point(
    data = bind_rows(
      filter(mean_length_df, abn_threshold == "5", length_type == "all"),
      filter(potential_mean_length_df, abn_threshold == "5", length_type == "potential")) %>%
      mutate(site = fct_reorder(site, order)),
    mapping = aes(y = 0, x = mean_duration, color = length_type),
    inherit.aes = FALSE,
    size = 3, alpha = 0.85) +
  scale_alpha_manual(values=c(0.5))+
  facet_grid(vars(site), scales = "free_x",
             labeller = as_labeller(fancy_labels)
             ) + 
  scale_color_discrete(labels = c(all = "Observed",
                                   potential = "Potential")) +  
  theme_classic() + 
  scale_y_continuous(
    # breaks = c(0, 0.1, 0.2), labels = c("0", "0.1", "0.2")
                     n.breaks = 3) + 
  labs(y = "Proportion of abandonment periods by duration", 
       x = "Abandonment duration (years)",
       color = "Site-level \nmean of:", linetype = NULL, fill = NULL) +
  
  scale_x_continuous(breaks = c(5, 10, 15, 20, 25, 30), expand = c(0, 0)) +
  
  guides(color = guide_legend(order = 1, nrow = 2, 
                              override.aes = list(linetype = 0)), 
         linetype = guide_legend(order = 2, label.vjust = 0)) + 
  coord_cartesian(clip = "off") +
  theme(axis.line.x = element_blank(),
        axis.text.y = element_text(vjust = 0.35, size = 6),
        strip.text.y.right = element_text(size = 8, angle = 0, hjust = 0, vjust = 0),
        strip.background = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        plot.background = element_blank(),
        panel.spacing = unit(1, "mm"),
        panel.grid.major.x = element_line(),
        legend.spacing.x = unit(0.5, "mm"),
        legend.position = c(0.65, -0.155),
        legend.direction = "horizontal",
        legend.box = "horizontal",
        plot.margin = margin(5, 0, 40, 5),
        legend.background = element_blank(),
        legend.box.just = "left",
        legend.text.align = 0,
        legend.title.align = 1)


# save plot
ggsave(plot = gg_fig2_manual_overlap_potential,
       filename = paste0(p_plots, "/fig2_overlap_potential", run_label, ".pdf"), 
    width = 3.5, height = 5, units = "in")

```


```{r man-fig2-recult}

# save plot
gg_fig2_manual_recult <-
  recult_length_distill_df %>% 
  bind_rows(
    tibble(length = 0, freq = 0, site = site_df$site, length_type = "recult"),
    tibble(length = 26, freq = 0, site = site_df$site, length_type = "recult")) %>%
  left_join(mean_recult_length_df %>% filter(recult_threshold == 1)) %>%
  mutate(height = freq / n_abn_periods,
         site = fct_reorder(site, order)) %>%
  
  # plot
  ggplot() + 
  geom_col(aes(x = length, y = height, group = site),
           fill = "gray70", color = "gray70") +
  geom_step(aes(x = length, y = height, group = site),
            direction = "mid", size = 0.5) +
  geom_hline(yintercept = 0) +
  geom_point(data = . %>% select(site, mean_duration, length_type) %>% unique(),
             mapping = aes(y = 0, x = mean_duration, color = length_type),
             inherit.aes = FALSE,
             size = 3, alpha = 0.85) +
  scale_alpha_manual(values=c(0.5))+
  facet_grid(vars(site), scales = "free_x",
             labeller = as_labeller(fancy_labels)
             ) + 
  theme_classic() + 
  scale_y_continuous(
    # breaks = c(0, 0.1, 0.2), labels = c("0", "0.1", "0.2")
    n.breaks = 3
                     ) + 
  labs(y = "Proportion of abandonment periods by duration", 
       x = "Recultivation duration (years)",
       color = "Site-level \nmean of:", linetype = NULL) +
  geom_vline(
    data = . %>% select(site, mean_duration) %>% unique() %>% 
      summarise(overall_mean = mean(mean_duration, na.rm = TRUE)),
    mapping = aes(xintercept = overall_mean,
                  linetype = "Mean across\nall sites"),
    color = hue_pal()(3)[1]) +
  
  scale_linetype_manual(values = c("Mean across\nall sites" = "dashed")) +
  scale_color_discrete(labels = c(recult = "All recultivation\nperiods")) +
  
  scale_x_continuous(breaks = c(5, 10, 15, 20, 25, 30), expand = c(0, 0)) +
  guides(color = guide_legend(order = 1, nrow = 1, 
                              override.aes = list(linetype = 0)), 
         linetype = guide_legend(order = 2, nrow = 1)) + 
  coord_cartesian(clip = "off") +
  theme(axis.line.x = element_blank(),
        axis.text.y = element_text(vjust = 0.35, size = 6),
        strip.text.y.right = element_text(size = 8, angle = 0, hjust = 0, vjust = 0),
        strip.background = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        plot.background = element_blank(),
        panel.spacing = unit(1, "mm"),
        panel.grid.major.x = element_line(),
        legend.spacing.x = unit(0.5, "mm"),
        legend.position = c(0.6, -0.14),
        legend.direction = "horizontal",
        legend.box = "horizontal",
        plot.margin = margin(5, 0, 33, 5),
        legend.background = element_blank(),
        legend.box.just = "left",
        legend.text.align = 0,
        legend.title.align = 1)

  
# save plot
ggsave(plot = gg_fig2_manual_recult,
       filename = paste0(p_plots, "/fig2_recult", run_label, ".pdf"), 
    width = 3.5, height = 5, units = "in")

```
# Area abandoned

```{r total-cropland-extent}
# ------------------------------------------------------ #
# First, create a raster showing pixels that were classified 
# as cropland at least once during the time series: "total_crop_mask"
# ------------------------------------------------------ #
# terraOptions(todisk = TRUE)
terraOptions()

# prepared input rasters (derived by Chris)
lcc


tic()
lcc_total_crop_mask <- lapply(1:11, function(i) {
  reclassified <- classify(lcc[[i]], rcl = tribble(
  ~"is", ~"becomes", 
  0, NA,
  1, NA,
  2, NA,
  3, 1,
  4, NA))
  
  lcc_crop_mask <- max(reclassified, na.rm = TRUE)
  rm(reclassified)
  
  # write new total_crop_mask raster to file:

  terra::writeRaster(lcc_crop_mask,
              filename = paste0(p_dat_derived, "total_crop_mask/", 
                                site_df$site[i], "_total_crop_mask_clean",
                                run_label, ".tif"),
              overwrite = TRUE
              )

  lcc_crop_mask
})
toc()
# 2033 seconds

names(lcc_total_crop_mask) <- site_df$site

```


This next chunk accomplishes the following:
1. Calculate area abandoned summary stats:
2. Calculate the area at each site that is abandoned at any point during the time series (>= 5 years)
3. Load the maximum extent of all cropland ever cultivated during time series
4. What percentage of the abandonment periods are recultivated?
5. Proportion of each site's area in non-vegetated land
6. Merge all together
7. Save the summary file

`area_summary_df` has 16 variables. All area values are given in hectares (ha) unless stated otherwise:
1. *site*
2. *total_site_area_ha_2017* - the total site area in hectares in 2017 
3. *cropland_area_1987* - the area in cropland in 1987 (ha)
4. *area_abn_ha_2017* - the area abandoned as of 2017 (ha)
5. *area_ever_abn_ha* - the area of those pixels that were abandoned at least once during the time series
6. *total_crop_extent_ha* - the area of those pixels that were classified as cropland at least once during the time series
7. *total_area_abn_remaining_2017* - duplicate of *area_abn_ha_2017,* the area abandoned as of 2017 (ha)
8. *total_initial_area_abn* - the sum of the initial area of each cohort of abandonment when it is first classified as "abandoned," i.e., at the 5 year mark. Note that this is cumulative, and because it counts those pixels that were abandoned more than once, it is therefore larger than *area_ever_abn_ha*.
9. *total_area_abn_recultivated_2017* - the area of abandoned land that was recultivated as of 2017 (cumulatively, i.e., *total_initial_area_abn* - *area_abn_ha_2017*)
10. *proportion_recultivated* - the proportion of all abandoned cropland (including multiple periods per pixel) that was recultivated by 2017 
11. *area_2017_as_prop_site* - area abandoned as of 2017 as a proportion of the total site area
12. *area_2017_as_prop_total_crop* - area abandoned as of 2017 as a proportion of the total crop extent
13. *area_2017_as_prop_crop87* - area abandoned as of 2017 as a proportion of cropland area in 1987
14. *area_ever_abn_as_prop_site* - area ever abandoned as a proportion of the total site area
15. *area_ever_abn_as_prop_total_crop* - area ever abandoned as a proportion of the total crop extent
16. *area_ever_abn_as_prop_crop87* - area ever abandoned as a proportion of cropland area in 1987

```{r area-abn-summary}

# 1. Calculate area abandoned summary stats:
# ----------------------------------------------------------- #
area_summary_df <- 
  area_dat %>% 
  filter(lc %in% c("Non-veg.", "Woody veg.", "Cropland", "Grassland"),
         year == 2017) %>%
  group_by(site) %>%
  summarise(total_site_area_ha_2017 = sum(area_ha)) %>% 
  left_join(., 
            left_join(filter(area_dat, lc == "Cropland", year == 1987) %>% 
                        select(site, cropland_area_1987 = area_ha), 
                      filter(area_dat, lc == "Abandoned (>=5)", year == 2017) %>% 
                        select(site, area_abn_ha_2017 = area_ha),
                      by = c("site")), 
            by = c("site"))


# ----------------------------------------------------------- #
# 2. Calculate the area at each site that is abandoned at any point during the time series (>= 5 years)
# ----------------------------------------------------------- #
# load max age rasters, see _util_files.R
max_age_t

tic()
area_ever_abn <- lapply(1:11, function(i) {
  max_age_mask <- max_age_t[[i]] %>%
    classify(., 
             rcl = tibble("is" = 0:31, 
                          becomes = if_else(is >= 5, 1, NA_real_)))
  
  area_ha <- cellSize(max_age_mask, unit = "ha", mask = TRUE) %>%
    global(., fun = "sum", na.rm = TRUE)
  
  tibble(site = site_df$site[i],
         area_ever_abn_ha = as.numeric(area_ha))
} 
) %>% bind_rows()
toc() # 75.81 sec

area_ever_abn
area_ever_abn %>% summarise(total_ever_abn_Mha = sum(area_ever_abn_ha)/1e6)


# ----------------------------------------------------------- #
# 3. Load the maximum extent of all cropland ever cultivated during time series
# ----------------------------------------------------------- #


lcc_total_crop_mask <- lapply(
  list.files(paste0(p_dat_derived, "total_crop_mask"), full.names = TRUE) %>%
    grep("clean", ., value = TRUE, invert = FALSE), 
  function(i) {
    rast(i)
    })

names(lcc_total_crop_mask) <- site_df$site

# calculate area
lcc_total_crop_mask_area_ha <- sapply(1:11, function(i) {
  size <- cellSize(lcc_total_crop_mask[[i]], unit = "ha", mask = TRUE)
  # product <- size * lcc_total_crop_mask[[i]]
  area_ha <- global(size, fun = "sum", na.rm = TRUE)
  
  as.numeric(area_ha)
  })


total_crop_mask_area <- tibble(site = site_df$site,
       total_crop_extent_ha = lcc_total_crop_mask_area_ha
       )

total_crop_mask_area


# ---------------------------------------------------------------------- #
# 4. What percentage of the abandonment periods are recultivated?
# ---------------------------------------------------------------------- #

persistence_dat %>% filter(site == "shaanxi", year == 2017)
# sum up the initial_area_abn to find the total area ever abandoned, then
# sum up the area_ha in 2017 (the last year of the time series). 
# Divide the two and subtract from 1, we have the proportion of initial area abandoned that 
# was recultivated during the time series. 

area_recultivated_df <- 
  persistence_dat %>% 
  filter(year == 2017) %>%
  group_by(site) %>%
  summarise(total_area_abn_remaining_2017 = sum(area_ha),
            total_initial_area_abn = sum(initial_area_abn),
            total_area_abn_recultivated_2017 = 
              total_initial_area_abn - 
              total_area_abn_remaining_2017) %>%
  mutate(proportion_recultivated = total_area_abn_recultivated_2017/total_initial_area_abn)


# ----------------------------------------------------------- #
# 5. Proportion of each site's area in non-vegetated land
# ----------------------------------------------------------- #
area_dat %>% 
  filter(lc %in% c("Non-veg.", "Woody veg.", "Cropland", "Grassland"), 
         # year == 2017
         ) %>%
  group_by(site) %>%
  summarise(total_site_area_ha = sum(area_ha)) %>% 
  left_join(area_dat, ., by = c("site")) %>% 
  mutate(prop_total_area = area_ha/total_site_area_ha) %>%
  group_by(site, lc) %>%
  # summarise(max_prop = max(prop_total_area)) %>%
  # arrange(max_prop) %>%
  
# plot
  # filter(., lc == "Non-veg.") %>%
  ggplot(data = .) +
    theme_classic() +
    labs(y = expression("Area  (10"^{6}*" ha)"), x = "Year", 
         title = "Area by Land Cover, with Abandonment",
         color = NULL) + 
    geom_line(mapping = aes(x = year, y = prop_total_area, #area_ha / (10^6), # convert to millions of ha
                            group = lc, color = lc),
              size = 1.5) + 
    theme(legend.position = "bottom",
          legend.title = element_text(size = 10)) +
    guides(color = guide_legend(nrow = 2, ncol = 3)) +
    
    scale_color_manual(values = plot_cols) + 
  facet_wrap(vars(site))

# ----------------------------------------------------------- #
# 6. Merge all together
# ----------------------------------------------------------- #

area_summary_df <- area_summary_df %>% 
  left_join(., area_ever_abn, by = c("site")) %>%
  left_join(., total_crop_mask_area, by = c("site")) %>%
  left_join(., area_recultivated_df, by = c("site"))


# calculate area as various proportions
names(area_summary_df)
area_summary_df <- area_summary_df %>%
  mutate(area_2017_as_prop_site = area_abn_ha_2017/total_site_area_ha_2017,
         area_2017_as_prop_total_crop = area_abn_ha_2017/total_crop_extent_ha,
         area_2017_as_prop_crop87 = area_abn_ha_2017/cropland_area_1987,
         area_ever_abn_as_prop_site = area_ever_abn_ha/total_site_area_ha_2017,
         area_ever_abn_as_prop_total_crop = area_ever_abn_ha/total_crop_extent_ha,
         area_ever_abn_as_prop_crop87 = area_ever_abn_ha/cropland_area_1987)

# ----------------------------------------------------------- #
# 7. save the summary file
# ----------------------------------------------------------- #
write_csv(area_summary_df, file = paste0(p_derived2, "area_summary_df", run_label, ".csv"))
# area_summary_df <- read_csv(file = paste0(p_derived2, "area_summary_df", run_label, ".csv"))

# area_summary_df_old_area <- read_csv(file = paste0(p_dat_derived, run_label, "_old_area", "/", "area_summary_df", run_label, ".csv"))
```

```{r total-area-panel-fig}

# Constructing a panel figure showing:
# 1. Area abandoned as of 2017 
# 2. Area abandoned at least once between 1987-2017
# 3. Area abandoned as of 2017 as a proportion of total cropland extent


# exploring these plots

# how much land is abandoned by the end of the time series?
area_dat %>% 
  filter(lc == "Abandoned (>=5)",
         year == 2017) %>%
  ggplot(mapping = aes(y = site, x = area_ha/(10^6))) + 
  geom_col()


# plot the area abandoned in 2017:
ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(area_abn_ha_2017)), 
                     x = area_abn_ha_2017/(10^6))) + 
  geom_col() +
  theme_classic() + 
  labs(x = "Area Abandoned (millions ha)", y = "Site")


# plot area abandoned as a proportion of the total site area:
ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(area_abn_ha_2017/total_site_area_ha_2017)),
                     x = area_abn_ha_2017/total_site_area_ha_2017)) + 
  geom_col() +
  theme_classic() + 
  labs(x = "Proportion of total site area abandoned", y = "Site")


# plot area abandoned as a proportion of the amount of cropland at the start of the time series:
ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(area_abn_ha_2017/cropland_area_1987)), 
                     x = area_abn_ha_2017/cropland_area_1987)) + 
  geom_col() +
  theme_classic() + 
  labs(x = "Proportion of cropland in 1987 that is abandoned", y = "Site")



# panel plot

area_a <-
  ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(area_2017_as_prop_total_crop)), 
                     x = area_abn_ha_2017/(10^3))) + 
  geom_col() +
  theme_classic() + 
  labs(#x = expression("Area abandoned ">="5 years \nas of 2017 (10"^{3}*" ha)"),
    x = "Area abandoned as of 2017\n(10\u00b3 ha)",
       y = "Site") + 
  scale_y_discrete(expand = c(0.02, 0),
                   labels = fancy_labels)

# plot area abandoned at any point between 1987-2017 (thousands ha):
area_b <-
  ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(area_2017_as_prop_total_crop)), 
                     x = area_ever_abn_ha/(10^3))) + 
  geom_col() +
  theme_classic() + 
  labs(x = "Area abandoned at any point \nbetween 1987-2017 (10\u00b3 ha)", y = "Site") + 
  theme(axis.text.y = element_blank(), axis.title.y = element_blank(),
        plot.margin = margin(3, 8, 3, 15))

# plot area ever abandoned as a proportion of the total cropland extent:
area_c <- ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(area_2017_as_prop_total_crop)),
                     x = area_2017_as_prop_total_crop)) + 
  geom_col() +
  theme_classic() + 
  labs(x = "Area abandoned as of 2017 \n as a proportion of \ntotal cropland extent", y = "Site") + 
  theme(axis.text.y = element_blank(), axis.title.y = element_blank(),
        plot.margin = margin(3, 6, 3, 15))

names(area_summary_df)


# save panel plot

ggsave(plot = plot_grid(area_a, area_b, area_c, 
                        nrow = 1, ncol = 3,
                        align = "h", axis = "l",
                        # hjust = 0,
                        rel_widths = c(1.5, 1, 1),
                        labels = "auto"),
       filename = paste0(p_plots, run_label, "/area_abn_panel", run_label, ".pdf"), 
    width = 8, height = 4, units = "in"
)

```


## Mean proportion of abandoned land that is recultivated

```{r prop-recultivated}
# see chunk: area-abn-summary
# note, area_recultivated_df has been joined to area_summary_df
area_recultivated_df %>% names()

area_recultivated_df %>% arrange(proportion_recultivated)

area_summary_df %>% names()

ggplot(data = area_summary_df,
       mapping = aes(y = fct_reorder(site, desc(proportion_recultivated)), 
                     x = proportion_recultivated)) + 
  geom_col() +
  theme_classic() + 
  labs(x = "Proportion of abandoned lands \nrecultivated by 2017", y = "Site") + 
  theme(#axis.text.y = element_blank(), axis.title.y = element_blank(),
        plot.margin = margin(3, 6, 3, 15))


# averaging across site to be consistent with summary_stats_all_sites
area_summary_df %>% select(site, proportion_recultivated) %>% arrange(proportion_recultivated)

area_summary_df %>% 
  summarise(mean_proportion = mean(proportion_recultivated))
```

```{r recultivation-by-threshold}
# How does the proportion that gets recultivated change based on the abandonment threshold?
# see chunk: area-abn-summary
area_recultivated_df

area_recult_threshold <- lapply(c(5, 7, 10), function(i) {
  persistence_dat %>% 
  left_join(.,
            persistence_dat %>% filter(age == i) %>% 
              select(site, year_abn, initial_area_abn_alt_thr = area_ha), 
            by = c("site", "year_abn")) %>%
  filter(year == 2017, age >= i) %>% # this restricts to only land abandoned for at least 5, 7, or 10 years
  group_by(site) %>%
  summarise(total_area_abn_remaining_2017 = sum(area_ha),
            total_initial_area_abn = sum(initial_area_abn_alt_thr),
            total_area_abn_recultivated_2017 = 
              total_initial_area_abn - 
              total_area_abn_remaining_2017) %>%
  mutate(proportion_recultivated = total_area_abn_recultivated_2017/total_initial_area_abn,
         abn_threshold = i) 
}) %>% 
  bind_rows() %>%
  left_join(., 
            filter(., abn_threshold == 5) %>% 
              arrange(proportion_recultivated) %>% 
              mutate(order = 1:n()) %>% select(site, order)
            )

# note: the sum of the "initial area abandoned" values is higher than the "area ever abandoned" because some pixels were abandoned multiple times (and recultivated multiple times), and therefore show up more than once in the initial area abandoned. 
area_recult_threshold %>% 
  group_by(abn_threshold) %>%
  # filter(abn_threshold == 5) %>%
  summarise(sum_abn_2017 = sum(total_area_abn_remaining_2017) / 10^6,
            sum_init_abn = sum(total_initial_area_abn) / 10^6) 

# data description:
names(area_recult_threshold)



round(area_summary_df %>% summarise(total_ever_abn_Mha = sum(area_ever_abn_ha)/1e6), digits = 2)


# ----------------------------------------------------------- #
# save the summary file
# ----------------------------------------------------------- #
write_csv(area_recult_threshold, file = paste0(p_derived2, "area_recult_threshold", run_label, ".csv"))
# area_recult_threshold <- read_csv(file = paste0(p_derived2, "area_recult_threshold", run_label, ".csv"))



# ----------------------------------------------------------- #
# save plot
# ----------------------------------------------------------- #

gg_recultivation <-
  ggplot(data = area_recult_threshold,
       mapping = aes(y = fct_reorder(site, desc(order)), 
                     x = proportion_recultivated,
                     group = as_factor(abn_threshold), fill = as_factor(abn_threshold))) + 
  geom_col(position = position_dodge(), width = 0.6) +
    
    scale_y_discrete(expand = c(0.02, 0),
                      labels = fancy_labels) +

  geom_vline(data = area_recult_threshold %>% 
               group_by(abn_threshold) %>%
               summarise(overall_mean = mean(proportion_recultivated, na.rm = TRUE)), 
             mapping = aes(xintercept = overall_mean, 
                           linetype = "Mean",
                           color = as_factor(abn_threshold)),
             show.legend = TRUE, #color = "black",
             size = 0.75
             ) + 
  scale_linetype_manual(values = c("Mean" = "dashed"), labels = "Mean\nacross\nsites",
                        name = NULL) +
    scale_fill_viridis(discrete = TRUE) + 
    scale_color_viridis(discrete = TRUE) + 
    
    # scale_fill_manual(values = viridis(n = 3)) + 
    # scale_color_manual(values = viridis(n = 3)) + 

  theme_classic() + 
  labs(x = "Proportion of abandoned lands recultivated by 2017", y = "Site",
       fill = "Abandonment\nthreshold") + 
  theme(#axis.text.y = element_blank(), axis.title.y = element_blank(),
        # plot.margin = margin(3, 6, 3, 15),
    legend.spacing.y = unit(1, "mm"),
        legend.position = c(0.87, 0.75),
        legend.background = element_blank(),
        # legend.box.background = element_rect(colour = "black")
        ) + 
    guides(fill = guide_legend(override.aes = list(linetype = 0)),
           color = "none")

ggsave(plot = gg_recultivation,
       filename = paste0(p_plots, run_label, "/recultivation_rates_by_threshold", run_label, ".pdf"), 
       width = 5.5, height = 4.5, units = "in")

```


## Proportion of abandoned land in each noncrop land cover class

What proportion of abandoned land is classified as woody vegetation vs. grassy vegetation? 
See "cluster/4_lc_of_abn.R"

```{r}
abn_lc_area_2017 <- read_csv(file = paste0(p_derived2, "abn_lc_area_2017", run_label, ".csv"))
abn_prop_lc_2017 <- read_csv(file = paste0(p_derived2, "abn_prop_lc_2017", run_label, ".csv"))
```


# Comparing methods for identifying cropland abandonment: annual trajectory vs. two-year method

```{r 2yr-abn-method}
# replicating commonly used "two-year" methods for identifying abandonment, by simply comparing land cover maps for two years, and identifying pixels classified as cropland in the first time step but not in the second time step.

# ---------------------------------------------------------------- #
# 1. load land cover rasters
# ---------------------------------------------------------------- #
# prepared input rasters (derived by Chris)
lcc # see _util_files.R


# ---------------------------------------------------------------- #
# 2-4 all together:
# ---------------------------------------------------------------- #
# run_label, "/", 
# ", run_label, "
abn_2yr_mask <- lapply(1:11, 
                  function(i) {
  # 2. Extract only 2017 & 1987
  lcc_87_17_tmp <- lcc[[i]] %>% 
    subset(., c("y1987", "y2017"),
           filename = paste0(p_dat_derived, "abn_2yr/", run_label, "/", 
                             site_df$site[i], "_87_17.tif"),
           overwrite=TRUE
           )
  
  # 3. reclassify the rasters, so that it's just cropland in 1987 (0-1), and just cropland in 2017 (0-1)
  lcc_87_17_tmp <- classify(lcc_87_17_tmp, 
           rcl = tribble(
             ~"is", ~"becomes", 
             NA, NA,
             0, NA,
             1, NA,
             2, 0,
             3, 1,
             4, 0),
           filename = paste0(p_dat_derived, "abn_2yr/", run_label, "/", site_df$site[i], "_87_17_rcl.tif"),
           overwrite=TRUE)
  
  # 4. Subtract the two (1987 - 2017), so that pixels that were cropland (1) in 1987, but not in 2017 (0) (i.e. are "abandoned" by this definition) have a value of 1 - 0 = 1.
  # Pixels that are cropland in both will be 1 - 1 = 0 (as will pixels that are non cropland in both 0 - 0 = 0)
  # Pixels that are cropland in 2017 (1) but not in 1987 (0) will have a value of 0 - 1 = -1
  subtract <- lcc_87_17_tmp$y1987 - lcc_87_17_tmp$y2017
  
  terra::writeRaster(subtract, 
                     filename = paste0(p_dat_derived, "abn_2yr/", run_label, "/", 
                                       site_df$site[i], "_87_17_sub.tif"),
                     overwrite=TRUE)
  
  # 5. Reclassify into 0-1 abandonment mask
  mask <- classify(subtract, rcl = 
                     tribble(
                       ~"is", ~"becomes", 
                       -1, NA_real_, # non-cropland in 1987 (0) but cropland in 2017 (1), 0-1 = -1
                       0, NA_real_, # cropland (1) in both 1987 & 2017 or non-cropland (0) in both, 1-1 = 0-0 = 0
                       1, 1), # cropland in 1987 (1) but not in 2017 (0), 1 - 0 = 1
                   filename = paste0(p_dat_derived, "abn_2yr/", run_label, "/", 
                                     site_df$site[i], "_abn_2yr_mask.tif"),
                   overwrite=TRUE
                   )
  mask
  }
  )

names(abn_2yr_mask) <- site_df$site
plot(abn_2yr_mask$shaanxi)

# ---------------------------------------------------------------- #
# 6. calculate the area "abandoned" at each site
# ---------------------------------------------------------------- #

abn_2yr_mask <- lapply(1:11, function(i) {
  rast(x = paste0(p_dat_derived, "abn_2yr/", run_label, "/",
                  site_df$site[i], "_abn_2yr_mask.tif"))
  })

names(abn_2yr_mask) <- site_df$site # (note that the layer name still have y1987, since it was calculated as y1987 - y2017)


abn_2yr_area_ha <- sapply(1:11, 
                          function(i) {
  size <- cellSize(abn_2yr_mask[[i]], unit = "ha", mask = TRUE)
  # product <- abn_2yr_mask[[i]] * size
  area_ha <- global(size, fun = "sum", na.rm = TRUE)
  as.numeric(area_ha)
  })
```


```{r abn_cc_mask}
# ---------------------------------------------------------------- #
# Select out the areas identified as abandoned as of 2017
# using the annual trajectories and full time series ("abn_cc_2017")
# ---------------------------------------------------------------- #

# load abandonment age rasters
# abandonment age maps (produced by Chris)
age_t # see _util_files.R

# area in 2017
freq(age_t$shaanxi$y2017)
plot(age_t$shaanxi$y1988)


# make abn mask
abn_cc_2017_mask <- lapply(1:11, function(i) {
  classify(age_t[[i]]$y2017, 
           # takes way longer to change values to 0 than to NA_real_
           rcl = tibble("is" = 0:30, becomes = if_else(is < 5, NA_real_, 1)), 
           filename = paste0(p_dat_derived, "abn_2yr/", run_label, "/",
                             site_df$site[i], "_abn_cc_2017_mask.tif"),
                   overwrite=TRUE)
           })

# reload
abn_cc_2017_mask <- lapply(1:11, function(i) {
  rast(x = paste0(p_dat_derived, "abn_2yr/", run_label, "/",
                  site_df$site[i], "_abn_cc_2017_mask.tif"))
  })

names(abn_cc_2017_mask) <- site_df$site
# plot(abn_cc_2017_mask$shaanxi)

# calculate area abandoned in 2017
abn_cc_2017_area_ha <- sapply(1:11, 
                         function(i) {
  size <- cellSize(abn_cc_2017_mask[[i]], unit = "ha", mask = TRUE)
  # product <- size * abn_cc_mask[[i]]
  area_ha <- global(size, fun = "sum", na.rm = TRUE)
  as.numeric(area_ha)
  })

# ---------------------------------------------------------------- #
# 6. Compare to area abandoned ever, and the area abandoned in 2017
# ---------------------------------------------------------------- #

area_summary_df %>% 
  select(site, area_abn_ha_2017) %>% 
  mutate(area_ha_2017_terra = abn_cc_2017_area_ha#,
         #area_ha_2017_2yr = abn_area_2017_2yr
         ) %>%
  mutate(percent_diff = 100*(area_ha_2017_terra - area_abn_ha_2017)/area_abn_ha_2017)
```


```{r jaccard-similarity}
# ---------------------------------------------------------------- #
# 7. Calculate Jaccard Similarity (a metric of overlap between two sets)
# for the two abandonment methods
# ---------------------------------------------------------------- #

# load two abandonment masks

abn_2yr_mask <- lapply(1:11, function(i) {rast(paste0(p_dat_derived, "abn_2yr/", run_label, "/",
                                                      site_df$site[i], "_abn_2yr_mask.tif"))})
names(abn_2yr_mask) <- site_df$site

abn_cc_2017_mask <- lapply(1:11, function(i) {rast(paste0(p_dat_derived, "abn_2yr/", run_label, "/",
                                                          site_df$site[i], "_abn_cc_2017_mask.tif"))})
names(abn_cc_2017_mask) <- site_df$site


# Add the two abandonment maps to each other, and calculate the Jaccard Similarity
# J = intersection of A and B / union of A and B
# sum of cells that = 2 / sum of cells > 0

# i <- 9 # for testing
jaccard_similarity <- sapply(1:11, function(i) {
  # both masks are 1 for abandoned (in 2017) and NA for not abandoned.
  # adding to an NA value yields NA, so this is now the intersection, 
  # where cells where both masks show abandonment have a value of 
  # 1 + 1 = 2
  
  # calculate intersection, if necessary:
  intersection <- abn_2yr_mask[[i]] + abn_cc_2017_mask[[i]]
  terra::writeRaster(intersection,
                     filename = paste0(p_dat_derived, "abn_2yr/", run_label, "/",
                                       site_df$site[i], "_abn_2yr_cc_intersection.tif"),
                     overwrite=TRUE)

  # # load, if intersection has already been created:
  # intersection <- rast(paste0(p_dat_derived, "abn_2yr/", run_label, "/",
  #                           site_df$site[i], "_abn_2yr_cc_intersection.tif"))
  
  intersection_area <- intersection %>% 
    cellSize(., unit = "ha", mask = TRUE) %>% 
    global(., fun = "sum", na.rm = TRUE) %>%
    as.numeric()
  
  a_area <- abn_2yr_mask[[i]] %>% 
    cellSize(., unit = "ha", mask = TRUE) %>% 
    global(., fun = "sum", na.rm = TRUE) %>%
    as.numeric()
  
  b_area <- abn_cc_2017_mask[[i]] %>% 
    cellSize(., unit = "ha", mask = TRUE) %>% 
    global(., fun = "sum", na.rm = TRUE) %>%
    as.numeric()
  
  # add area calculation for both a and b, so you can do j =  int / (a + b - int)
  jaccard <- intersection_area / (a_area + b_area - intersection_area)
  
  as.numeric(jaccard)
})

jaccard_similarity
```


```{r check-jaccard-with-shaanxi}
# double check with shaanxi -- checks out.


# reclassify both to 0 and 1.
freq(abn_2yr_mask$shaanxi)
freq(abn_cc_2017_mask$shaanxi)

t_a <- classify(abn_cc_2017_mask[[9]], 
           # takes way longer to change values to 0 than to NA_real_
           rcl = tribble(
                       ~"is", ~"becomes", 
                       NA_real_, 0,
                       1, 1), 
           filename = paste0(p_dat_derived, "abn_2yr",  run_label, "/test/", 
                             site_df$site[9], "_abn_cc_2017_mask_01.tif"),
                   overwrite=TRUE)

t_b <- classify(abn_2yr_mask[[9]], 
           # takes way longer to change values to 0 than to NA_real_
           rcl = tribble(
                       ~"is", ~"becomes", 
                       NA_real_, 0,
                       1, 1), 
           filename = paste0(p_dat_derived, "abn_2yr",  run_label, "/test/", 
                             site_df$site[9], "_abn_2yr_mask_01.tif"),
                   overwrite=TRUE)

plot(t_b)
# add them
t_add <- t_a + t_b
plot(t_add)

t_union <- classify(t_add, 
           # takes way longer to change values to 0 than to NA_real_
           rcl = tribble(
                       ~"is", ~"becomes", 
                       0, NA_real_,
                       1, 1,
                       2, 1), 
           filename = paste0(p_dat_derived, "abn_2yr",  run_label, "/test/", 
                             site_df$site[9], "_t_union.tif"),
                   overwrite=TRUE)

t_intersection <- classify(t_add,
           # takes way longer to change values to 0 than to NA_real_
           rcl = tribble(
                       ~"is", ~"becomes", 
                       0, NA_real_,
                       1, NA_real_,
                       2, 1), 
           filename = paste0(p_dat_derived, "abn_2yr",  run_label, "/test/", 
                             site_df$site[9], "_t_intersection.tif"),
                   overwrite=TRUE)

t_union_area <- t_union %>% 
    cellSize(., unit = "ha", mask = TRUE) %>% 
    global(., fun = "sum", na.rm = TRUE) %>%
    as.numeric()

t_intersection_area <- t_intersection %>% 
    cellSize(., unit = "ha", mask = TRUE) %>% 
    global(., fun = "sum", na.rm = TRUE) %>%
    as.numeric()

t_intersection_area / t_union_area


# reclassify to 0 and 1

abn_2yr_overestimation
```


```{r prop-two-yr-abn-less-than-5-yrs}
# Of the pixels that are identified as abandonment by the two-year method (2017 - 1987), what proportion are "young," as in, less than 5 years old (and therefore not meeting the five-year abandonment threshold)?

# The following code produces:
abn_2yr_ages_df
total_area_by_site
abn_2yr_ages_summary_df

# load 2 yr mask (2017) rasters (produced up above):
abn_2yr_mask <- lapply(1:11, function(i) {
  rast(x = paste0(p_dat_derived, "abn_2yr/",  run_label, "/", site_df$site[i], "_abn_2yr_mask.tif"))
  })

names(abn_2yr_mask) <- site_df$site
# (note that the layer name still have y1987, since it was calculated as y1987 - y2017)


# load abandonment age rasters (produced by Chris)
age_t

# ---------------------------------------------------------------- #
# calculate the age of the "abandonment" identified in the 2-yr-mask (i.e. those pixels that were cropland in 1987 and non-crop in 2017) for all sites.
# ---------------------------------------------------------------- #
# i <- 9 # for testing

abn_2yr_ages_df <- lapply(1:11, FUN = function(i){
  # temp SpatRaster
  
  tmp_b <- c(
    abn_2yr_mask[[i]], # abandonment identified using the two-year method (2017-1987)
    age_t[[i]]$y2017, # abandonment age in 2017, calculated using annual time series, five year abandonment definition
    terra::cellSize(abn_2yr_mask[[i]], unit = "ha", mask = FALSE) # area (ha), calculated as SpatRaster, then converted to Raster*
               )
  names(tmp_b) <- c("twoyr_2017", "age_2017", "area_ha")

  # convert to data.table
  
  tmp_dt <- spatraster_to_dt(tmp_b)
  
  # calculate the sum of area, pixels, etc.
  twoyr_age_df <- tmp_dt[twoyr_2017 > 0, 
                       .(count = .N, sum_area_ha = sum(area_ha)), 
                       by = age_2017
                       ][order(age_2017)] %>%
    as_tibble() %>% mutate(site = site_df$site[i])
  
  twoyr_age_df
}) %>% bind_rows()

# write to file:
write_csv(abn_2yr_ages_df, 
          file = paste0(p_derived2, "abn_2yr_ages_df", run_label, ".csv"))


abn_2yr_ages_df <- read_csv(file = paste0(p_derived2, "abn_2yr_ages_df", run_label, ".csv"))

total_area_by_site <- 
  abn_2yr_ages_df %>%
  group_by(site) %>%
  # filter(age_2017 > 0) %>%
  summarise(total_pixels = sum(count),
            total_area_ha = sum(sum_area_ha)) #%>% as.numeric()

abn_2yr_ages_summary_df <- 
  abn_2yr_ages_df %>%
  left_join(total_area_by_site, by = "site") %>%
  mutate(percent_px = 100 * count/total_pixels,
         percent_area = 100 * sum_area_ha/total_area_ha) %>%
  filter(age_2017 < 5) %>%
  group_by(site) %>%
  summarise(sum_p_px = sum(percent_px),
            percent_2yr_abn_area_less_than_5yrs_old = sum(percent_area)
  )










# ---------------------------------------------------------------- #
# testing code:
# ---------------------------------------------------------------- #
s_2yr_age <- terra::mask(x = age_t$shaanxi$y2017, mask = abn_2yr_mask$shaanxi, maskvalues = 1, inverse = TRUE)


par(mfrow = c(2, 1))
plot(is.na(age_t$shaanxi$y2017), main = "NA values in age raster")
plot(is.na(abn_2yr_mask$shaanxi), main = "NA values in 2 year mask")

plot(abn_2yr_mask$shaanxi, main = "2 year method mask")
plot(age_t$shaanxi$y2017 >= 0, main = "pixels with age value >= 0")

in_2yr <- abn_2yr_mask$shaanxi %>% 
  classify(., 
           # takes way longer to change values to 0 than to NA_real_
           rcl = tribble(
                       ~"is", ~"becomes", 
                       NA_real_, 0))

in_age <- age_t$shaanxi$y2017 >= 0
in_age <- in_age %>% 
  classify(., 
           # takes way longer to change values to 0 than to NA_real_
           rcl = tribble(
                       ~"is", ~"becomes", 
                       NA_real_, 0))
plot(in_2yr - in_age)

tdf <- freq(in_2yr - in_age) %>% as_tibble()
tdf

plot(abn_2yr_mask$shaanxi - s_2yr_age2 >= 0)


freq(abn_2yr_mask$shaanxi)
freq(s_2yr_age)
freq(s_2yr_age>0)
(5079796 - 4544511) / 5079796 * 100 # 10%!


# About 10% of the pixels in the two-year abandonment map have NA age values.
# Two year method: cropland in 1987, non-cropland in 2017.
# Age calculation: cropland at any point, then transition to non-crop after. Starts counting age in the first year, even though our definition of abandonment means that we only include abandonment >= 5 years as "abandoned."

# Why might some of these pixels have NA values? Answers:
# - The NA values in the age are a mix of non-abandonment (always non-veg or always crop), or pixels that are non-veg at ANY point during the time series.
# - As a result, there could be pixels that transition between cropland, non-veg, and non-cropland, which would show up as "abandoned" in the two-year abandonment map, but would be excluded from the age layer.
# - The 2-year method also doesn't incorporate some of the cleaning steps we took, like our 5- and 8-year moving windows.
# - Together, these factors could result in a substantial number of pixels that are excluded in the age layer, but that aren't in the 2-year map. 
# -One idea for future work would be to apply this 5- and 8-year moving window filters to all land cover classes, not just crop and non-crop pixels. Specifically, we could potentially recover some pixels that include only one year in non-veg, in the middle of a longer string of a different land cover code. For example, 33133 (non-veg), 22122 (grassland), or 44144 (woody)


terra::animate(lcc$shaanxi, 
     levels = plot_cols1$name,
     col = plot_cols1$color,
     type = "classes",
     maxcell = 500000,
     n = 1)


plot(abn_2yr_mask$shaanxi - age_t$shaanxi$y2017)

```


```{r 2yr-overestimation}
abn_2yr_overestimation <- 
  data.frame(site = site_df$site,
             area_ha_cc = abn_cc_2017_area_ha,
             area_2yr = abn_2yr_area_ha
             ) %>%
  mutate(percent_diff = 100 * (area_2yr - area_ha_cc) / area_ha_cc,
         jaccard = jaccard_similarity)

abn_2yr_overestimation <- abn_2yr_overestimation %>% 
  left_join(select(abn_2yr_ages_summary_df, -sum_p_px), by = "site")



# ----------------------------------------------------------- #
# save the summary file
# ----------------------------------------------------------- #
write_csv(abn_2yr_overestimation, file = paste0(p_derived2, "abn_2yr_overestimation", run_label, ".csv"))
# abn_2yr_overestimation <- read_csv(file = paste0(p_derived2, "abn_2yr_overestimation", run_label, ".csv"))

# ----------------------------------------------------------- #
# Plot the overestimation
# ----------------------------------------------------------- #
abn_2yr_overestimation <- 
  read_csv(file = paste0(p_derived2, 
                         "abn_2yr_overestimation", run_label, ".csv"))



# install.packages("ggforce")
library(ggforce)


ggplot(data = abn_2yr_overestimation #%>% filter(site != "mato_grosso")
       , 
       mapping = aes(x = fct_reorder(site, desc(percent_diff)), y = percent_diff)) + 
  geom_col() +
  # coord_flip() +
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 330, vjust = 1, hjust = 0)) +
  scale_x_discrete(labels = c("belarus" = "Vitebsk, Belarus /\n Smolensk, Russia",
                              "bosnia_herzegovina" = "Bosnia & Herzegovina",
                              "chongqing" = "Chongqing, China",
                              "goias" = "Goiás, Brazil",
                              "iraq" = "Iraq",
                              "mato_grosso" = "Mato Grosso, Brazil",
                              "nebraska" = "Nebraska/Wyoming, USA",
                              "orenburg" = "Orenburg, Russia /\n Uralsk, Kazakhstan",
                              "shaanxi" = "Shaanxi/Shanxi, China",
                              "volgograd" = "Volgograd, Russia",
                              "wisconsin" = "Wisconsin, USA")) #+ 
  # facet_zoom(ylim = c(-35, 25))




# ------------------------------------------------------------------------ #
# Produce table
# ------------------------------------------------------------------------ #
install.packages("gt")
library(gt)

abn_2yr_overestimation <- 
  read_csv(file = paste0(p_derived2, 
                         "abn_2yr_overestimation", run_label, ".csv")) %>%  
  arrange(percent_diff) %>% 
  mutate(
    # area_ha_cc = area_ha_cc %>% round(digits = 0) %>% prettyNum(big.mark = ","),
         # area_2yr = area_2yr %>% round(digits = 0) %>% prettyNum(big.mark = ","),
         # percent_diff = percent_diff %>% round(digits = 2) %>% paste0(., "%"),
         # jaccard = jaccard %>% round(digits = 2)
         ) %>% 
  left_join(., site_df %>% select(site, description), by = "site")


abn_2yr_overestimation %>% 
  select(description, area_ha_cc, area_2yr, percent_diff, jaccard) %>%

  knitr::kable(., 
             col.names = c("Site", "Area (annual)", "Area (2017-1987)", 
                           "Percent Difference", "Jaccard Similarity"),
             align = "lrrrr", format = "simple",
             caption = "Cropland abandonment (in Mha) as of 2017 as identified using a) our annual time series and a five-year abandonment definition and b) the difference between only two years.") 



gt_2yr <- abn_2yr_overestimation %>%
  select(description, area_ha_cc:jaccard) %>%
  gt() %>%
  tab_header(title = "Table 1. Cropland abandonment (in Mha) as of 2017 as identified using a) our annual time series and a five-year abandonment definition and b) the difference between only two years.") %>%
  tab_options(heading.align = "left") %>%
  cols_label(
    description = "Site",
    area_ha_cc = "Area (annual, as of 2017)",
    area_2yr = "Area (2017-1987)",
    percent_diff = "Percent Difference",
    jaccard = "Jaccard Similarity") %>%
  cols_align(align = "left", columns = description) %>%
  cols_align(align = "right", columns = c("area_ha_cc", "area_2yr", "percent_diff", "jaccard")) %>%
  fmt_percent(columns = percent_diff, decimals = 2, scale_values = FALSE) %>%
  fmt_number(., columns = c("area_ha_cc", "area_2yr"), use_seps = TRUE, sep_mark = ",", decimals = 0) %>%
  fmt_number(., columns = c("jaccard"), use_seps = TRUE, sep_mark = ",", decimals = 2)

gt_2yr

gtsave(gt_2yr, filename = paste0(p_plots, run_label, "/", 
                         "table_1_abn_2yr_overestimation.png"))

# note, can also use kableExtra::kable_as_image() to save the table



```

# Calculating cluster times

```{r seff}
# extract run codes:
list.files("scripts/cluster/slurm_out") %>% grep("*potential*", ., value = TRUE) %>%
  str_extract(., "[0-9]{6,10}_[0-9]{1,2}") %>%
  paste0("seff ", .) %>%
  paste(., collapse = "; ") %>%
  cat()

1202/60 # 20 minutes for shaanxi


cat(
  paste("seff", paste0("39004952_", 1:27), collapse = "; "),
  "> text2.txt"
    )
# Use the above to make seff calls to get the full times and memory usage for the full array run
# then copy and paste this output into a 

paste("seff", paste0("39537809_", c(1, 7, 8, 10)), collapse = "; ")


txt <- read_csv("scripts/ref/seff_out_5_potential_age.csv")
txt <- read_csv("scripts/ref/seff_out_run_1_2.2.csv")

txt %>% print(n = 15)


seff <- tibble(
  # core_index = 1:11,
  job_id = str_subset(txt$out, "^Job ID:"),
  array_id = str_subset(txt$out, "Array Job ID:"),
  time = str_subset(txt$out, "CPU Utilized"), 
  mem_used = str_subset(txt$out, "Memory Utilized"),
  mem_given = str_subset(txt$out, "Memory Efficiency"),
  state = str_subset(txt$out, "State: ")
  ) %>%
  mutate(
    job_id = str_extract(job_id, "[0-9]{6,10}") %>% as.numeric(),
    array_id = str_extract(array_id, "[0-9]{6,10}_[0-9]{1,2}"),
    array = str_extract(array_id, "_[0-9]{1,2}") %>% gsub("_", "", .) %>% as.numeric(),
    time = str_extract(time, "[0-9][0-9]:[0-9][0-9]:[0-9][0-9]"),
    mem_used = str_extract(mem_used, "[0-9]{2,3}.[0-9][0-9]..."),# %>% as.numeric(),
    mem_used_p = str_extract(mem_given, "[0-9]{1,2}.[0-9][0-9]%"),
    mem_given = str_extract(mem_given, "[0-9]{2,3}.[0-9][0-9] GB") %>% 
      gsub(" GB", "", .) %>% as.numeric()
    )

seff %>% arrange(array) %>% print(n = 30)
seff %>% filter(grepl("COMPLETED", state)) %>% arrange(array) %>% print(n = 30)



lcc

```
