==> scripts/4_figures.Rmd <==
---
title: "Figures"
output: html_notebook
editor_options: 
  chunk_output_type: console
---

# Exploratory data analysis and visualization

This document includes code for exploratory data analysis and the production of various figures for the Crawford, et al. See: https://github.com/chriscra/abandonment_trajectories. 

Not all of the code is necessary; this script is included largely for exploratory purposes.
Code chunks are meant to be walked through sequentially.

```{r initialize}
source("/Users/christophercrawford/work/projects/abandonment_trajectories/scripts/0_start.R")
```

```{r clean-up}
os <- get_sizes(ls())

os %>% 
  summarise(total_env_size = sum(size)) %>% 
  mutate(gb = total_env_size / 1024^3)

os %>% print(n = 30)

rm(list = os$object[c(1:3)])
# rm(list = os$object[c(3:18)])
get(os$object[c(30)])


# terra and raster temp files
terra::tmpFiles(current=TRUE, orphan=FALSE, old=FALSE, 
                remove=TRUE)

terra::tmpFiles(current=TRUE, orphan=TRUE, old=TRUE)

terra::tmpFiles(current=TRUE, orphan=TRUE, old=TRUE, 
                remove=TRUE)

raster::rasterTmpFile()
raster::showTmpFiles()
raster::removeTmpFiles()

warnings()

```

```{r load-data}
run_label # check. See "_util_main.R"


# for use on cluster:
# p_dat <- "/scratch/network/clc6/abandonment_trajectories/data_derived/"
# p_dat_derived <- "/scratch/network/clc6/abandonment_trajectories/data_derived/"
# p_output <- paste0("/scratch/network/clc6/abandonment_trajectories/output/")
list.files(p_dat_derived)
list.files(p_dat)

site_df <- read.csv(file = paste0(p_dat_derived, "site_df.csv"))

site_df_w_size <- read_csv(file = paste0(p_dat_derived, "site_df_w_size.csv"))


# -------------------------- load summarized data.frames --------------------------- #
# load data, produced from the following functions:
# cc_calc_persistence() # which is called within
# cc_calc_persistence_all() # which is called within
# cc_summarize_abn_dts(), which is the meta function called
# within the cluster script, "2_analyze_abn.R."

# load combined persistence, area, and turnover datasets (see "1_summary_stats.Rmd"):

area_dat <- read_csv(file = paste0(p_dat_derived, "area_dat", run_label, ".csv"))
turnover_dat <- read_csv(file = paste0(p_dat_derived, "turnover_dat", run_label, ".csv"))

# ------------------------------------------ #
# when reloading, must restructure factors:
persistence_dat <- read_csv(file = paste0(p_dat_derived, "persistence_dat", run_label, ".csv")) %>%
  mutate(site = as.factor(site),
         bins = as.factor(bins),
         year_abn_bins = as.factor(year_abn_bins),
         cohort = as.factor(cohort))

# when is the first year abandonment can take place?
persistence_dat %>% filter(area_ha > 0) %>% group_by(site) %>% summarise(first_year = min(year_abn))


```


```{r load-rasters}
# Land use class codes:
#       1. Non-vegetated area (e.g. water, urban, barren land)
#       2. Woody vegetation
#       3. Cropland 
#       4. Herbaceous land (e.g. grassland)

# -------------------------- development rasters --------------------------- #
# small testers:
bs <- brick(paste0(p_dat, "Abandonment/belarus_small.tif"))
bt <- brick(paste0(p_dat_derived, "belarus_subset.tif"))
names(bs) <- paste0("y", 1987:2017)
names(bt) <- paste0("y", 1987:2017)

# raw rasters
s <- brick(paste0(p_dat_derived, "input_rasters/shaanxi.tif"))
b <- brick(paste0(p_dat_derived, "input_rasters/belarus.tif")) # merged version

# age
s_age_r <- brick(paste0(p_dat_derived, "shaanxi_age.tif"))
b_age_r <- brick(paste0(p_dat_derived, "belarus_age.tif"))

# max_length
s_max_length_r <- brick(paste0(p_dat_derived, "shaanxi_max_length.tif"))
b_max_length_r <- brick(paste0(p_dat_derived, "belarus_max_length.tif"))

# update year names 1987 - 2017
names(s) <- paste0("y", 1987:2017)
names(b) <- paste0("y", 1987:2017)
names(s_age_r) <- paste0("y", 1987:2017)
names(b_age_r) <- paste0("y", 1987:2017)


# ------------------------------- load land cover maps --------------------------------------- #
# prepared input rasters (derived by Chris)
site_input_raster_files <- list.files(paste0(p_dat_derived, "input_rasters"), full.names = TRUE) %>%
  grep(".tif", ., value = TRUE) #%>% grep("age", ., value = TRUE, invert = TRUE)

lc <- lapply(seq_along(site_input_raster_files), function(i) {
  # raster::brick(site_input_raster_files[i]) # as rasters
  terra::rast(site_input_raster_files[i]) # as SpatRasters
  })
names(lc) <- site_df$site

# rename raster layers:
for (i in 1:11) {
  if (names(lc[i]) == "nebraska") {
    names(lc[[i]]) <- paste0("y", 1986:2018)
  } else {
    if (names(lc[i]) == "wisconsin") {
      names(lc[[i]]) <- paste0("y", 1987:2018)
    } else {
      # everything else, just 1987:2017
      names(lc[[i]]) <- paste0("y", 1987:2017)
    }}}



# ----------------------------- load abandonment age rasters ---------------------------- #
# abandonment age maps (produced by Chris)
age_files <- list.files(paste0(p_dat_derived, "age_rasters/", run_label), 
                        full.names = TRUE) %>%
  grep(".tif", ., value = TRUE) #%>% grep("age", ., value = TRUE, invert = FALSE)


age_r <- lapply(seq_along(age_files), function(i) {raster::brick(age_files[i])})
names(age_r) <- site_df$site
for (i in seq_along(age_r)) {names(age_r[[i]]) <- paste0("y", 1987:2017)} # remember: these are just 1987:2017


age_t <- lapply(seq_along(age_files), function(i) {terra::rast(age_files[i])})
names(age_t) <- site_df$site
for (i in seq_along(age_t)) {names(age_t[[i]]) <- paste0("y", 1987:2017)} # remember: these are just 1987:2017


# load max age rasters
max_age_files <- list.files(paste0(p_dat_derived, "max_age/", run_label), 
                        full.names = TRUE) %>%
  grep(".tif", ., value = TRUE)

max_age_r <- lapply(seq_along(max_age_files), function(i) {
  brick(max_age_files[i])
  })
names(max_age_r) <- site_df$site

max_age_t <- lapply(seq_along(max_age_files), function(i) {
  rast(max_age_files[i])
  })
names(max_age_t) <- site_df$site
for (i in seq_along(max_age_t)) {names(max_age_t[[i]]) <- "max_age"} # remember: these are just 1987:2017



# year of first abandonment maps (from He)
yoa_files <- list.files(paste0(p_dat, "Abandonment/year_of_abandonment/"))
```


```{r load-data.tables}
# -------------------------- data.tables --------------------------- #
b_age <- fread(input = paste0(p_dat_derived, "belarus_age.csv"))
names(b_age)
s_age <- fread(input = paste0(p_dat_derived, "shaanxi_age.csv"))


# original data
s_dt <- fread(input = paste0(p_dat_derived, "shaanxi.csv"))
names(s_dt) <- gsub(pattern = "andcover", replacement = "y", names(s_dt))


# input_data.tables folder contains the input land cover maps, as csv files

b_dt <- fread(input = paste0(p_dat_derived, "input_data.tables/belarus.csv")) # caution - huge file! 8.4 GB at least. 
b_dt <- fread(input = paste0(p_dat_derived, "input_data.tables/belarus.csv"),
              select = 3,
              nrows = 200) # caution - huge file! 8.4 GB at least. 
nrow(b_dt)

# length and data.tables for each site can be found: "data_derived/length"
```






# Exploration and Development


```{r simple-plots}

# -------------------------- plot raw data --------------------------- #
plot(b$y1987, main = "Belarus 1987", breaks = c(0, plot_cols$breaks), col = plot_cols$color)
legend("bottomleft", cex = 0.6, inset = 0,
       legend = plot_cols$name, 
       fill = plot_cols$color)

plot(b$y2017, main = "Belarus 2017", breaks = c(0, plot_cols$breaks), col = plot_cols$color)
legend("bottomleft", cex = 0.6, inset = 0,
       legend = plot_cols$name, 
       fill = plot_cols$color)


plot(bs$y1987, main = "Belarus small 1987", breaks = c(0, 1:4), col = plot_cols)
legend("bottomleft", cex = 0.6, inset = 0,
       legend = names(plot_cols)[1:4], fill = plot_cols[1:4])

plot(bs$y1987, main = "Belarus small 1987", breaks = c(0, plot_cols1$breaks), col = plot_cols1$color)
legend("bottomleft", cex = 0.6, inset = 0,
       legend = plot_cols1$name, 
       fill = plot_cols1$color)

# ---------------------- plot abandonment age ---------------------- #
plot(b_age_r$y2017, main = "Belarus Age of Abandonment 2017")
plot(b_age_r[[28:31]])



# -------------------------- animate --------------------------- #
# --------------
# https://www.rdocumentation.org/packages/raster/versions/3.1-5/topics/animate

animate(bt, pause = 0.5, zlim = c(1, 4), maxpixels=5000, n=1,
        breaks = c(0, plot_cols$breaks), col = plot_cols$color)

```



```{r animations}
site_index <- 9

lc_plot_cols
terra::animate(lcc$shaanxi, 
     levels = lc_plot_cols$name, # see "_util/_util_misc.R"
     col = lc_plot_cols$color,
     type = "classes",
     maxcell = 500000,
     colNA = "gray95",
     n = 1)

# ------------------------------------------- #
# save as gif: age of abandonment
# ------------------------------------------- #
lapply(1:11, function(site_index) {
  file_out_tmp <- paste0(p_plots, "animations/", "abn_age",
                       site_df$label[site_index], ".gif")
  saveGIF(
    terra::animate(age_t[[site_index]][[c(2,2:31)]], 
                   # levels = rcl_df$labels %>% unique() %>% .[-1],
                   main = paste0(inset_map_titles[site_index], ", Abandonment Age, ", 1987:2017),
                   type = "continuous",
                   maxcell = 500000,
                   range = c(5, 30),
                   col = viridis(n = 25, direction = -1),
                   colNA = "gray95",
                   # legend = FALSE,
                   # plg=list(title.cex=1.25),
                   mar = c(mval+0.5, mval, mval + 0.2, mval+1.5),
                   n = 1),
    movie.name = file_out_tmp,
    # ani.height = 350, ani.width = 350
)
  
  # Reload:
  tmp_gif <- image_animate(image_read(file_out_tmp), fps = 5)
  image_write(tmp_gif, file_out_tmp)
})



# ------------------------------------------- #
# now for land cover:
# ------------------------------------------- #
lapply(3, function(site_index) {
  file_out_tmp <- paste0(p_plots, "animations/", "land_cover",
                       site_df$label[site_index], "3.gif")
  saveGIF(
    terra::animate(lcc[[site_index]], 
                   levels = lc_plot_cols$name, # see "_util/_util_misc.R"
                   col = lc_plot_cols$color,
                   type = "classes",
                   main = paste0(inset_map_titles[site_index], ", ",
                                 # "Land Cover, ",
                                 1987:2017),
                   maxcell = 500000,
                   colNA = "gray95",
                   legend = FALSE,
                   mar = c(mval+0.5, mval, mval, mval-0.1),
                   n = 1),
    movie.name = file_out_tmp,
    # ani.dev = "pdf", ani.type = "pdf", 
    # ani.height = 350, ani.width = 350,
    # ani.res = 100 #units = "px"
    )
  
  # Reload:
  tmp_gif <- image_animate(image_read(file_out_tmp), fps = 5)
  image_write(tmp_gif, file_out_tmp)
})



# ------------------------------------------- #
# old
# ------------------------------------------- #

lapply(c(1,3,9), function(site_index) {
  tic()
  cc_save_gif(raster = age_r[[site_index]], 
            npixels = 500000, 
            titles = paste0("Abandonment Age, ", 
                            capwords(site_df$site[site_index]), " ", 1987:2017),
            frames_per_second = 5,
            file_out = paste0(p_plots, "/", run_label, "/", "animations/age",
                              site_df$label[site_index], ".gif"))
toc()
})


# ----------------------------------------------- #
# Land cover alone
# ------------------------------------------- #

lapply(c(1,3,9), function(site_index) {
  file_out_tmp <- paste0(p_plots, "/", run_label, "/", "animations/lcr",
                              site_df$label[site_index], ".gif")
  saveGIF(
    raster::animate(lcr[[site_index]], 
                   # levels = lc_plot_cols$name, # see "_util/_util_misc.R"
                   col = lc_plot_cols$color,
                   main = paste0(inset_map_titles[site_index], ", ", 1987:2017),
                   # main = paste0("Land Cover, ", capwords(site_df$site[site_index]), " ", 1987:2017),
                   maxpixels = 500000,
                   colNA = "gray95",
                   legend = TRUE, n = 1),
    movie.name = file_out_tmp
    )
  
  # Reload:
  tmp_gif <- image_animate(image_read(file_out_tmp), fps = 5)
  image_write(tmp_gif, file_out_tmp)
})

```




```{r plot-trajectory-per-pixel}
bs <- brick(paste0(p_dat, "Abandonment/belarus_small.tif"))
names(bs) <- gsub("smolensk", "y", names(bs))
bs_dt <- as.data.table.raster(bs)

# subset the data.table, for plotting
sub <- bs_dt[1:10, -c(1,2)]
sub[, pixel := c(1:10)]


sub_melt <- melt(sub, id.vars = "pixel", 
                 variable.name = "year", 
                 value.name = "land_use", na.rm = TRUE)

sub_melt$year <- gsub("y", "", sub_melt$year)

str(sub_melt)

# set colors for plotting
show_col(terrain.colors(9))

lc_cols <- scale_color_manual(name = "Land Cover",
                     labels = c("1" = "1. Non-veg",
                                "2" = "2. Woody veg",
                                "3" = "3. Crop",
                                "4" = "4. Grassland"),
                     values = c("1" = "gray80",
                                "2" = terrain.colors(9)[1], # dark green
                                "3" = terrain.colors(9)[5], # gold
                                "4" = terrain.colors(9)[3] # light green
                                )
                     ) 

# this plot shows the land-use trajectories of 10 individual pixels.
gg_10_px_traj <- ggplot(data = sub_melt, 
                        mapping = aes(x = year, y = land_use, group = pixel)) +
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0)) +
  geom_line(mapping = aes(color = factor(land_use)), size = 2) +
  facet_grid(rows = vars(pixel), scales = "free_x", switch = "x") + 
  ylim(1, 4) + lc_cols

gg_10_px_traj
#       1. Non-vegetated area (e.g. water, urban, barren land)
#       2. Woody vegetation
#       3. Cropland 
#       4. Herbaceous land (e.g. grassland)


```

```{r check-temporal-filter-counts}
# Not strictly necessary - check cluster outputs and the number of pixels affected by the temporal filters

# note that in the run: _2021-03-05, the temporal filter counter calculated the percent_affected without multiplying by 100, so percent_affected = value/nrow(dt), and therefore, nrow(dt) = value/percent_affected

# for run: _2021_03_13, the percent was fixed, so nrow(dt) = 100 * value / (percent_affected)
x <- "shaanxi"

percent_affected_df <- 
  lapply(site_df$site, function(x) {
    # load the counts list
    load(file = paste0(p_dat_derived, run_label, "/", x, 
                       "_temporal_filter_counts_l", run_label, ".rds"), verbose = TRUE)
    
    cat(fill = TRUE, 
        "number of rows:", 
        temporal_filter_counts_l$edge_nrow_affected %>%
               mutate(nrow = 100 * value/percent_affected) %>% 
               .$nrow %>% unique())
    
    percent_affected_df <- 
      bind_rows(temporal_filter_counts_l$nrow_affected_df,
                temporal_filter_counts_l$edge_nrow_affected %>% 
                  filter(edge == "start") %>% 
                  select(filter = edge, nrow_affected = value)
                ) %>%
      mutate(site = x,
             nrow_abn = temporal_filter_counts_l$edge_nrow_affected %>%
               mutate(nrow = 100 * value/percent_affected) %>% 
               .$nrow %>% round(digits = 0) %>% unique(),
             # percent_affected = 100 * nrow_affected/nrow
             ) %>% 
      as_tibble() 
    
    # percent_affected_df %>% mutate(filter = recode(filter, start = "starting_edge" ))

    percent_affected_df
    
}) %>% bind_rows()

site_df_w_size <- read_csv(file = "/Users/christophercrawford/Google_Drive/_Projects/abandonment_trajectories/data_derived/site_df_w_size.csv") 

percent_affected_df <- percent_affected_df %>% 
  left_join(. , select(site_df_w_size, site, nrow_total = ncell)) %>%
  select(-nrow) %>% 
  pivot_longer(cols = c("nrow_abn", "nrow_total"), 
               names_to = "relative_to", names_prefix = "nrow_", 
               values_to = "nrow") %>%
  mutate(percent_affected = 100 * nrow_affected/nrow)

# write to file:

write_csv(percent_affected_df, file = paste0(p_dat_derived, run_label, "/derived_data/temporal_filters_percent_affected_df", run_label,".csv"))



# plot #
gg_temp_filter_counts <- 
  ggplot(data = percent_affected_df# %>% filter(filter %in% c("either", "starting_edge"))
         ,
       mapping = aes(x = fct_reorder(site, percent_affected), y = percent_affected, color = filter)) +
  geom_point() +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 340, vjust = 1, hjust = 0)) + 
  labs(x = "Site", y = "Percent of abandoned pixels affected by filters",
       title = "Scope of pixels affected by moving window temporal filters") + 
  scale_colour_discrete(labels = c(
    eight_yr = "8-year filter",
    five_yr = "5-year filter",
    either = "Either 5- or 8-year filters", 
    starting_edge = "Starting edge filter")) + 
  facet_grid(rows = vars(relative_to), scales = "free_y", 
             labeller = labeller(relative_to = c(abn = "Relative to abandoned pixels", #old = "new",
                                                 total = "Relative to total pixels")))

# save
ggsave(plot = gg_temp_filter_counts,
       filename = paste0(p_output, "plots/", run_label, "/temp_filter_counts_plot_full.pdf"), 
       width = 8, height = 6, units = "in")

# 
percent_affected_df %>% group_by(filter, relative_to) %>% summarize(max = max(percent_affected))

```



# Save SI Figures

## Facet grid figures

```{r abn-area-by-age-bin-grid}
# A figure showing the area abandoned at each site over time, by age class

# plot
gg_age_class_bins_grid <-
    ggplot(data = persistence_dat) + 
    theme_classic() +
    labs(y = expression("Area abandoned (10"^{3}*" ha)") , 
         x = "Year", 
         #title = "Area of Abandoned Land, by Age Class",
         #subtitle = subtitle,
         fill = "Age (years)") +
    geom_col(mapping = aes(x = year, y = area_ha / 10^3, 
                           fill = fct_reorder(bins, age)),
             position = position_stack(reverse = FALSE)) +
    theme(
      #legend.position = c(0.1, 0.8),
      #legend.position = "bottom",
      legend.title = element_text(size = 10)
      ) +
    scale_fill_brewer(palette = "Greens", direction = 1) +
      facet_wrap(vars(site), scales = "free_y", labeller = as_labeller(cap_update_labels)) +  # see _util_misc.R
    theme(legend.position = c(0.885, 0.12), 
        legend.spacing = unit(2, units = "mm"), 
        legend.margin = margin(t = 0, r = 0, b = 0, l = 0),
        strip.text.x = element_text(size = 8)) + 
      guides(fill = guide_legend(nrow = 3, ncol = 2))




# save
ggsave(plot = gg_age_class_bins_grid,
       filename = paste0(p_plots, run_label,
       "/abn_area_by_age_grid", run_label, ".pdf"),
       width = 7, height = 5, units = "in", dpi = 400
)



# add line for area:

gg_age_class_bins_grid

gg_age_class_bins_grid + 
  geom_line(data = area_dat %>% filter(lc == "Abandoned (>=5)"),
            mapping = aes(x = year, y = area_ha / (10^3)),
            size = 2, alpha = 0.8)


# save a version of the above, with the total area abandoned (>= 5, and >=1 years)
ggsave(plot = gg_age_class_bins_grid + 
         geom_line(data = area_dat %>% filter(grepl("Abandoned", lc)),
                   mapping = aes(x = year, y = area_ha / (10^3),
                                 group = lc, color = lc), 
                   size = 1.5) +
         scale_color_manual(values = plot_cols[grep("Abandoned", names(plot_cols))],
                            name = "Total area abandoned"),
       filename = paste0(p_plots, run_label,
                         "/abn_area_by_age_grid_w_total", run_label, ".pdf"),
       width = 7.5, height = 6, units = "in", dpi = 400
)



# FINAL VERSION: save a version of the above, with the total area abandoned (>= 5) and the potential area abandoned
ggsave(plot = gg_age_class_bins_grid + 
         geom_line(data = 
                     bind_rows(
                       area_dat %>% mutate(type = "Observed"),
                       potential_area_dat %>% mutate(type = "Potential")
                       ) %>%
                     filter(lc %in% c("Abandoned (>=5)")),
                   mapping = aes(x = year, y = area_ha / (10^3),
                                 group = type, linetype = type), 
                   size = 1.5) + 
         labs(linetype = "Scenario", fill = "Observed age (years)") + theme(legend.position = c(0.88, 0.11)),
       filename = paste0(p_plots, run_label,
                         "/abn_area_by_age_grid_w_obs_pot_lines", run_label, ".pdf"),
       width = 7.2, height = 5.5, units = "in", dpi = 400
)




# version for ppt, with just the observed area:
ggsave(plot =
         ggplot(data = persistence_dat) + 
           theme_classic() +
           labs(y = expression("Area abandoned (10"^{3}*" ha)") , 
                x = "Year", fill = "Age (years)") +
           theme(legend.title = element_text(size = 10)) +
           scale_fill_brewer(palette = "Greens", direction = 1) +
           facet_wrap(vars(site), scales = "free_y", labeller =
                        as_labeller(cap_update_labels)) +
           theme(legend.position = c(0.885, 0.12), 
                 legend.spacing = unit(2, units = "mm"),
                 legend.margin = margin(t = 0, r = 0, b = 0, l = 0),
                 strip.text.x = element_text(size = 8)) + 
           guides(fill = guide_legend(nrow = 3, ncol = 2)) + 
           geom_line(data = 
                       bind_rows(
                       area_dat %>% mutate(type = "Observed"),
                       potential_area_dat %>% mutate(type = "Potential")
                       ) %>%
                     filter(lc %in% c("Abandoned (>=5)")),
                   mapping = aes(x = year, y = area_ha / (10^3),
                                 group = type, 
                                 alpha = type,
                                 linetype = type
                                 ), 
                   size = 1.5
                   ) + 
         labs(linetype = "Scenario", 
              fill = "Observed age (years)",
              alpha = "Scenario") + 
           scale_alpha_discrete(range = c(0, 1)) +
         theme(legend.position = c(0.88, 0.11)),
       filename = paste0(p_plots, run_label,
                         "/abn_area_by_age_grid_w_pot_area_line_only", run_label, ".pdf"),
       width = 7.2, height = 5.5, units = "in", dpi = 400
)

ggsave(plot =
         ggplot(data = persistence_dat) + 
           theme_classic() +
           labs(y = expression("Area abandoned (10"^{3}*" ha)") , 
                x = "Year", fill = "Age (years)") +
           theme(legend.title = element_text(size = 10)) +
           scale_fill_brewer(palette = "Greens", direction = 1) +
           facet_wrap(vars(site), scales = "free_y", labeller =
                        as_labeller(cap_update_labels)) +
           theme(legend.position = c(0.885, 0.12), 
                 legend.spacing = unit(2, units = "mm"),
                 legend.margin = margin(t = 0, r = 0, b = 0, l = 0),
                 strip.text.x = element_text(size = 8)) + 
           guides(fill = guide_legend(nrow = 3, ncol = 2)) + 
           geom_line(data = 
                       bind_rows(
                       area_dat %>% mutate(type = "Observed"),
                       potential_area_dat %>% mutate(type = "Potential")
                       ) %>%
                     filter(lc %in% c("Abandoned (>=5)")),
                   mapping = aes(x = year, y = area_ha / (10^3),
                                 group = type, linetype = type), 
                   size = 1.5) + 
         labs(linetype = "Scenario", 
              fill = "Observed age (years)",
              alpha = "Scenario") + 
         theme(legend.position = c(0.88, 0.11)),
       filename = paste0(p_plots, run_label,
                         "/abn_area_by_age_grid_w_obs_area_line_only", run_label, ".pdf"),
       width = 7.2, height = 5.5, units = "in", dpi = 400
)

ggsave(plot =
         gg_age_class_bins_grid + 
         geom_line(data = 
                     bind_rows(
                       area_dat %>% mutate(type = "Observed"),
                       potential_area_dat %>% mutate(type = "Potential")
                       ) %>%
                     filter(lc %in% c("Abandoned (>=5)")),
                   mapping = aes(x = year, y = area_ha / (10^3),
                                 group = type, alpha = type), 
                   size = 1.5) + 
         labs(alpha = "Scenario", fill = "Observed age (years)") + 
           scale_alpha_discrete(range = c(1, 0)) + 
         theme(legend.position = c(0.88, 0.11)),
       filename = paste0(p_plots, run_label,
                         "/abn_area_by_age_grid_w_obs_area_line", run_label, ".pdf"),
       width = 7.2, height = 5.5, units = "in", dpi = 400
)

```

```{r area-land-cover-facet-grid}
# ---------------------------------------------------------------------- #
# plot area in each land cover class in a facet grid
# ---------------------------------------------------------------------- #
gg_area_lc_facet_grid <-
  ggplot(data = area_dat %>% filter(lc != "Abandoned (>=1)")
         ) +
  theme_classic() +
  # theme(axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0)) +
  labs(y = expression("Area  (10"^{6}*" ha)") , 
       x = "Year", 
       color = NULL) + 
  geom_line(mapping = aes(x = year, y = area_ha / (10^6), # convert to millions of ha
                          group = lc, color = lc),
            size = 1.5) + 
  theme(legend.position = c(0.9, 0.12),
        legend.title = element_text(size = 10)) +
  scale_color_manual(values = plot_cols[-1]
                     ) +
  guides(color = guide_legend(ncol = 1)) +
  facet_wrap(vars(site),
             labeller = as_labeller(cap_update_labels),
             scales = "free")



ggsave(plot = gg_area_lc_facet_grid,
       filename = paste0(p_plots, run_label, "/area_by_lc_grid_all", run_label, ".pdf"), 
    width = 7, height = 5, units = "in"
)


ggsave(plot = gg_area_lc_facet_grid,
       filename = paste0(p_plots, run_label, "/area_by_lc_grid", run_label, ".pdf"), 
    width = 7, height = 5, units = "in"
)

```


```{r potential-vs-obs-area-abn}
# ---------------------------------------------------------------------- #
# plot observed area abandoned vs. potential area abandoned, assuming no recultivation
# ---------------------------------------------------------------------- #
gg_potential_area_facet_grid <-
  ggplot(data = 
           bind_rows(area_dat %>% mutate(type = "Observed"),
                     potential_area_dat %>% mutate(type = "Potential")) %>%
           filter(grepl("Abandoned", lc)),
         mapping = aes(x = year, y = area_ha / (10^6), # convert to millions of ha
                       # group = interaction(lc, type), 
                       color = lc, linetype = type)) +
  theme_classic() +
  # theme(axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0)) +
  labs(y = expression("Area  (10"^{6}*" ha)") , 
       x = "Year", 
       # title = "Observed Abandonment vs. Potential Abandonment without recultivation",
       color = "Abandonment\ndefinition", linetype = "Scenario") + 
  geom_line(size = 1.5) + 
  theme(legend.position = c(0.88, 0.11),
        legend.title = element_text(size = 9),
        legend.spacing = unit(-1, units = "mm"),
        legend.background = element_blank(),
        # legend.box.background = element_rect(colour = "black")
        ) +
  guides(color = guide_legend(nrow = 2),
         linetype = guide_legend(nrow = 2)) +
    # guides(color = guide_legend(nrow = 5, ncol = 1)) +

  # scale_color_viridis(discrete = TRUE) +
  scale_color_manual(values = brewer_pal(palette = "Greens")(7)[c(2,6)]) +
  # scale_color_manual(values = plot_cols[1:2]) +
  facet_wrap(vars(site), scales = "free", labeller = as_labeller(fancy_labels))



ggsave(plot = gg_potential_area_facet_grid,
       filename = paste0(p_plots, run_label, "/potential_area_grid", run_label, ".pdf"), 
    width = 7, height = 6, units = "in"
)


show_col(brewer_pal(palette = "Greens")(7))
```

```{r total-carbon}
# load 
# final version
# site_carbon_df <- read_csv(file = paste0(p_derived2, "/carbon_df", run_label,".csv"))
# site_carbon_df_terra <- read_csv(file = paste0(p_derived2, "/carbon_df_terra.csv"))

site_carbon_df0
site_carbon_df %>%
  mutate(
    # old20 = site_carbon_df0$C_up_to_20,
         old21 = site_carbon_df0$C_21_30,
         oldtotal = site_carbon_df0$total_C_Mg,
         # diff20 = all.equal(old20, C_up_to_20),
         diff21 = 100 * (C_21_30 - old21) / C_21_30,
         difftotal = 100 * (total_C_Mg - oldtotal) / total_C_Mg
         ) %>%
  filter(threshold == 5, type == "Observed",
         year %in% c(2012:2017))

# plot:

gg_total_c <-
  ggplot(data = site_carbon_df %>% filter(threshold == 5), 
         mapping = aes(x = year, y = total_C_Mg / 10^6,
                       # group = interaction(type, threshold),
                       # color = as_factor(threshold),
                       linetype = type
                       )) +
  geom_line(size = 1.5, 
            col = brewer_pal(palette = "Greens")(7)[6]) + 
  # scale_color_manual(values = brewer_pal(palette = "Greens")(7)[c(3, 6)]) +
  labs(x = "Year", 
       y = expression("Carbon Accumulated (10"^{6}*" Mg C)"),
       linetype = "Scenario"
       ) +
  theme_classic() + 
  theme(
    legend.position = c(0.88, 0.11),
    # legend.position = "bottom",
        legend.title = element_text(size = 9),
        legend.spacing = unit(-1, units = "mm"),
        legend.background = element_blank(),
    strip.text.x = element_text(size = 8)
        # legend.box.background = element_rect(colour = "black")
        ) +
  guides(linetype = guide_legend(ncol = 1)) +
    facet_wrap(vars(site), scales = "free_y", labeller = as_labeller(cap_update_labels))

# figure for MS:
ggsave(gg_total_c,
       file = paste0(p_plots, run_label, "/carbon/carbon_accumulated_over_time.pdf"), 
       width = 6.5, height = 5, units = "in", dpi = 400)

# figure for PPT:
# just potential, to be shown before both lines together
gg_total_c_pot <-
  ggplot(data = site_carbon_df %>% filter(threshold == 5), 
         mapping = aes(x = year, y = total_C_Mg / 10^6,
                       linetype = type, alpha = type
                       )) +
  geom_line(size = 1.5, 
            col = brewer_pal(palette = "Greens")(7)[6]) + 
  labs(x = "Year", 
       y = expression("Carbon Accumulated (10"^{6}*" Mg C)"),
       linetype = "Scenario", alpha = "Scenario"
       ) +
  theme_classic() + 
    theme(
      legend.position = c(0.88, 0.11),
        legend.title = element_text(size = 9),
        legend.spacing = unit(-1, units = "mm"),
        legend.background = element_blank(),
    strip.text.x = element_text(size = 8)
        ) +
    scale_alpha_discrete(range = c(0, 1)) +
    guides(linetype = guide_legend(ncol = 1)) +
    facet_wrap(vars(site), scales = "free_y", labeller = as_labeller(cap_update_labels))

ggsave(gg_total_c_pot,
       file = paste0(p_plots, run_label, "/carbon/carbon_accumulated_over_time_pot_only.pdf"), 
       width = 6.5, height = 5, units = "in", dpi = 400)

# ---------- #
site_carbon_tmp <- 
  site_carbon_df %>% 
  filter(#type == "Observed", 
         threshold == 5) %>%
  left_join(
    bind_rows(potential_area_dat %>% mutate(type = "Potential"),
              area_dat %>% mutate(type = "Observed")) %>% 
      filter(grepl(">=5", lc)) %>%
      mutate(threshold = 5) %>% 
      select(site, year, area_abn_ha = area_ha, type, threshold),
    by = c("site", "year", "type", "threshold")) %>%
  left_join(area_summary_df %>% 
              select(site,
                     area_abn_ha_2017,
                     total_site_area_ha_2017, 
                     total_crop_extent_ha,
                     area_2017_as_prop_total_crop), by = "site") %>%
  mutate(var = total_C_Mg /
           # 10^6 /
           area_abn_ha,
         # area_abn_ha_2017
         # total_crop_extent_ha
         # area_2017_as_prop_total_crop
         )

site_carbon_tmp <- 
  site_carbon_tmp %>%
  left_join(site_carbon_tmp %>%
              filter(year == 2017) %>%
              select(site, type, var_2017 = var), 
  by = c("site", "type")
  )

# add inset 
gg_carbon_inset <-
  ggplot(data = site_carbon_tmp %>% filter(!is.na(var)), 
         mapping = aes(x = year, y = var, linetype = type,
                       color = fct_reorder(site, var_2017, .desc = TRUE))) +
  geom_line(
    size = 1
            ) + 
  labs(x = "Year", 
       y = expression(#"10"^{6}*
         "Mg C / ha"),
       linetype = "Scenario", color = "Site"
       ) +
  scale_y_continuous(breaks = seq(from = 0, to = 70, by = 5)) +
  scale_x_continuous(breaks = seq(from = 1990, to = 2020, by = 5)) +
  scale_color_discrete(labels = fancy_labels) + 
  theme_classic() + 
  theme(legend.title = element_text(size = 9),
        legend.spacing = unit(-1, units = "mm"),
        legend.background = element_blank()) +
  guides(color = guide_legend(ncol = 1, reverse = FALSE))

# save 
ggsave(gg_carbon_inset,
       file = paste0(p_plots, run_label, "/carbon/carbon_accumulated_per_ha.pdf"), 
       width = 6, height = 6, units = "in", dpi = 400)
  


# ---------------------------------------------- #
# Save the combination figure:
ggsave(ggdraw() +
         draw_plot(gg_total_c, width = 0.75) +
         draw_plot(gg_carbon_inset, x = 0.575, y = .1, width = .45, height = .3),
       file = paste0(p_plots, run_label,
                     "/carbon/carbon_accumulated_over_time_inset.pdf"), 
       width = 8, height = 5.5, units = "in", dpi = 400)

```


```{r potential-age-bins}
gg_potential_area_age_bins

potential_area_dat
potential_persistence_dat

gg_potential_area_age_bins <-
    ggplot(data = potential_persistence_dat) + 
    theme_classic() +
    labs(y = expression("Area abandoned (10"^{3}*" ha)") , 
         x = "Year", 
         #title = "Area of Abandoned Land, by Age Class",
         #subtitle = subtitle,
         fill = "Age (years)",
         linetype = NULL) +
    geom_col(mapping = aes(x = year, y = area_ha / 10^3, 
                           fill = fct_reorder(bins, age)),
             position = position_stack(reverse = FALSE)) +
    theme(
      #legend.position = c(0.1, 0.8),
      #legend.position = "bottom",
      legend.title = element_text(size = 10)
      ) +
    scale_fill_brewer(palette = "Greens", direction = 1) +
      
      facet_wrap(vars(site), scales = "free_y", labeller = as_labeller(cap_update_labels)) +  # see _util_misc.R
    theme(legend.position = c(0.885, 0.12), 
        legend.spacing = unit(2, units = "mm"), 
        legend.margin = margin(t = 0, r = 0, b = 0, l = 0)) + 
      guides(fill = guide_legend(nrow = 3, ncol = 2)) + 
      
      
      # add total area:
  geom_line(data = 
           bind_rows(area_dat %>% mutate(type = "Observed"),
                     potential_area_dat %>% mutate(type = "Potential")) %>%
           filter(lc == "Abandoned (>=5)"
             # grepl("Abandoned", lc)
                  ),
         mapping = aes(x = year, y = area_ha / (10^3), # convert to millions of ha
                       # group = interaction(lc, type), 
                       # color = lc, 
                       linetype = type),
         size = 1.5)



# save
ggsave(plot = gg_potential_area_age_bins,
       filename = paste0(p_plots, run_label,
       "/potential_area_age_bins", run_label, ".pdf"),
       width = 7, height = 5, units = "in", dpi = 400
)

```


```{r turnover-grid}
turnover_dat

# ---------------------------------------------------------------------- #
# plot facet grid
# ---------------------------------------------------------------------- #

gg_turnover_facet_grid <-
  ggplot() +
    theme_classic() + 
    labs(y = expression("Annual area turnover (10"^{3}*" ha)"), 
         x = "Year", 
         #title = "Change in Area of Abandoned Land",
         #subtitle = subtitle,
         fill = NULL, #"Annual Change",
         color = NULL) + 
    scale_color_manual(values = c("black"),
                       labels = "Net Change") + 
    scale_fill_manual(values = brewer_pal(palette = "Greens")(5)[3:2],
                      labels = c("Area Gained", "Area Lost")) + 
    geom_col(data = turnover_dat %>% filter(change != "net"),
             mapping = aes(x = year, y = area_ha / (10^3), 
                           group = change, fill = change)) + 
    geom_line(data = turnover_dat %>% filter(change == "net"),
              mapping = aes(x = year, y = area_ha / (10^3), color = "Net Change in Area"),
              size = 1) + 
  facet_wrap(vars(site),
             labeller = as_labeller(cap_update_labels),
             scales = "free") +
    theme(legend.position = c(0.9, 0.12),
          legend.title = element_text(size = 10))
  
gg_turnover_facet_grid


ggsave(plot = gg_turnover_facet_grid,
       filename = paste0(p_plots, run_label, "/turnover_grid", run_label, ".pdf"), 
    width = 7, height = 5, units = "in"
)
```


```{r decay-facet-grid}
# See "2_decay_models.Rmd" chunk "decay-mod-facet-grid"
# File path: 
paste0(p_plots, run_label, "/decay/", "decay_mod_grid", run_label, ".pdf")
```


## Misc
```{r save-si_panel-composite-figs}
# save "si_panel" composite figures

# these show:
# a) accumulation of abandoned land by age class, 
# b) decay of abandoned land by year abandoned, 
# c) the area in each land cover class through time (including both land that has been abandoned for five or more years, as well as any land abandoned for 1 or more years, therefore including short-term fallows),
# d) the annual turnover of abandoned land through time, and 
# e) the distribution of abandonment duration for all periods of cropland abandonment (top) and the maximum duration of abandonment at each pixel (bottom).

# note that the 

lapply(1:11, function(i) {
  cc_si_panel_plots(
    input_path = paste0(p_dat_derived, run_label), 
    output_path = paste0(p_output, "plots/", run_label, "/"),
    outfile_label = paste0(run_label, site_df$label[i]),
    site = site_df$site[i]
  )
})


# Shaanxi main text panel

cc_si_panel_plots(
    input_path = paste0(p_dat_derived, run_label), 
    output_path = paste0(p_output, "plots/", run_label, "/"),
    outfile_label = paste0(run_label, site_df$label[9]),
    site = site_df$site[9],
    main_text_panel = TRUE
  )
```

```{r save-indiv-figs}
# run plots loop:

# This saves individual versions of the panels in the si_panel composite figures

if(!dir.exists(paste0(p_output, "plots/", run_label))) {
  dir.create(paste0(p_output, "plots/", run_label))
}

lapply(1:11, function(i) {
  cc_save_area_persistence_plots(
    input_path = paste0(p_dat_derived, run_label), 
    output_path = paste0(p_output, "plots/", run_label, "/"),
    site_label = site_df$label[i], 
    run_label = run_label,
    outfile_label = paste0(run_label, site_df$label[i]),
    site = site_df$site[i],
    subtitle = paste0(site_df$description[i]), # ", 5- and 8-year temp. filter"
    subtitle_all = paste0(site_df$description[i], ", all abandonment (no threshold)"),
    save_all = FALSE
  )
})
```

```{r produce-summary-figures-by-site}
# Produce a set of summary figures for each site, in sequence, mostly for presentations


input_path <- paste0(p_dat_derived, run_label)
output_path <- paste0(p_output, "plots/", run_label, "/presentation/")
outfile_label <- paste0(run_label, site_df$label[9])
site <- site_df$site[9]
col_palette <- plot_cols


cc_presentation_summary_plots <- function(input_path, 
                             output_path,
                             outfile_label = paste0(run_label, site_label),
                             site,
                             width = 9, height = 11,
                             col_palette = plot_cols,
                             subtitle = NULL,
                             site_desc) {
  
  area_df <- read_csv(file = paste0(input_path, "/", site, "_result_area",  run_label, ".csv"))
  persistence_df <- read.csv(file = paste0(input_path, "/", site, "_result_persistence",  run_label, ".csv"))
  turnover_df <- read.csv(file = paste0(input_path, "/", site, "_result_turnover",  run_label, ".csv"))
  length_distill_df <- read.csv(file = paste0(input_path, "/", "length_distill_df", run_label, ".csv"))
  mean_length_df <- read.csv(file = paste0(input_path, "/", "mean_length_df", run_label, ".csv"))
  # site <- "shaanxi"
  input_site <- site
  
  # ------------- calculate total area per lc, with abandonment ---------------- #
  gg_lc_abn_area <-
    ggplot(data  = area_df #%>% filter(lc != "Abandoned (>=1)")
    ) +
    theme_classic() +
    # theme(axis.text.x = element_text(angle = 320, vjust = 1, hjust = 0)) +
    labs(y = expression("Area  (10"^{6}*" ha)") , 
         x = "Year", 
         #title = "Area by Land Cover, with Abandonment",
         #subtitle = subtitle,
         color = NULL #"Land Cover"
         ) + 
    geom_line(mapping = aes(x = year, y = area_ha / (10^6), # convert to millions of ha
                            group = lc, color = fct_reorder(lc, lc, .desc = TRUE)),
              size = 1.5) + 
    theme(legend.position = "bottom",
          legend.title = element_text(size = 10)) +
    guides(color = guide_legend(nrow = 2, ncol = 3)) +
    
    scale_color_manual(values = col_palette)
  # scale_color_brewer(palette = "Greens")
  # scale_color_viridis(discrete = TRUE)
  
  
  # ------------------------ abandonment persistence --------------------------- #
  gg_persistence <-
    ggplot(data = persistence_df) + 
    theme_classic() + 
    labs(#title = "Persistence of Abandoned Land",
      #subtitle = subtitle,
      color = "Year \nAbandoned \n(Cohort)") + 
    scale_color_distiller(palette = "Greens") +
    theme(legend.position = c(0.9, 0.7),
      #legend.direction = "horizontal"
      #legend.position = "bottom", legend.text = element_text(angle = 330, vjust = 1, hjust = 0)
      legend.title = element_text(size = 10)
    ) +
    geom_line(mapping = aes(x = age, y = proportion,
                            group = year_abn, color = year_abn), 
              size = 1.25) + 
    labs(y = "Proportion remaining abandoned", 
         x = "Years since initial abandonment")
  
  
  # -------------------- calculate the abandonment area turnover ------------------- #
  gg_turnover <-
    ggplot() +
    theme_classic() + 
    labs(y = expression("Annual area turnover (10"^{3}*" ha)"), 
         x = "Year", 
         #title = "Change in Area of Abandoned Land",
         #subtitle = subtitle,
         fill = NULL, #"Annual Change",
         color = NULL) + 
    scale_color_manual(values = c("black"),
                       labels = "Net Change") + 
    scale_fill_manual(values = brewer_pal(palette = "Greens")(5)[3:2],
                      labels = c("Area Gained", "Area Lost")) + 
    theme(legend.position = "bottom",
      #legend.position = c(0.2, 0.7),
      # legend.direction = "horizontal",
      legend.title = element_text(size = 10)
    ) +
    geom_col(data = filter(turnover_df, change != "net"),
             mapping = aes(x = year, y = area_ha / (10^3), 
                           group = change, fill = change)) + 
    geom_line(data = filter(turnover_df, change == "net"),
              mapping = aes(x = year, y = area_ha / (10^3), color = "Net Change in Area"),
              size = 1.5)
  
  
  # -------------------- plot abandonment area by age class ------------------- #
  gg_age_class_bins <-
    ggplot(data = persistence_df) + 
    theme_classic() +
    labs(y = expression("Area abandoned (10"^{3}*" ha)") , 
         x = "Year", 
         #title = "Area of Abandoned Land, by Age Class",
         #subtitle = subtitle,
         fill = "Age \n(years)") +
    geom_col(mapping = aes(x = year, y = area_ha / 10^3, 
                           fill = fct_reorder(bins, age)),
             position = position_stack(reverse = FALSE)) +
    theme(
      #legend.position = c(0.1, 0.8),
      #legend.position = "bottom",
      legend.title = element_text(size = 10)
      ) +
    scale_fill_brewer(palette = "Greens", direction = 1)
  
  
  # histogram:
  gg_hist <-
    ggplot(data = filter(length_distill_df, site == input_site,# site_df$site[i], 
                         length > 0)) + 
    theme_classic() +
    labs(color = "Mean") +
    xlab("Time abandoned (years)") + ylab(expression("Count  (10"^{6}*" pixels)")) +
    geom_col(mapping = aes(x = length, y = freq/(10^6)), fill = "gray70") +
    facet_grid(rows = vars(length_type), scales = "free",
               labeller = labeller(length_type = c(all = "all abn. periods", #old = "new",
                                                   max = "max. length \nper pixel"))) +
    # means
    geom_vline(data = filter(mean_length_df, site == input_site, #site_df$site[i], 
                             abn_threshold != 3), 
               mapping = aes(xintercept = mean_length, color = as_factor(abn_threshold)#, linetype = "mean (>1)"
                             ),
               show.legend = TRUE, size = 1, linetype = "dashed"
               ) + 
    
    scale_color_manual(values = c("5" = "#C51B7D", "1" = "#F1B6DA"), 
                       labels = c("5" = "Mean (>=5)", "1" = "Mean (>1)")) +
    
    theme(#legend.position = c(0.85, 0.95)
          )
  
  title <- ggdraw() + 
              draw_label(
                site_desc,
                fontfamily = "Tahoma",
                x = 0, hjust = 0) +
              theme(# add margin on the left of the drawing canvas,
                # so title is aligned with left edge of first plot
                plot.margin = margin(0, 0, 0, 5))
  
  plot_combo1 <- plot_grid(gg_lc_abn_area + theme(legend.position = "top"), 
                     gg_age_class_bins + theme(legend.position = c(0.2, 0.7)),
                     ncol = 2, rel_widths = c(1, 1), align = "h", axis = "b")
  
  plot_combo2 <- plot_grid(gg_persistence, 
                     gg_turnover + theme(legend.position = "top"),
                     ncol = 2, rel_widths = c(1, 1), align = "h", axis = "b")
  
  # save fig 1:
  ggsave(filename = paste0(output_path, "summary_panel_1", outfile_label, ".png"),
         plot = plot_grid(title, plot_combo1,
                          ncol = 1, nrow = 2, rel_heights = c(0.2, 4)),
       width = 9, height = 4.5, units = "in", dpi = 400)
  
  # save fig 2:
  ggsave(filename = paste0(output_path, "summary_panel_2", outfile_label, ".png"),
         plot = plot_grid(title, plot_combo2,
                          ncol = 1, nrow = 2, rel_heights = c(0.2, 4)),
       width = 9, height = 4.5, units = "in", dpi = 400)
  
  
  # for main preso: shaanxi
  # ggsave(filename = paste0(output_path, "summary_panel_1", outfile_label, "_main_preso.png"),
  #        plot = plot_combo1,
  #      width = 9, height = 4.5, units = "in", dpi = 400)
  # 
  # ggsave(filename = paste0(output_path, "summary_panel_2", outfile_label, "_main_preso.png"),
  #        plot = plot_combo2,
  #      width = 9, height = 4.5, units = "in", dpi = 400)
  
}

lapply(1:11, function(i) {
  cc_presentation_summary_plots(
    input_path = paste0(p_dat_derived, run_label), 
    output_path = paste0(p_output, "plots/", run_label, "/presentation/"),
    outfile_label = paste0(run_label, site_df$label[i]),
    site = site_df$site[i],
    site_desc = site_df$description[i]
  )
})

```


```{r extrapolation-combo}
# see chunk 'extrapolation-panel' in 2_decay_models.Rmd
# just b and c, for presentation
ggsave(plot = 
         plot_grid(gg_extrapolation_panel_age_b + 
                     scale_x_continuous(breaks = seq(from = 2000, to = 2120, by = 40)) #+
                     # theme(legend.position = "bottom") + 
                     # guides(col = guide_legend(nrow = 4, ncol = 3))
                   , 
                     gg_extrapolation_panel_area_c + scale_x_continuous(breaks = seq(from = 2000, to = 2120, by = 40)) #+ theme(legend.position = "bottom")
                   ,
                     #rel_widths = c(1, 1.2),
                     # labels = c("b", "c"),
                   ncol = 2
                     ),
       filename = paste0(p_plots, run_label, "/presentation/", "extrapolation_combo", run_label, ".png"),
       width = 9, height = 3.5, units = "in", dpi = 400
)
```


# Global Site Map

```{r load-world-data}
r_extent <- lapply(r, raster::extent)
plot(r_extent$belarus)

site_sf <- lapply(age_r, FUN = function(x) {
  x %>% 
    extent() %>% 
    as(., 'SpatialPolygons') %>%
    st_as_sf()
  }) %>% 
  bind_rows() %>% 
  mutate(site = site_df$site, 
         label = gsub("_", "", site_df$label),
         lab_desc = paste0("(", label, ") ", site_df$description),
         desc = site_df$description)


dev.off()

plot(site_sf$geometry)

# basemap
world <- ne_countries(scale = "small", returnclass = "sf") #%>% st_make_valid() # can set returnclass to sf or sp.
world_10 <- ne_countries(scale = "large", returnclass = "sf") %>% st_make_valid() # can set returnclass to sf or sp.
world_states <- ne_states(returnclass = "sf") %>% st_make_valid()


# set crs for site_sf 
st_crs(site_sf) <-  "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

# -------------------------------------------------------------------- #
# save site_sf
st_write(site_sf, dsn = paste0(p_dat_derived, "site_sf.shp"), append = FALSE)

# load back in:
site_sf <- st_read(paste0(p_dat_derived, "site_sf.shp"))
st_crs(site_sf) <-  "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
site_sf <- site_sf %>% st_make_valid()
# bind sf with df
left_join(site_sf, site_df, by = "site")
site_sf$site[9] == site_df$site[9]
```

```{r simplify-states}
install.packages("rmapshaper")
library(rmapshaper)

world_states
world_10

object_size(world_states)
object_size(world_10)

plot(world_10_simple)

world_states_simple <- ms_simplify(world_states, keep = 0.1,
                                   keep_shapes = FALSE)
world_states_simple <- st_make_valid(world_states_simple)

world_10_simple <- ms_simplify(world_10, keep = 0.1,
                               keep_shapes = FALSE)
world_10_simple <- st_make_valid(world_10_simple)

world_states_simple %>% st_drop_geometry() %>% as_tibble() %>% select(name, name_alt, type_en, adm0_label, adm0_a3, geonunit)

# --------------------------- #
# union state boundaries
states_union_simple <- world_states_simple %>% 
  group_by(admin) %>%
  summarise(dissolved = "what") %>% st_cast()

states_union <- world_states %>% 
  group_by(admin) %>%
  summarise(dissolved = "what") %>% st_cast()

states_union %>% filter(geonunit == "Afghanistan") %>% st_geometry() %>% plot()
world_states_simple %>% filter(geonunit == "Afghanistan") %>% st_geometry() %>% plot()

states_union %>% filter(geonunit == "Bosnia and Herzegovina") %>% st_geometry() %>% plot()
world_states_simple %>% filter(geonunit == "Bosnia and Herzegovina") %>% st_geometry() %>% plot()

world_states %>% filter(geonunit == "Bosnia and Herzegovina") %>% st_geometry() %>% plot()
world_states %>% filter(name_en == "Bosnia and Herzegovina") %>% st_geometry() %>% plot()
world_10 %>% filter(name_en == "Bosnia and Herzegovina") %>% st_geometry() %>% plot()


world_states %>% filter(grepl("Bosnia", admin)) %>% st_geometry() %>% plot()
world_states_simple %>% filter(grepl("Bosnia", admin)) %>% st_geometry() %>% plot()
states_union %>% filter(grepl("Bosnia", admin)) %>% st_geometry() %>% plot()
states_union_simple %>% filter(grepl("Bosnia", admin)) %>% st_geometry() %>% plot()
states_union_simple
names(world_states)
states_union$geonunit
```

```{r si-fig0-global-site-map}
my_width_s <- 8.7
my_width_m <- 11.4
my_width_l <- 17.8
sf_width <- 0.25
sf_width2 <- 0.1

gg_simple_globe <- 
  ggplot() + 
  # theme_classic() +
  theme_minimal() +
  geom_sf(data = world %>% filter(continent != "Antarctica"), 
          fill = NA, color = "gray50", size = sf_width) +
  geom_sf(data = site_sf, mapping = aes(fill = desc), size = sf_width2) + 
  labs(#subtitle = "Map projection: longlat", 
       x = NULL, y = NULL) +
  theme(legend.position = "none") + 
  geom_label_repel(data = site_sf,
                   aes(label = label, geometry = geometry),
                   stat = "sf_coordinates",
                   min.segment.length = 0.0, segment.alpha = 1,
                   point.padding = 0.2, box.padding = 0.2,
                   label.size = 0.3, color = "black") + 
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "eqearth")))



ggsave(plot = gg_simple_globe,
       filename = paste0(p_plots, "fig0_new.pdf"),
       width = 5, height = 2.5, units = "in")


# now, add labels:

gg_globe_labels <- 
  ggplot() + 
  theme_minimal() +
  geom_sf(data = world %>% filter(continent != "Antarctica"), 
          fill = NA, color = "gray60", size = sf_width) +
  geom_sf(data = site_sf %>%
            mutate(fancy_labels = fancy_labels), 
          mapping = aes(fill = fancy_labels), size = sf_width2) + 
  labs(#subtitle = "Map projection: longlat", 
       x = NULL, y = NULL) +
  theme(legend.position = "none") + 
  
  geom_label_repel(data = site_sf,
                   aes(label = fancy_labels, geometry = geometry),
                   stat = "sf_coordinates",
                   min.segment.length = 0.0, segment.alpha = 1,
                   point.padding = 0.2, box.padding = 0.2,
                   label.padding = 0.2,
                   label.size = 0.2, color = "black", direction = "both",
                   size = 3) + 
    coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "eqearth"))) +
  guides(fill = guide_legend(nrow = 4, ncol = 3))


ggsave(plot = gg_globe_labels,
       filename = paste0(p_plots, "fig0_labels.pdf"),
       width = 7, height = 3.35, units = "in")




gg_biome_world <- 
  ggplot() +
  theme_minimal() +
  geom_sf(data = biomes2017, 
          mapping = aes(fill = BIOME_NAME)) + 
    theme(legend.position = "bottom") + 
    guides(fill = guide_legend(ncol = 3)) + 
    coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "eqearth")))

ggsave(plot = gg_biome_world,
       filename = paste0(p_plots, "fig0_biomes.pdf"),
       width = 7, height = 5, units = "in")
```

```{r misc-global-site-map}
# -------------------------------------------------------------------- #
gg_globe <-
  ggplot() + 
  theme_classic() +
  geom_sf(data = world %>% filter(continent != "Antarctica"), fill = NA, color = "gray") +
  geom_sf(data = site_sf, mapping = aes(fill = desc)) + 
  labs(#subtitle = "Map projection: longlat", 
       x = NULL, y = NULL) +
  theme(legend.position = "bottom") + 
  geom_label_repel(data = site_sf,
                   aes(label = label, geometry = geometry),
                   stat = "sf_coordinates",
                   min.segment.length = 0.0, segment.alpha = 1,
                   point.padding = 0.2, box.padding = 0.2,
                   label.size = 0.3, color = "black")


gg_globe

proj_text <- "moll"
gg_globe1 <- gg_globe +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", proj_text))) +
  labs(subtitle = paste0("Map projection: ", proj_text))


proj_text <- "eck4"
gg_globe2 <- gg_globe +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", proj_text))) +
  labs(subtitle = paste0("Map projection: ", proj_text))

# use this one:
proj_text <- "eqearth"
gg_globe3 <- gg_globe +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", proj_text))) +
  labs(subtitle = paste0("Map projection: ", proj_text))

proj_text <- "natearth2"
gg_globe4 <- gg_globe +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", proj_text))) +
  labs(subtitle = paste0("Map projection: ", proj_text))


plot_grid(gg_globe, gg_globe1, gg_globe2, gg_globe3, gg_globe4)
warnings()
# -------------------------------------------------------------------- #
# save map
png(filename = paste0(p_output, "plots/site_locations_w_labels.png"), 
    width = 9, height = 6, units = "in", res = 400)
print(gg_globe4 + labs(subtitle = NULL))
dev.off()


# long labels:
# gg_site_locations_long <-
  ggplot() + 
  theme_classic() +
  geom_sf(data = world %>% filter(continent != "Antarctica"), fill = NA, color = "gray") +
  geom_sf(data = site_sf, mapping = aes(fill = lab_desc)) + 
  labs(x = NULL, y = NULL, fill = "") +
  theme(legend.position = "bottom") + 
  geom_label_repel(data = site_sf,
                   aes(label = desc,  
                       # label = label, # use this for single letter code
                       geometry = geometry),
                   stat = "sf_coordinates",
                   min.segment.length = 0.0, segment.alpha = 1,
                   point.padding = 0.2, box.padding = 0.2,
                   label.padding = 0.2,
                   label.size = 0.2, color = "black", direction = "both") +
  # labs(subtitle = paste0("Map projection: ", "natearth2")) + 
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "natearth2"))) + 
  guides(fill = guide_legend(nrow = 4, ncol = 3))


ggsave(plot = gg_site_locations_long, 
       filename = paste0(p_output, "plots/site_locations_w_labels_long.png"),
       width = 9, height = 5, units = "in", dpi = 400)

# just chinese sites
ggplot() + 
  geom_sf(data = filter(world, name %in% c("China")), fill = NA, color = "gray") +
  geom_sf(data = site_sf %>% filter(site %in% c("shaanxi", "chongqing")), mapping = aes(fill = site)) +
  # geom_sf_label(data = site_sf, aes(label = site)) + 
  labs(subtitle = paste0("Map projection: ", "natearth2")) +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "natearth2"))) + 

  theme_classic() +
  theme(legend.position = "bottom")







# just a simple map

my_width_s <- 8.7
my_width_m <- 11.4
my_width_l <- 17.8
sf_width <- 0.25
sf_width2 <- 0.1
```

```{r crop-world}

# cropping the map:
world %>% st_make_valid
world_crop <- st_crop(st_make_valid(world), extent(site_sf) )#+ c(50, 20))

# errors!

# gg_world_crop <-
  ggplot() + 
  theme_classic() +
  geom_sf(data = world, 
          fill = NA, color = "gray") +
  geom_sf(data = site_sf, mapping = aes(fill = desc)) + 
  labs(x = NULL, y = NULL) +
  theme(legend.position = "bottom") + 
  geom_label_repel(data = site_sf,
                   aes(label = label, geometry = geometry),
                   stat = "sf_coordinates",
                   min.segment.length = 0.0, segment.alpha = 1,
                   point.padding = 0.2, box.padding = 0.2,
                   label.size = 0.3, color = "black") + 
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "natearth2"))) +
  # coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "eqearth"))) +
  theme(legend.position = "none",
        plot.margin = margin(0, 0, 0, 0, "mm"))


gg_world_crop

```




# Abandonment Maps

```{r classify-into-bins}
# reclassify age rasters into bins of 5 years.

# load rasters:
age_t # see _util_files.R


# reclassification tibble:
rcl_df <-
  tibble("is" = 0:30) %>% 
  mutate(becomes = cut(is, breaks = seq(from = 5, to = 30, by = 5), include.lowest = TRUE, labels = FALSE),
         interval = cut(is, breaks = seq(from = 5, to = 30, by = 5), include.lowest = TRUE),
         labels = case_when(
           is %in% 5:10 ~ "5-10",
           is %in% 11:15 ~ "11-15",
           is %in% 16:20 ~ "16-20",
           is %in% 21:25 ~ "21-25",
           is %in% 26:30 ~ "26-30")) 



# reclassify age rasters into bins
age_t_bins <- lapply(1:11, function(i) {
  terra::classify(
  age_t[[i]]$y2017,
  rcl = select(rcl_df, is, becomes), 
  filename = paste0(p_dat_derived, "age_rasters/", run_label, "/2017_bins/", 
                    site_df$site[i], "_y2017_bins", run_label, ".tif"),
  overwrite = TRUE)
}
)
names(age_t_bins) <- site_df$site



# reclassify max_age rasters into bins
max_age_t_bins <- lapply(1:11, function(i) {
  terra::classify(
  max_age_t[[i]],
  rcl = select(rcl_df, is, becomes), 
  filename = paste0(p_dat_derived, "max_age/", run_label, "/bins/", 
                    site_df$site[i], "_max_age_bins", run_label, ".tif"),
  overwrite = TRUE)
}
)

names(max_age_t_bins) <- site_df$site


```


```{r *fig1-final}

inset_map_titles <- paste0("(", gsub("_", "", site_df$label), ") ", site_df$description)
# cat(inset_map_titles, sep = "; ")
new_titles <- inset_map_titles
new_titles[1] <- "(b) Vitebsk, Belarus /\nSmolensk, Russia"
new_titles[8] <- "(o) Orenburg, Russia /\nUralsk, Kazakhstan"

age_t_bins # see "_util_files.R"

# ----------------------------------------------------- #
# horizontal
# ----------------------------------------------------- #
mval <- 1
include_borders_simple <- FALSE
include_borders <- FALSE

{
  pdf(file = paste0(p_output, "plots/fig1", 
                    if(include_borders) {"_borders"},
                    if(include_borders_simple) {"_borders_simple"},
                    run_label, ".pdf"),
      width = 7, height = 6)
  
  par(mfrow = c(3, 4), 
      # mar=c(0,0,0,0), 
      oma=c(1,1,0,0), 
      adj = 0,
      cex.main = 1#,
      # tcl = -0.2
      )
  
  for (i in 1:11) {
    terra::plot(age_t_bins[[i]], type = "classes",
       levels = rcl_df$labels %>% unique() %>% .[-1],
       col = viridis(n = 5, direction = -1), 
       colNA = "gray95",
       main = new_titles[i], #inset_map_titles[i],
       legend = FALSE, 
       cex = 1, # controls title sizes
       mar = c(mval+0.5, mval, mval + 1.2, mval) # bltr     
       )
    
    if(include_borders) {
      plot(
        world_states %>% 
          {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
          st_geometry(), 
        add = T, border = "gray50")
      
      plot(        
        states_union %>%
          {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
          st_geometry(), add = T, border = "black")
    }
    
    if(include_borders_simple) {
      plot(
        world_states_simple %>% 
          {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
          st_geometry(), 
        add = T, border = "gray50")
      
      plot(        
        states_union_simple %>%
          {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
          st_geometry(), add = T, border = "black")
    }
    
  
  }
  
  plot.new()
  plot(age_t_bins[[i]], type = "classes",
       levels = rcl_df$labels %>% unique() %>% .[-1],
         legend = "left", #legend.shrink = 5,
         plg = list(cex = 1.5, title = "Abandonment\nDuration (Years)", title.adj = 0),
         col = viridis(n = 5, direction =  -1), colNA = "gray95",
       legend.only = TRUE
       )
  
  dev.off()

}

```


```{r si-fig1-max-age-maps}
max_age_t_bins # see _util_files.R


mval <- 1
include_borders <- FALSE

{
  pdf(file = paste0(p_output, "plots/fig1_max", #run_label, "/maps/max_age_maps", 
                    if(include_borders) {"_borders"},
                    run_label, ".pdf"),
    width = 7, height = 6)
  
  par(mfrow = c(3, 4), 
      # mar=c(0,0,0,0), 
      oma=c(1,1,0,0), 
      adj = 0,
      cex.main = 1#,
      # tcl = -0.2
      )
  
  for (i in 1:11) {
    terra::plot(max_age_t_bins[[i]], type = "classes",
       levels = rcl_df$labels %>% unique() %>% .[-1],
       col = viridis(n = 5, direction = -1), colNA = "gray95",
       main = new_titles[i], #inset_map_titles[i],
       legend = FALSE, cex = 1,
       mar = c(mval+0.5, mval, mval + 1.2, mval) # bltr     
       )
    
    if(include_borders) {
      plot(world_states %>%
             {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
             st_geometry(), add = T, border = "gray50")
      
      plot(world_10 %>%
             {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
             st_geometry(), add = T, border = "black")
      }
  
  }


  plot.new()
  plot(max_age_t_bins[[i]], type = "classes",
       levels = rcl_df$labels %>% unique() %>% .[-1],
         legend = "bottomleft", legend.shrink = 5,
         plg = list(cex = 1.5, title = "Max. Abandonment\nDuration (Years)", title.adj = 0),
         col = viridis(n = 5, direction =  -1), colNA = "gray95",
       legend.only = TRUE)
  
  dev.off()

}


```



```{r generate-age-maps-thumb}
# Unused, but potentially useful.
# Save individual maps for each site, to put together manually into a figure.


i <- 9

# titles
inset_map_titles <- paste0("(", gsub("_", "", site_df$label), ") ", site_df$description)


# ------------------------------------------------------------------ #
# base terra plot, with bins
plot(age_t_bins[[9]])
# ------------------------------------------------------------------ #
cc_save_fig1_inset <- function(i, include_borders = FALSE, titles_vector = inset_map_titles, maxpixels = 5e6,
                               myh = 4.5, myw = 4.5) {
 pdf(file = paste0(p_output, "plots/fig1/", 
                   "fig1_inset",
                    if(!include_borders) {""} else {"_borders"}, 
                    site_df$label[i], run_label, ".pdf"),
    width = myw, height = myh)

  par(bg=NA, mar=c(0,0,0,0), oma=c(0,0,0,0), adj = 0)


  plot(age_t_bins[[i]], type = "classes",
     levels = rcl_df$labels %>% unique() %>% .[-1],
     col = viridis(n = 5, direction = -1), colNA = "gray95",
     main = titles_vector[i],
     legend = FALSE, #maxcell = maxpixels
     mar = c(1.5, 1.5, 1.5, 0.5) # bltr     
     )

  if(include_borders) {
    plot(world_states %>%
           {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
           st_geometry(), add = T, border = "gray50")
    
    plot(world_10 %>%
           {if (i == 6) . else  st_crop(., extent(site_sf[i, ]))} %>%
           st_geometry(), add = T, border = "black")
    }

  dev.off() 
}


lapply(1:11, function(i) cc_save_fig1_inset(i = i, myh = 4.25, myw = 4.25))




# ------------------------------------------------------------------ #
# Save legend only
# ------------------------------------------------------------------ #

{
pdf(file = paste0(p_output, "plots/fig1/fig1_inset_legends", run_label, ".pdf"), 
    width = 4, height = 5)
plot.new()
par(bg=NA, mar=c(0,0,0,0), oma=c(0,0,0,0))

plot(age_t_bins[[i]], type = "classes",
     levels = rcl_df$labels %>% unique() %>% .[-1],
       legend = "right", legend.shrink = 5, 
       plg = list(cex = 3),
       col = viridis(n = 5, direction =  -1), colNA = "gray95",
     legend.only = TRUE)
dev.off()
}


# ------------------------------------------------------------------ #
# Save world map
# ------------------------------------------------------------------ #

gg_world_thumb <- ggplot() +
  theme_classic() +
  geom_sf(data = world %>% filter(continent != "Antarctica") #%>%
            # st_crop(., #extent(site_sf) + c(50, 20)
            #         extent(c(-130, 140, -60, 160))
            #         )
          , 
          fill = NA, color = "gray") +
  geom_sf(data = site_sf, mapping = aes(fill = desc)) + 
  labs(#subtitle = "Map projection: longlat", 
       x = NULL, y = NULL) +
  geom_label_repel(data = site_sf,
                   aes(label = label, geometry = geometry),
                   stat = "sf_coordinates",
                   min.segment.length = 0.0, segment.alpha = 1,
                   point.padding = 0.2, box.padding = 0.2,
                   label.size = 0.3, color = "black") + 
  coord_sf(expand = TRUE, crs = st_crs("+proj=natearth2")
           ) +
  theme(legend.position = "none",
        plot.margin = margin(0, 0, 2, 0, "mm"))

ggsave(plot = gg_world_thumb,
       filename = paste0(p_plots, "/fig1/fig1_world.pdf"),
       width = 5, height = 3)

```
```{r age-levelplots}
# Save abandonment age levelplots with histograms

for(i in 1:11) {cc_age_levelplot_hist(site_index = i, year_or_max = "max")}
for(i in 1:11) {cc_age_levelplot_hist(site_index = i, year_or_max = "2017")}

```

```{r aggregate-abn-age}
# aggregate abandonment maps to a coarser resolution


# aggregate using max function
age_2017_aggregated_10_max_l <- lapply(1:11, function(i) {
  temp <- age_r[[i]][["y2017"]]
  temp <- raster::aggregate(temp, fact = 100, fun = max)
  temp
})

# update names of the list items, and rename the raster layers after each site
names(age_2017_aggregated_10_max_l) <- site_df$site
for (i in 1:11) {
  names(age_2017_aggregated_10_max_l[[i]]) <- site_df$site[i]
}


# aggregate with mean, after removing short-term fallowing 
age_2017_aggregated_10_mean_l <- lapply(1:11, function(i) {
  temp <- age_r[[i]][["y2017"]]
  temp[temp < 5] <- NA # exclude short-term fallowing:
  temp <- raster::aggregate(temp, fact = 10, fun = mean, na.rm = TRUE)
  temp
})
# update names of the list items, and rename the raster layers after each site
names(age_2017_aggregated_10_mean_l) <- site_df$site
for (i in 1:11) {
  names(age_2017_aggregated_10_mean_l[[i]]) <- site_df$site[i]
}


# plot(age_2017_aggregated_10_max_l$mato_grosso)
# plot(age_r$mato_grosso$y2017)
plot(age_2017_aggregated_10_mean_l$mato_grosso)
plot(age_r$mato_grosso$y2017)

# set lattice levelplot settings
lattice.options(
  layout.heights=list(bottom.padding=list(x=0), top.padding=list(x = 0)),
  layout.widths=list(left.padding=list(x=0), right.padding=list(x=0))
  )

# create a color palette with the NA color (the plot panel background) set to gray
myTheme <- viridisTheme()
myTheme$panel.background$col <- 'gray95' 

# save level plots to objects
for (i in 1:11) {
  assign(
  x = paste0("lp_ag_mean", site_df$label[i]),
  value =
      levelplot(
        # age_2017_aggregated_10_max_l[[i]],
        age_2017_aggregated_10_mean_l[[i]],
        main = site_sf$lab_desc[i],
        colorkey = TRUE,
        scales=list(draw=TRUE),
        par.settings = myTheme,
        at = seq(5, 31, by = 1), 
        xlab = NULL, ylab = NULL,
        margin = FALSE
        )
  )
}


# map of site locations:
gg_site_locations <- gg_globe +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "natearth2"))) +
  theme(legend.position = "none",
        plot.margin = margin(0, 0, 0, 0, "cm"))

# create plot_grid elements
pan_top <- plot_grid(lp_ag_w, lp_ag_b, lp_ag_v, lp_ag_o, nrow = 1)
pan_l <- plot_grid(lp_ag_n, lp_ag_mg, ncol = 1)
pan_r <- plot_grid(lp_ag_s, lp_ag_c, ncol = 1)
pan_bottom <- plot_grid(lp_ag_g, lp_ag_bh, lp_ag_i, nrow = 1, ncol = 3)

# create plot_grid elements
pan_top <- plot_grid(lp_ag_mean_w, lp_ag_mean_b, lp_ag_mean_v, lp_ag_mean_o, nrow = 1)
pan_l <- plot_grid(lp_ag_mean_n, lp_ag_mean_mg, ncol = 1)
pan_r <- plot_grid(lp_ag_mean_s, lp_ag_mean_c, ncol = 1)
pan_bottom <- plot_grid(lp_ag_mean_g, lp_ag_mean_bh, lp_ag_mean_i, nrow = 1, ncol = 3)

pan_middle <- plot_grid(pan_l, gg_site_locations, pan_r, NULL,
          rel_widths = c(1, 2, 1, 0.05),
          nrow = 1, ncol = 4)

ggsave(plot = plot_grid(NULL, pan_top, pan_middle, pan_bottom, 
                        ncol = 1, nrow = 4,
                        rel_heights = c(0.05, 1, 1.8, 1)),
       filename = paste0(p_plots, run_label, #"/presentation", 
       "/site_locations_w_abn_age_ag_mean_10_gray", run_label, ".pdf"),

       width = 14, height = 10, units = "in", dpi = 400
       )


```
```{r old-fig1-w-global-map}
# An old version of Figure 1, which showed a map of abandonment for each site,
# surrounding a global map.

myTheme <- viridisTheme()
myTheme$regions$col <- viridis(n = 100, direction = -1)
myTheme$panel.background$col = 'gray95' 

lattice.options(
  layout.heights=list(bottom.padding=list(x=0), top.padding=list(x = 0)),
  layout.widths=list(left.padding=list(x=0), right.padding=list(x=0))
  )

# for testing
for (i in 1:11) {
  assign(
  x = paste0("tp", i),
  value =
      levelplot(
        age_2017_l[[i]],
        # age_r[[i]][["y2017"]],
        main = site_sf$lab_desc[i],
        colorkey = TRUE, scales=list(draw=TRUE),
        par.settings = myTheme,
        # col.regions = map_colors,
        # col.regions = viridis(n = 31),
        at = seq(5, 31, by = 1), 
        xlab = NULL, ylab = NULL, margin = FALSE
        )
  )
}

# --------------------------------------------------- #
# make levelplots for full rasters
# --------------------------------------------------- #
for (i in 1:11) {
  assign(
  x = paste0("lp", site_df$label[i]),
  value =
      levelplot(
        age_r[[i]][["y2017"]],
        main = site_sf$lab_desc[i],
        colorkey = TRUE,
        scales=list(draw=TRUE),
        par.settings = myTheme, # use myTheme for gray background
        at = seq(5, 31, by = 1), 
        xlab = NULL, ylab = NULL,
        margin = FALSE
        )
  )
}

gg_site_locations <- gg_globe +
  coord_sf(expand = TRUE, crs = st_crs(paste0("+proj=", "natearth2"))) +
  theme(legend.position = "none",
        plot.margin = margin(0, 0, 2, 0, "mm"))

# for testing:
# pan_top <- plot_grid(tp11, tp1, tp10, tp8, nrow = 1)
# pan_l <- plot_grid(tp7, tp6, ncol = 1)
# pan_r <- plot_grid(tp9, tp3, ncol = 1)
# pan_bottom <- plot_grid(tp4, tp2, tp5, nrow = 1, ncol = 3)

# with full rasters, gray background
pan_top <- plot_grid(lp_w, lp_b, lp_v, lp_o, nrow = 1)
pan_l <- plot_grid(lp_n, lp_mg, ncol = 1)
pan_r <- plot_grid(lp_s, lp_c, ncol = 1)
pan_bottom <- plot_grid(lp_g, lp_bh, lp_i, nrow = 1, ncol = 3)

pan_middle <- plot_grid(pan_l, gg_site_locations, pan_r, NULL,
          rel_widths = c(1, 2, 1, 0.05),
          nrow = 1, ncol = 4)

# save composite figure with map
ggsave(plot = plot_grid(NULL, pan_top, pan_middle, pan_bottom, 
                        ncol = 1, nrow = 4,
                        rel_heights = c(0.05, 1, 1.8, 1)),
       filename = paste0(p_plots, run_label, #"/presentation", 
       "/site_locations_w_abn_age_gray", run_label, ".pdf"),
       width = 14, height = 10, units = "in", dpi = 400
       )



# lp1:lp11 are with regular viridis theme, with white as the background
# lp_b : lp_w are with a gray background

# -------------------------------------------------------- #
# save plain composite figure ---------------------------- #
# 

ggsave(plot = 
         # plot_grid(lp1, lp2, lp3, lp4, lp5, lp6, lp7, lp8, lp9, lp10, lp11, NULL, nrow = 3, ncol = 4),
         plot_grid(lp_b, lp_bh, lp_c, lp_g, lp_i, lp_mg, lp_n, lp_o, lp_s, lp_v, lp_w, NULL, nrow = 3, ncol = 4),
       filename = paste0(p_plots, run_label, #"/presentation", 
       # "/abn_age_maps", 
       "/abn_age_maps_2017_rev", 
       run_label, ".pdf"),
       width = 14, height = 10, units = "in", dpi = 400
       )

ggsave(plot = 
         plot_grid(gg_site_locations #+ labs(title = "Site locations")
                   ,
                   plot_grid(lp_b, lp_bh, lp_c, 
                             lp_g, lp_i, lp_mg, 
                             lp_n, lp_o, lp_s, 
                             lp_v, lp_w, NULL, 
                             nrow = 4, ncol = 3),
                   rel_heights = c(1.5, 4),
                   nrow = 2, ncol = 1
                   ),
       filename = paste0(p_plots, run_label, #"/presentation", 
       "/abn_age_maps_test2", 
       run_label, ".pdf"),
       width = 11, height = 13, units = "in", dpi = 400
       )
```



# Quadrant plot
```{r quadrant}
install.packages("geomtextpath")
library(geomtextpath)
# The idea: a measure of abandonment decay on the y-axis, and a measure of abandonment extent on the x-axis


# load required files:



# -------- x-axis ----------- #
# abandonment as of 2017, as a proportion of total cropland extent
area_summary_df %>% 
  select(site, area_abn_ha_2017, area_ever_abn_ha, area_2017_as_prop_total_crop)



# -------- y-axis ----------- #
# half life on the y-axis
time_to_l3$p %>% unique()

time_to_l3 %>% filter(p == 0.5) %>%
  select(site, time_to_value_low, time_to_value, time_to_value_high, p) %>% 
  # mutate(time_to_value_high = ifelse(is.na(time_to_value_high), 0, time_to_value_high)) %>%
  arrange(time_to_value)

time_to_combo %>% filter(mod == "l3", type == "mean", proportion == 0.5)


quadrant_df <- area_summary_df %>% 
  select(site, area_abn_ha_2017, area_ever_abn_ha, area_2017_as_prop_total_crop) %>%
  left_join(
    time_to_combo %>% filter(mod == "l3", type == "mean", proportion == 0.5),
    # time_to_l3 %>% 
  #             filter(p == 0.5) %>%
  #             select(site, time_to_value_low, time_to_value, time_to_value_high, p),
            by = "site") %>%
  left_join(., site_df %>% select(site, description), by = "site") %>%
  mutate(description = ifelse(site == "belarus", "Vitebsk, Belarus /\nSmolensk, Russia", description),
         description = ifelse(site == "orenburg", "Orenburg, Russia /\nUralsk, Kazakhstan", description))



# plot
gg_quadrant <-
  ggplot(data = quadrant_df, 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = time_to,
                     label = description,
                     size = area_abn_ha_2017/10^3,
                     # color = site
                     )) + 
  geom_vline(xintercept = mean(quadrant_df$area_2017_as_prop_total_crop), 
             linetype="dashed", alpha = 0.4, colour = "red") +
  geom_hline(yintercept = mean(quadrant_df$time_to), 
             linetype="dashed", alpha = 0.4, colour = "red") +
  geom_point(alpha = 0.6) + 
  geom_text_repel(
                  box.padding   = 1,
                  point.padding = 1,
                  min.segment.length = 0,
                  size = 3,
                  # nudge_x = 5,
                  # direction = "x",
                  # force = 1,
                  segment.color = 'grey10'
                  ) +
  # geom_textpath(size = 3, hjust = -.3, vjust = -0.5) + 
  
  labs(x = "Area abandoned as a percent (%) of total cropland extent",
       y = "Half-life (years)",
       size = "Area abandoned \n(2017, kha)",
       caption = "Dashed red lines represent mean values.\nHalf lives derived from mean trajectories.") +
  theme_classic() + 
    theme(legend.position = c(0.15, 0.8)) + 
    scale_size_continuous(range = c(0.1, 10))


# save 
ggsave(plot = gg_quadrant,
       filename = paste0(p_plots, run_label, "/quadrant_fig", run_label, ".pdf"), 
       # width = 9, height = 7,
       width = 7, height = 6,
       # width = 3.42, height = 3.4, 
       units = "in")

```

```{r quad-endpoints}

quadrant_endpoints <-
  endpoint_half_lives %>%
  left_join(area_summary_df %>% 
              select(site, area_abn_ha_2017, area_ever_abn_ha, area_2017_as_prop_total_crop),
  by = "site")

# derive mean half-live
quad_end_mean <- 
  quadrant_endpoints %>%
  group_by(site, area_2017_as_prop_total_crop, area_abn_ha_2017) %>%
  summarise(mean_hl = mean(mean),
            median_hl = median(mean)
            )

quadrant_endpoints %>%
  group_by(site) %>%
  summarise(mean_hl = mean(mean),
            median_hl = median(mean))


# ------------------------------------------------------------------------------- #
# Quadrant, with endpoint derived half-lives
# ------------------------------------------------------------------------------- #
gg_quadrant_endpoints_all <-
  ggplot(data = quadrant_endpoints, 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = mean,
                     size = area_abn_ha_2017/10^3,
                     # group = site,
                     color = endpoint, 
                     group = site
                     )) + 
  geom_vline(xintercept = mean(quadrant_endpoints$area_2017_as_prop_total_crop), 
             linetype="dashed",# alpha = 0.4, 
             colour = "red") +
  geom_hline(yintercept = mean(quadrant_endpoints$mean, na.rm = TRUE), 
             linetype="dashed", #alpha = 0.4, 
             colour = "red") +
    
    # add half-lives by endpoint:
  geom_point(alpha = 0.6,
             # position = position_dodge(0.2)
             ) +
    # geom_boxplot(notch = TRUE, alpha = 0.5) + 
    scale_color_viridis(direction = -1) +


    # add point for mean half-live across 
    geom_point(data = quad_end_mean, # see mean below,
               mapping = aes(x = area_2017_as_prop_total_crop,
                             y = mean_hl,
                             size = area_abn_ha_2017/10^3
                             ),
               inherit.aes = F) +
    
    geom_text_repel(data = quad_end_mean, # see mean below,
               mapping = aes(x = area_2017_as_prop_total_crop,
                             y = mean_hl,
                             label = cap_update_labels(site),
                             size = area_abn_ha_2017/10^3
                             ),
               inherit.aes = FALSE,
               box.padding = 1, point.padding = 1, min.segment.length = 0,
                    size = 3, segment.color = 'grey10') +
  labs(x = "Area abandoned in 2017 as a percent (%) of total cropland extent",
       y = "Half-life (years)",
       size = "Area abandoned\n(2017, kha)",
       caption = "Dashed red lines represent mean values",
       color = "Endpoint\n(Years)") +
  theme_classic() +
      theme(legend.position = c(0.15, 0.6))


# save 
ggsave(plot = gg_quadrant_endpoints_all,
       filename = paste0(p_plots, run_label, "/quadrant_endpoints_all", run_label, ".pdf"), 
       # width = 9, height = 7,
       width = 7, height = 6,
       # width = 3.42, height = 3.4, 
       units = "in")



# ------------------------------------------------------------------------------- #
# just the mean values derived from the endpoints analysis:
# ------------------------------------------------------------------------------- #
  
gg_quadrant_endpoints_mean <- 
  ggplot(data = quad_end_mean, 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = mean_hl,
                     label = cap_update_labels(site),
                     size = area_abn_ha_2017/10^3,
                     # color = site
                     )) + 
  geom_vline(xintercept = mean(quad_end_mean$area_2017_as_prop_total_crop), 
             linetype="dashed", #alpha = 0.4, 
             colour = "red") +
  geom_hline(yintercept = mean(quad_end_mean$mean_hl), 
             linetype="dashed",# alpha = 0.4,
             colour = "red") +
  geom_point(alpha = 0.5) + 
  geom_text_repel(
                  box.padding   = 1,
                  point.padding = 1,
                  min.segment.length = 0,
                  size = 3,
                  # nudge_x = 5,
                  # direction = "x",
                  # force = 1,
                  segment.color = 'grey10'
                  ) +

  labs(x = "Area abandoned as a percent (%) of total cropland extent",
       y = "Mean half-life (years) across endpoints",
       size = "Area abandoned \n(2017, kha)",
       caption = "Dashed red lines represent mean values") +
  theme_classic() + 
    theme(legend.position = c(0.15, 0.8)) +
    scale_size_continuous(range = c(0.1, 10))

# ------- save -------- #
# save 
ggsave(plot = gg_quadrant_endpoints_mean,
       filename = paste0(p_plots, run_label, "/quadrant_endpoints_mean", run_label, ".pdf"), 
       width = 7, height = 6, units = "in")

# ------------------------------------------------------------------------------- #
# just the median values derived from the endpoints analysis:
# ------------------------------------------------------------------------------- #
  
gg_quadrant_endpoints_median <-
  ggplot(data = quad_end_mean, 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = median_hl,
                     label = cap_update_labels(site),
                     size = area_abn_ha_2017/10^3,
                     # color = site
                     )) + 
  geom_vline(xintercept = mean(quad_end_mean$area_2017_as_prop_total_crop), 
             linetype="dashed", #alpha = 0.4, 
             colour = "red") +
  geom_hline(yintercept = mean(quad_end_mean$mean_hl), 
             linetype="dashed",# alpha = 0.4,
             colour = "red") +
  geom_point(alpha = 0.5) + 
  geom_text_repel(box.padding   = 1,
                  point.padding = 1,
                  min.segment.length = 0,
                  size = 3,
                  segment.color = 'grey10'
                  ) +

  labs(x = "Area abandoned as a percent (%) of total cropland extent",
       y = "Median half-life (years) across endpoints",
       size = "Area abandoned \n(2017, kha)",
       caption = "Dashed red lines represent mean values") +
  theme_classic() + 
    theme(legend.position = c(0.15, 0.8)) +
    scale_size_continuous(range = c(0.1, 10))

# ------- save -------- #
# save 
ggsave(plot = gg_quadrant_endpoints_median,
       filename = paste0(p_plots, run_label, "/quadrant_endpoints_median", run_label, ".pdf"), 
       width = 7, height = 6, units = "in")


# with box plots:
quadrant_endpoints %>% 
  select(site, area_2017_as_prop_total_crop) %>% 
  unique()

# gg_quadrant_endpoints_box <-
  ggplot(data = quadrant_endpoints, 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = mean,
                     # size = area_abn_ha_2017/10^3,
                     # group = site,
                     color = endpoint, 
                     group = site
                     )) + 
  geom_vline(xintercept = mean(quadrant_endpoints$area_2017_as_prop_total_crop), 
             linetype="dashed", #alpha = 0.4, 
             colour = "red") +
  geom_hline(yintercept = mean(quadrant_endpoints$mean, na.rm = TRUE), 
             linetype="dashed", #alpha = 0.4, 
             colour = "red") +
    
    # add half-lives by endpoint:
 
    geom_boxplot(notch = F, alpha = 0.5) + 
    scale_color_viridis(direction = -1) +
    geom_rug(sides = "bl") + 
    geom_text(data = quadrant_endpoints %>% 
                select(site, x = area_2017_as_prop_total_crop) %>% 
                unique() %>% mutate(y = 0),
              mapping = aes(x, y, label = site), inherit.aes = FALSE,
              angle = 270, vjust = 1, hjust = 0, )
              

    geom_text_repel(data = quad_end_mean, # see mean below,
               mapping = aes(x = area_2017_as_prop_total_crop,
                             y = mean_hl,
                             label = cap_update_labels(site),
                             size = area_abn_ha_2017/10^3
                             ),
               inherit.aes = FALSE,
               box.padding = 1, point.padding = 1, min.segment.length = 0,
                    size = 3, segment.color = 'grey10') +
  labs(x = "Area abandoned in 2017 as a percent (%) of total cropland extent",
       y = "Half-life (years)",
       size = "Area abandoned\n(2017, kha)",
       caption = "Dashed red lines represent mean values",
       color = "Endpoint\n(Years)") +
  theme_classic() +
      theme(legend.position = c(0.15, 0.6))




# subsets of endpoints:
i <- 15
i <- c(10, 15, 20, 25)
for (i in c(10, 15, 20, 25)){
  gg_quadrant_endpoints_subset <-
  ggplot(data = quadrant_endpoints %>% filter(endpoint %in% i), 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = mean,
                     label = cap_update_labels(site),
                     size = area_abn_ha_2017/10^3,
                     # color = site, 
                     # color = endpoint,
                     group = site
                     )) + 
  geom_vline(xintercept = mean(quadrant_endpoints %>% filter(endpoint %in% i) %>%
                                 .$area_2017_as_prop_total_crop), 
             linetype="dashed", alpha = 0.4, colour = "red") +
  geom_hline(yintercept = mean(quadrant_endpoints %>% filter(endpoint %in% i) %>% 
                                 .$mean, na.rm = TRUE), 
             linetype="dashed", alpha = 0.4, colour = "red") +
    
    # add half-lives by endpoint:
  geom_point(alpha = 0.6) + 

    geom_text_repel(box.padding = 1, point.padding = 1, min.segment.length = 0,
                    size = 3, segment.color = 'grey10', color = "black") +
  labs(x = "Area abandoned in 2017 as a percent (%) of total cropland extent",
       y = "Half-life (years)",
       size = "Area abandoned\n(2017, kha)",
       subtitle = paste0("Endpoint: ", i, " years"),
       caption = "Dashed red lines represent mean values",
       color = "Endpoint\n(Years)") +
  theme_classic() +
      theme(legend.position = c(0.15, 0.7))


# save 
ggsave(plot = gg_quadrant_endpoints_subset,
       filename = paste0(p_plots, run_label, "/quadrant_endpoints_", i, "yr", run_label, ".pdf"), 
       width = 7, height = 6,
       units = "in")
}

quadrant_endpoints %>% filter(endpoint %in% c(10, 15, 20, 25),
                              site == "shaanxi")

# Final figure for manuscript:
# subset 5, 10, 15, 20
gg_quadrant_endpoints_subset_10_25 <-
  ggplot(data = quadrant_endpoints %>% filter(endpoint %in% c(10, 15, 20, 25)), 
       mapping = aes(x = area_2017_as_prop_total_crop,
                     y = mean,
                     label = cap_update_labels(site),
                     size = area_abn_ha_2017/10^3,
                     # color = site, 
                     color = as_factor(endpoint),
                     group = site
                     )) + 
  geom_vline(xintercept = mean(quadrant_endpoints %>% 
                                 filter(endpoint %in% c(10, 15, 20, 25)) %>%
                                 .$area_2017_as_prop_total_crop), 
             linetype="dashed", #alpha = 0.4,
             colour = "red") +
  geom_hline(yintercept = mean(quadrant_endpoints %>% 
                                 filter(endpoint %in% c(10, 15, 20, 25)) %>%
                                 .$mean, na.rm = TRUE), 
             linetype="dashed", #alpha = 0.4, 
             colour = "red") +
    
    # add half-lives by endpoint:
  geom_point(alpha = 0.6) + 
    scale_color_viridis(direction = -1, discrete = TRUE) +
    
  # add mean label
    geom_text_repel(data = 
                      quad_end_mean, # see mean below,
                    # quadrant_endpoints %>% filter(endpoint %in% c(10, 15, 20, 25)),
               mapping = aes(x = area_2017_as_prop_total_crop,
                             # y = mean,
                             y = mean_hl,
                             label = cap_update_labels(site),
                             size = area_abn_ha_2017/10^3
                             ),
               inherit.aes = FALSE,
               box.padding = 1, point.padding = 1, 
               force = 3, 
               max.time = 2,
               min.segment.length = 0,
                    size = 3, segment.color = 'grey10') +

  labs(x = "Area abandoned in 2017 as a proportion of total cropland extent",
       y = "Half-life (years)",
       size = "Area abandoned\n(2017, kha)",
       caption = "Dashed red lines represent mean values",
       color = "Endpoint\n(Years)") +
  theme_classic() +
      theme(legend.position = c(0.15, 0.7)) + 
    guides(color = guide_legend(order = 1,
                              override.aes = list(size = 4, linetype = NULL)))


# save 
ggsave(plot = gg_quadrant_endpoints_subset_10_25,
       filename = paste0(p_plots, run_label, "/quadrant_endpoints_10_25yr", run_label, ".pdf"), 
       width = 7, height = 6,
       units = "in")
```
