==> scripts/4.1_predicts.rmd <==
---
title: "PREDICTS"
output: html_notebook
editor_options: 
  chunk_output_type: console
---
```{r start}
source("scripts/0_start.R")

env_size(ls())

p_dat

```


```{r load-data}
load()

predicts <- readRDS(paste0(p_dat, "Bd/PREDICTS/database.rds")) # PREDICTS itself

# predicts_dt <- fread(paste0(p_dat, "Bd/PREDICTS/database.csv")) # PREDICTS as a data.table
sites <- readRDS(paste0(p_dat, "Bd/PREDICTS/sites.rds")) # Site-level summaries
columns <- read.csv(paste0(p_dat, "Bd/PREDICTS/columns.csv")) # meta-data on variables, column names, etc.
references <- read.csv(paste0(p_dat, "Bd/PREDICTS/references.csv")) # references

predicts %>%
  filter(!is.na(Years_since_fragmentation_or_conversion)) %>%
  summary()

class(predicts)
str(sites)

str(predicts)
length(predicts) # 67 variables
names(predicts) # variable names

columns

as_tibble(predicts)
head(predicts)





predicts %>%
  filter(Country == "Brazil")


sites
head(sites)
names(sites)


# r packages - 


```

```{r predicts}

predicts
predicts %>%
  filter(Country == "Brazil")


class(predicts)
names(predicts)

# data.table
predicts_dt[,"Country"]
levels(as.factor(predicts_dt[,"Country"]))
summary(predicts_dt$Country)

# exploring the variables:
names(predicts)
levels(predicts$Diversity_metric)
levels(drop.levels(predicts$Country))
levels(predicts$Biome)
levels(predicts$Predominant_land_use)
levels(drop.levels(as.factor(predicts$Years_since_fragmentation_or_conversion)))
table(predicts$Years_since_fragmentation_or_conversion)

predicts_2nd <- predicts %>% filter(Predominant_land_use %in% c("Primary vegetation", "Young secondary vegetation", "Intermediate secondary vegetation", "Mature secondary vegetation", "Secondary vegetation (indeterminate age)"))
levels(drop.levels(predicts_2nd$Predominant_land_use))
hist(predicts_2nd$Years_since_fragmentation_or_conversion, xlim = c(0,150), breaks = 100, main = "Years since fragmentation or conversion", xlab = "Years since fragmentation or conversion")


dev.off()
par(oma = c(10, 2,0,0))
plot(predicts$Predominant_land_use, las = 2, main = "Predominant land use")#, xlim = c(0,150), breaks = 100, main = "Years since fragmentation or conversion", xlab = "Years since fragmentation or conversion")


levels(predicts$UN_subregion)
levels(predicts$Site_name) # not useful
levels(drop.levels(predicts$Ecoregion)) # there are 814 levels, for all 814 ecoregions, but according to the article (Hudson et al. 2017), only 281 have data.

levels(drop.levels(predicts$Hotspot))
levels(predicts$Realm)
levels(predicts$Class)
levels(drop.levels(predicts$Class))
levels(drop.levels(predicts$Higher_taxon))
levels(drop.levels(predicts$Class))
levels(drop.levels(predicts$Order))
levels(drop.levels(predicts$Phylum))
levels(drop.levels(predicts$Kingdom))


summary(predicts$Habitat_patch_area_square_metres)
levels(drop.levels(predicts$Study_common_taxon))

summary(predicts$Best_guess_binomial)

predicts %>%
  filter(Ecoregion == "Zambezian Halophytics") %>% # example of an ecoregion with no data associated with it. 
  as_tibble()


levels(predicts$Best_guess_binomial)

predicts[, 64]
names(predicts)

predicts %>% 
  filter(Class == "Insecta") %>%
  .$Best_guess_binomial %>%
  drop.levels() %>%
  levels()
# this means, 3164 species of birds included in the database.

predicts %>% 
  filter(Class == "Mammalia" #, Country == "Brazil"
         ) %>%
  .$Best_guess_binomial %>%
  drop.levels() %>%
  levels()

names(predicts)
levels(drop.levels(predicts$Class))
"Magnoliopsida" #8285

predicts %>% filter(Kingdom == "Plantae") %>% .$Higher_taxon %>% drop.levels() %>% levels()

log10(8285)
log10(562)

log10(3164)

predicts %>% select(Diversity_metric_type, Diversity_metric_unit, Measurement, Parsed_name, Taxon_name_entered, Best_guess_binomial, Higher_taxon) %>% head()

levels(predicts$Study_name)

10^3.5
log10(3162)
nrow(predicts)
predicts[1,]


names(predicts)
select(predicts, c(6:8, 11, 13:14, 67)) %>% head()



include_columns <- c(
  "Country",
  "Class",
  "Ecoregion",
  "Diversity_metric", 
  "Measurement", 
  "Effort_corrected_measurement",
  "Predominant_land_use",
  "Years_since_fragmentation_or_conversion",
  "Use_intensity",
  "Taxon_name_entered", 
  "Higher_taxon",
  "Genus",
  "Biome"

  )
predicts %>%
  select(include_columns) %>% 
  head()

predicts %>%
  filter(!is.na(Years_since_fragmentation_or_conversion)) %>%
  nrow() # 613991 records 

nrow(predicts$Years_since_fragmentation_or_conversion[!is.na(predicts$Years_since_fragmentation_or_conversion)])
```

```{r sites}
names(sites)
names(predicts)
levels(drop.levels(sites$Study_number))
head(sites$N_samples)
head(sites)

sites
nrow(predicts) #3,250,404

nrow(sites) #26,114 (125x fewer than PREDICTS)

str(sites)
names(sites)
levels()
```


```{r sec-veg}

levels(predicts$Predominant_land_use)
summary(predicts$Years_since_fragmentation_or_conversion)
nrow(predicts) - 2636413 # 613,991 records with information on the years since fragmentation

```


```{r predicts_number_samples}
# figuring out how many samples exist in different areas:
names(predicts)
levels(predicts$Biome)
levels(drop.levels(predicts$Ecoregion[predicts$Country == "Brazil" & 
                                        predicts$Class == "Aves"]
                   ))


# first, filter by area
pre_brazil <- predicts %>%
  filter(Country == "Brazil",
         Class == "Aves")

nrow(pre_brazil)
levels(drop.levels())

pre_brazil$Class %>%
  drop.levels() %>% levels()

```


```{r basemaps}
library(rnaturalearth)
world <- ne_countries(scale = "medium", returnclass = "sf") # can set returnclass to sf or sp.

russia <- world %>% filter(name == "Russia")
nepal <- world %>% filter(name == "Nepal")
boz <- world %>% filter(name == "Bosnia and Herz.")
boz1 <- world %>% filter(grepl("Bosnia", name))

usa <- world %>% filter(grepl("United States", name))
china <- world %>% filter(name == "China")
russia <- world %>% filter(name == "Russia")
russia <- world %>% filter(name == "Russia")
russia <- world %>% filter(name == "Russia")

plot(nepal$geometry)

china %>% st_geometry() %T>% plot()
plot(nepal$geometry)

plot(world$geometry[world$name == "Brazil"])

usa <- ne_countries(scale = "large", country = "United States of America", returnclass = "sf") # can set returnclass to sf or sp.
usa %>% st_geometry() %T>% plot()

world <- ne_countries(scale = "medium", returnclass = "sf", type = "mapunits") # can set returnclass to sf or sp.





# specific regions of the sites of interest
china <- ne_states(country = "China", returnclass = "sf")
nepal <- ne_states(country = "Nepal", returnclass = "sf")
russia <- ne_states(country = "Russia", returnclass = "sf")
bh <- ne_states(country = "Bosnia and Herzegovina", returnclass = "sf")
usa <- ne_states(country = "United States of America", returnclass = "sf")
brazil <- ne_states(country = "Brazil", returnclass = "sf")
italy <- ne_states(country = "Italy", returnclass = "sf")
iraq <- ne_states(country = "Iraq", returnclass = "sf")
uganda <- ne_states(country = "Uganda", returnclass = "sf")
belarus <- ne_states(country = "Belarus", returnclass = "sf")

italy %>% st_geometry() %T>% plot()
italy
italy["399", ]

italy[1, ]

china[china$name == "Shaanxi", ] %>% st_geometry() %T>% plot()
china %>% filter(name == "Shaanxi") %>% st_geometry() %T>% plot()

# sites
shaanxi <- china %>% filter(name == "Shaanxi")
cq <- china %>% filter(name == "Chongqing")
nebraska <- usa %>% filter(name == "Nebraska")
wisconsin <- usa %>% filter(name == "Wisconsin")
orenburg <- russia %>% filter(name == "Orenburg")
volgograd <- russia %>% filter(name == "Volgograd")
nepal
iraq
uganda
belarus
bh
sardinia <- italy %>% filter(name_alt == "Cerdeña|Sardenha")
goias <- brazil %>% filter(name == "Goiás")
mg <- brazil %>% filter(name == "Mato Grosso")

he_sites <- list(shaanxi, cq, nebraska, wisconsin, orenburg, volgograd, nepal, iraq, uganda, belarus, bh, sardinia, goias, mg)
deparse(quote(shaanxi))
he_sites_names <- c(
  "shaanxi", 
  "cq", 
  "nebraska", 
  "wisconsin", 
  "orenburg", 
  "volgograd", 
  "nepal", 
  "iraq", 
  "uganda", 
  "belarus", 
  "bh", 
  "sardinia", 
  "goias", 
  "mg")

brazil$name
plot(shaanxi$geometry)
```



```{r intersections}
# identify the ecoregions of the sites of interest. 
st_crs(ecoregions)
he_sites[2]

test_ <- st_intersection(ecoregions, st_union(he_sites[[1]]))

plot(test_$geometry)
names(he_sites) <- he_sites_names

he_sites_ecoreg_old
he_sites_ecoreg <- lapply(1:length(he_sites), 
                          FUN = function(i) {
                            st_intersection(ecoregions, 
                                            st_union(he_sites[[i]])
                                            )
                            }
                          )

length(he_sites_ecoreg)
length(he_sites_names)
names(he_sites_ecoreg) <- he_sites_names

plot(he_sites_ecoreg_old$shaanxi["ECO_NAME"])
plot(he_sites_ecoreg$shaanxi["ECO_NAME"])


plot(he_sites_ecoreg$shaanxi$geometry)

he_sites_ecoreg$shaanxi$ECO_NAME

he_sites_ecoreg_names_old
he_sites_ecoreg_names <- c()
for (i in 1:length(he_sites_ecoreg)) {
  he_sites_ecoreg_names <- c(he_sites_ecoreg_names, as.character(he_sites_ecoreg[[i]]$ECO_NAME))
}

he_sites_ecoreg
he_sites_names[1]

for (i in 1:length(he_sites_ecoreg)) {
  he_sites_ecoreg[[i]]$site <- he_sites_names[i]
}

he_sites_ecoreg_sf <- do.call("rbind", he_sites_ecoreg)
he_sites_ecoreg_sf %>% st_drop_geometry() %>% select(site, ECO_NAME)

names(he_sites_ecoreg_sf)
plot(he_sites_ecoreg_sf$geometry)


nrow(he_sites_ecoreg$shaanxi)

length(he_sites_ecoreg_names)
levels(as.factor(he_sites_ecoreg_names))



# ----------------------------
# intersection of He's sites (political units) with the G200 ecoregions
he_sites_g200 <- lapply(1:length(he_sites), 
                          FUN = function(i) {
                            st_intersection(g200, 
                                            st_union(he_sites[[i]])
                                            )
                            }
                          )

names(he_sites_g200) <- he_sites_names

plot(he_sites_g200$shaanxi$geometry)
plot(he_sites_g200$iraq["G200_REGIO"])
g200$G200_REGIO


# combine all of the list elements into a single sf file.
# first, remove the empty list elements, i.e. sites without a single G200 ecoregion (Wisconsin, Orenburg, Volgograd, and Belarus)
he_sites_g200[c(4:6, 10)]
he_sites_names[c(4:6, 10)]
he_sites_g200 <- he_sites_g200[-c(4:6, 10)]
for (i in 1:length(he_sites_g200)) {
  he_sites_g200[[i]]$site <- he_sites_names[-c(4:6, 10)][i]
}

he_sites_g200
he_sites_g200_sf <- do.call("rbind", he_sites_g200)

plot(he_sites_g200_sf$geometry)
he_sites_g200_sf %>% st_drop_geometry() %>% select(G200_REGIO, site)

mapView(filter(he_sites_ecoreg_sf, site == "shaanxi")["ECO_NAME"]) +
  mapView(filter(he_sites_g200_sf, site == "shaanxi")["G200_REGIO"])



he_sites_g200_sf %>%
  filter(site == "shaanxi") %>%
  st_geometry() %T>% plot(add =  T, border = "red")

# intersect the TNC ecoregion sf file for He's sites with the G200 sf file for He's sites:
plot(he_sites_ecoreg_sf$geometry)
plot(he_sites_g200_sf$geometry)

st_is_valid(he_sites_ecoreg_sf, reason = TRUE)
st_is_valid(he_sites_g200_sf, reason = TRUE)
he_sites_ecoreg_sf <- cc_make_valid(he_sites_ecoreg_sf)
he_sites_g200_sf <- cc_make_valid(he_sites_g200_sf)

he_sites_ecoreg_in_g200 <- st_intersection(he_sites_ecoreg_sf, he_sites_g200_sf)

mapView(filter(he_sites_ecoreg_in_g200, site == "cq"), zcol = "ECO_NAME")

he_sites_ecoreg_in_g200 %>% select(site, site.1, ECO_NAME) %>% st_drop_geometry()

names(ecoregions)

levels(drop.levels(he_sites_ecoreg$nepal$ECO_NAME))

# ------------------------------------------------------------------------------
# Map of ecoregions that overlap with He's sites ------------
# ------------------------------------------------------------------------------
mapView(filter(ecoregions, ECO_NAME %in% levels(drop.levels(he_sites_ecoreg_sf$ECO_NAME))), zcol = "ECO_NAME") + mapView(he_sites_ecoreg_sf, zcol = "ECO_NAME")

mapView(he_sites_ecoreg_in_g200[c("7", "2.1"),]$geometry)
levels(drop.levels(he_sites_ecoreg_in_g200$ECO_NAME))
levels(drop.levels(he_sites_ecoreg_sf$ECO_NAME))

# ------------------------------------------------------------------------------
write.csv(st_drop_geometry(he_sites_ecoreg_in_g200), file = "/Users/christophercrawford/Google Drive/_Projects/Ab_trajectories/test.csv")
```


```{r filter_predicts}
he_sites_ecoreg_sf %>% st_drop_geometry() %>% select(site, ECO_NAME)
# list out the ecoregions in each site

filter(he_sites_ecoreg_sf, site == "shaanxi") %>% .$ECO_NAME %>% drop.levels() %>% levels()
he_sites_ecoreg$orenburg$ECO_NAME %>% drop.levels() %>% levels()

num_eco <- c()
for (i in 1:length(he_sites_names)) {
  num_eco <- c(num_eco, 
               predicts %>% filter(Ecoregion %in% levels(drop.levels(he_sites_ecoreg[[i]]$ECO_NAME))) %>% nrow())
}


num_eco_bird <- c()
for (i in 1:length(he_sites_names)) {
  num_eco_bird <- c(num_eco_bird, 
               predicts %>% filter(
                 Ecoregion %in% levels(drop.levels(he_sites_ecoreg[[i]]$ECO_NAME)),
                 Class == "Aves") %>% nrow())
}

num_records_df <- data.frame(sites = he_sites_names, 
                      num_eco = num_eco,
                      num_eco_bird = num_eco_bird)
arrange(num_records_df, num_eco)
he_sites_names

he_sites_ecoreg_names <- levels(as.factor(he_sites_ecoreg_names))
levels(predicts$Ecoregion)
predicts %>%
  filter(Ecoregion == he_sites_ecoreg_names[2]) %>% # example of an ecoregion with no data associated with it. 
  as_tibble()

he_sites_ecoreg_names[3]
capwords <- function(s, strict = FALSE) {
    cap <- function(s) paste(toupper(substring(s, 1, 1)),
                  {s <- substring(s, 2); if(strict) tolower(s) else s},
                             sep = "", collapse = " " )
    sapply(strsplit(s, split = " "), cap, USE.NAMES = !is.null(names(s)))
}

he_sites_ecoreg_names <- capwords(he_sites_ecoreg_names)

he_sites_ecoreg_names[2]
levels(drop.levels(predicts$Ecoregion))[1:20]



test <- predicts %>%
  filter(grepl("Alto ", Ecoregion)) %>% # example of an ecoregion with no data associated with it. 
  as_tibble()

levels(drop.levels(test$Ecoregion))

intersect(he_sites_ecoreg_names, levels(drop.levels(predicts$Ecoregion)))
Reduce(intersect, list())

levels(drop.levels(predicts$Ecoregion))
summary(predicts$Ecoregion)

predicts %>% filter(Ecoregion == "(Other)") %>% as_tibble()

predicts$Ecoregion
nrow(filter(predicts,is.na(Ecoregion)))
levels(as.factor(as.character(predicts$Ecoregion)))
summary(predicts$Ecoregion)
summary(predicts$Country)

predicts_filter <- predicts %>%
  filter(Ecoregion %in% he_sites_ecoreg_names) 
as_tibble(predicts_filter)


nrow(predicts_filter) # 295411

levels(drop.levels(predicts$Country))
summary(predicts$Country)

# number of records in each country
predicts %>% filter(grepl("Braz", Country)) %>% nrow() # 59312
predicts %>% filter(grepl("United Sta", Country)) %>% nrow() # 135127
predicts %>% filter(grepl("China", Country)) %>% nrow() # 22536
predicts %>% filter(grepl("Nepal", Country)) %>% nrow() # 55332
predicts %>% filter(grepl("Italy", Country)) %>% nrow() # 62204
predicts %>% filter(grepl("Uganda", Country)) %>% nrow() # 9834
predicts %>% filter(grepl("Russia", Country)) %>% nrow() # 2130
# ----
predicts %>% filter(grepl("Iraq", Country)) %>% nrow() # 0
predicts %>% filter(grepl("Belarus", Country)) %>% nrow() # 0
predicts %>% filter(grepl("Bosnia ", Country)) %>% nrow() # 0


# number of bird records in each:
predicts %>% filter(grepl("Braz", Country), Class == "Aves") %>% nrow() # 10842
predicts %>% filter(grepl("United Sta", Country), Class == "Aves") %>% nrow() # 35151
predicts %>% filter(grepl("Italy", Country), Class == "Aves") %>% nrow() # 738
predicts %>% filter(grepl("China", Country), Class == "Aves") %>% nrow() # 14216
predicts %>% filter(grepl("Uganda", Country), Class == "Aves") %>% nrow() # 9564
# ----
predicts %>% filter(grepl("Nepal", Country), Class == "Aves") %>% nrow() # 0
predicts %>% filter(grepl("Russia", Country), Class == "Aves") %>% nrow() # 0
predicts %>% filter(grepl("Iraq", Country)) %>% nrow() # 0
predicts %>% filter(grepl("Belarus", Country)) %>% nrow() # 0
predicts %>% filter(grepl("Bosnia ", Country)) %>% nrow() # 0

# ----------------------------------------------------------------------------------------------
# number of records in each ecoregion

# the number of ecoregions in the PREDICTS database:
ecoregions_pred <- levels(as.factor(as.character(predicts$Ecoregion))) # 281 ecoregions are represented
he_sites_ecoreg_names





predicts %>% filter(grepl("Braz", Country)) %>% nrow() # 59312
predicts %>% filter(grepl("United Sta", Country)) %>% nrow() # 135127
predicts %>% filter(grepl("Nepal", Country)) %>% nrow() # 55332
predicts %>% filter(grepl("Italy", Country)) %>% nrow() # 62204
predicts %>% filter(grepl("Uganda", Country)) %>% nrow() # 9834
predicts %>% filter(grepl("Russia", Country)) %>% nrow() # 2130

he_sites_ecoreg[[1]]$ECO_NAME

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$shaanxi$ECO_NAME)))) %>% nrow() # 2809
# "Central China Loess Plateau Mixed Forests" 
# "Daba Mountains Evergreen Forests"
# "Huang He Plain Mixed Forests"         
# "Ordos Plateau Steppe"                  
# "Qin Ling Mountains Deciduous Forests"

levels(drop.levels(he_sites_ecoreg$cq$ECO_NAME))
predicts %>% filter(Ecoregion %in% levels(drop.levels(he_sites_ecoreg$cq$ECO_NAME))) %>% nrow() # 0
# "Daba Mountains Evergreen Forests"
# "Guizhou Plateau Broadleaf And Mixed Forests"
# "Sichuan Basin Evergreen Broadleaf Forests"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$volgograd$ECO_NAME)))) %>% nrow() # 0
# "East European Forest Steppe"
# "Pontic Steppe"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$orenburg$ECO_NAME)))) %>% nrow() # 0
# "East European Forest Steppe"
# "Kazakh Forest Steppe"
# "Kazakh Steppe"
# "Pontic Steppe"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$mg$ECO_NAME)))) %>% nrow() # 35975
# "Cerrado"
# "Chiquitano Dry Forests"
# "Madeira-Tapajós Moist Forests"
# "Mato Grosso Seasonal Forests"
# "Pantanal"
# "Tapajós-Xingu Moist Forests"
# "Xingu-Tocantins-Araguaia Moist Forests"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$goias$ECO_NAME)))) %>% nrow() # 20667
# "Alto Paraná Atlantic Forests"
# "Cerrado"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$wisconsin$ECO_NAME)))) %>% nrow() # 48326
# "Central Forest-grasslands Transition"
# "Central Tall Grasslands"
# "Lake"
# "Upper Midwest Forest-savanna Transition"
# "Western Great Lakes Forests"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$nebraska$ECO_NAME)))) %>% nrow() #13025
# "Central And Southern Mixed Grasslands"
# "Central Tall Grasslands"
# "Nebraska Sand Hills Mixed Grasslands"
# "Northern Mixed Grasslands"
# "Northern Short Grasslands"
# "Western Short Grasslands"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$belarus$ECO_NAME)))) %>% nrow() # 82478
# "Central European Mixed Forests"
# "Sarmatic Mixed Forests" 

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$bh$ECO_NAME)))) %>% nrow() # 16503
# "Balkan Mixed Forests"
# "Dinaric Mountains Mixed Forests"
# "Illyrian Deciduous Forests"
# "Pannonian Mixed Forests"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$sardinia$ECO_NAME)))) %>% nrow() # 30284
# "Tyrrhenian-Adriatic Sclerophyllous And Mixed Forests"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$nepal$ECO_NAME)))) %>% nrow() # 56166
# "Eastern Himalayan Alpine Shrub And Meadows"
# "Eastern Himalayan Broadleaf Forests"
# "Eastern Himalayan Subalpine Conifer Forests"
# "Himalayan Subtropical Broadleaf Forests"
# "Himalayan Subtropical Pine Forests"
# "Lower Gangetic Plains Moist Deciduous Forests"
# "Rock And Ice"
# "Terai-Duar Savanna And Grasslands"
# "Upper Gangetic Plains Moist Deciduous Forests"
# "Western Himalayan Alpine Shrub And Meadows"
# "Western Himalayan Broadleaf Forests"
# "Western Himalayan Subalpine Conifer Forests"

predicts %>% filter(Ecoregion %in% levels(drop.levels(he_sites_ecoreg$uganda$ECO_NAME))) %>% nrow() # 112660
predicts %>% filter(Ecoregion %in% c("Albertine Rift Montane Forests", 
                                     "East African Montane Forests", 
                                     "East African Montane Moorlands", 
                                     "East Sudanian Savanna", 
                                     "Lake: Afrotropic", 
                                     "Northern Acacia-Commiphora Bushlands And Thickets", 
                                     "Northern Congolian Forest-Savanna Mosaic", 
                                     "Ruwenzori-Virunga Montane Moorlands", 
                                     "Victoria Basin Forest-Savanna Mosaic")) %>% nrow() # 112660

# "Albertine Rift Montane Forests"
# "East African Montane Forests"
# "East African Montane Moorlands"
# "East Sudanian Savanna"
# "Lake"
# "Northern Acacia-Commiphora Bushlands And Thickets"
# "Northern Congolian Forest-savanna Mosaic"
# "Rwenzori-Virunga Montane Moorlands"
# "Victoria Basin Forest-savanna Mosaic"

predicts %>% filter(Ecoregion %in% capwords(levels(drop.levels(he_sites_ecoreg$iraq$ECO_NAME)))) %>% nrow() # 10198
# "Arabian Desert And East Sahero-Arabian Xeric Shrublands"
# "Eastern Mediterranean Conifer-sclerophyllous-broadleaf Forests"
# "Mesopotamian Shrub Desert"
# "Middle East Steppe"
# "Persian Gulf Desert And Semi-desert"
# "Red Sea Nubo-Sindian Tropical Desert And Semi-desert"
# "South Iran Nubo-Sindian Desert And Semi-desert"
# "Tigris-Euphrates Alluvial Salt Marsh"
# "Zagros Mountains Forest Steppe"

names(he_sites_ecoreg)
he_sites

plot(he_sites_ecoreg$shaanxi$geometry)
plot(he_sites_ecoreg$shaanxi["ECO_NAME"])

```

```{r match}
ecoregions_names_df <- as.data.frame(levels(as.factor(ecoregions$ECO_NAME))) # 827 ecoregions.

predicts_ecoregion_names_df <- as.data.frame(levels(predicts$Ecoregion))
levels(drop.levels(predicts$Ecoregion)) # 281 ecoregions
ecoregions_pred # 281 ecoregions

he_sites_ecoreg_names # 65 ecoregions

matches <- match(he_sites_ecoreg_names, ecoregions_pred)
test_df <- data.frame(matches = matches, ecoregions = he_sites_ecoreg_names)


ecoregions_pred[3]
filter(test_df, is.na(matches))
!is.na(test_df$matches)
ecoregions_pred

grep("Albert", ecoregions_pred)
eco_df <- data.frame(key = 1:length(ecoregions_pred), ecoregions_pred)

# fixing particular ecoregions:
he_sites_ecoreg_names[21] <- ecoregions_pred[76] # "Eastern Mediterranean Conifer-Sclerophyllous-Broadleaf Forests"


identical(ecoregions_pred[76], he_sites_ecoreg_names[21])
```



```{r tester}
italy_birds <- predicts %>% filter(grepl("Italy", Country), Class == "Aves")

italy_birds %>% select(include_columns) %>% head()

names(predicts)

```





```{r ecoregions}

ecoregions_wwf <- st_read(paste0(p_dat, "Habitats/TEOW/wwf_terr_ecos.shp"))

tic()
ecoregions <- st_read(paste0(p_dat, "Habitats/terr-ecoregions-TNC/tnc_terr_ecoregions.shp")) %>%
  st_make_valid()
toc() # 44.88

tic()
ecoregions <- st_make_valid(ecoregions) # very fast.
toc()

names(ecoregions)
names(ecoregions_wwf)
levels(as.factor(ecoregions_wwf$ECO_NAME)) # 827 ecoregions. (this is the WWF file)
length(as.factor(ecoregions$ECO_NAME)) # 814 ecoregions. (this is the TNC file)
levels(as.factor(ecoregions$WWF_MHTNAM)) # 16 biomes, but without "rock and ice" and "inland water," it's 14 like Hudson et al. 2014 states.
# MHT stands for Major Habitat Types

mapView(ecoregions)
plot(ecoregions["WWF_MHTNAM"])

levels(as.factor(ecoregions$BIOME)) # 14 levels of regions. 

object_size(ecoregions)

```

```{r g200-ecoregions}
g200 <- st_read(paste0(p_dat, "Habitats/global200ecoregions/g200_terr.shp"))
g200 <- st_make_valid(g200)


# exploring the datasets, g200
ecoregions$ECO_NAME
g200$G200_REGIO

match(g200$G200_REGIO, ecoregions$ECO_NAME)
match("Alaskan North Slope Coastal Tundra", levels(drop.levels(ecoregions$ECO_NAME)))

g200
names(g200)
object_size(g200)
levels(as.factor(g200$G200_BIOME)) # 14 levels of regions.
```


# Analysis Methods:
Ok, so how would I run a model to relate the different land use changes to biodiversity outcomes? 

```{r community-comp}

```

Sorensen similarity Index:
(From Wikipedia: https://en.wikipedia.org/wiki/S%C3%B8rensen%E2%80%93Dice_coefficient) Sørensen's original formula was intended to be applied to discrete data. Given two sets, X and Y, it is defined as

{\displaystyle DSC={\frac {2|X\cap Y|}{|X|+|Y|}}}{\displaystyle DSC={\frac {2|X\cap Y|}{|X|+|Y|}}}
Sørensen = 2 * (intersection between X and Y) / (X + Y)
where |X| and |Y| are the cardinalities of the two sets (i.e. the number of elements in each set). The Sørensen index equals twice the number of elements common to both sets divided by the sum of the number of elements in each set.

Jaccard = size of the intersection divided by the size of the union of the sample sets
J = intersection / union
$$
J = \frac{}{}
$$

```{r extinction-risk-sandbox}
z <- 0.25
A_0 <- 150 # original habitat, long ago
A_1 <- 125 # new habitat area, after loss, current time step
x <- 135 # new habitat area, after restoration



E_1 <- 1 - (A_1 / A_0)^z
E_2 <- 1 - ((x) / A_0)^z

r <- 1 - (x / A_1)^z
E_1
E_2
r
E_1 - r
E_2 - E_1 # this is somewhat close to r



```
